Module Arch

This module adds some code that is related to the Architecture specific modules but is in itself architecture independent.

include module type of struct include Sig end

type func_api = Sig.func_api = {

args : Ctype.t list;

ret : Ctype.t option;

}

Describe the C API of a function

type func_abi = Sig.func_abi = {

init : State.t -> State.t;
(*
Gives the initial state for verifying the function, from a given global register state. Only global registers are kept.
*)

}

Describe the ABI of a function

This is a record because I expect to add many other fields later.

type dwarf_reg_map = State.Reg.t array

The map of dwarf register: Which register number map to which ISA register

val supports : Config.Arch.t -> bool

Tells if this Arch module supports this architecture

val init : Config.Arch.t -> unit

If this arch module supports the architecture, then initialize read-dwarf state using this architecture

val initialized : unit -> Config.Arch.t option

Return Some(arch) is the loaded arch is arch and None if nothing is loaded yet.

val module_name : string

The name of the arch module. Must be the name of the module i.e. Config.arch_module

val loaded_name : string

For dynamic arch module, the name of the dynamically loaded module. Otherwise module_name

val address_size : int

The true size of addresses for memory operation

val dwarf_reg_map : unit -> dwarf_reg_map

Get the register map of the architecture

val is_local : State.Reg.t -> bool

Tell if a register is local for the ABI

val nop : unit -> Utils.BytesSeq.t

Give the opcode of the nop instruction (For Sail/Isla initialisation

val get_abi : func_api -> func_abi

Give the ABI of a function from it's C API

val pc : unit -> State.Reg.t

Give the register index for the program counter

val sp : unit -> State.Reg.t

Give the register index for the stack pointer

val assemble_to_elf : string -> string

Take an instruction string and give the name of an temporary ELF file created that contains the instruction at symbol instr.

val split_into_instrs : Elf.Symbol.data -> Elf.Symbol.data list

Split a byte-sequence into a list of instructions.

val is_ret : Utils.BytesSeq.t -> bool

Tell if an instruction is a return instruction.

val is_cmp : Utils.BytesSeq.t -> (State.Reg.t * Utils.BitVec.t) option

Tell if an instruction is a compare instruction. Returns Some (reg,bv) where the contents of reg are compared against the value bv if it is and None if not.

val is_bl : Utils.BytesSeq.t -> Utils.BitVec.t option

Tell if an instruction is an (unconditional) branch on immediate with link instructions. Returns Some bv where bv is the offset (from the address of this instruction, in the range +/-128MB) that is branched to if it is and None if not.

val ensure_loaded : Config.Arch.t -> unit

Ensure that the right architecture type is loaded

val pp_api : func_api -> Utils.Pp.document

val get : unit -> Config.Arch.t

Get the initialized architecture type. Fails (Failure) if not architecture was loaded

val get_config : unit -> Config.File.ArchConf.t

Get the configuration for the initialized architecture

val get_isla_config : unit -> Config.File.ArchConf.Isla.t

Get the Isla configuration for the initialized architecture

val load_elf_arch : Elf.File.t -> unit

Load the architecture of this Elf.File