Module Ast.Manip

This module provide generic facilities of expression and SMT statements provided by Ast. It is intended to only provide syntactic facilities over Ast types, in particular Ast.exp.

In particular it provides generic mapping and iteration function over expressions as well a function allowing to convert between the various Expression type parameter and options.

Warning: due to OCaml type system limitations, mainly issue #9456, this module is sometimes required to use Obj.magic in some specific cases. No other module should ever do that. If you need to use Obj.magic to bypass OCaml type system limitation about Ast type, add a function here instead.

Get annotations

val annot_exp : ('a, 'v, 'b, 'm) Base.exp -> 'a

Get the annotation of an expression

TODO: This would be much more efficient if the annotation was always the first member of the constructor and not the last (in that case the offset to fetch the annotation do not depend on the constructor index). This may require to modify ott.

Non-recursive maps and iters

This section is filled on demand.

direct_a_map_b take a function b -> b and applies it to all b in a, non-recursively. Then a new a with the same structure is formed.

direct_a_iter_b take a function b -> unit and applies it to all b in a, non-recursively.

val direct_exp_map_exp : (('a, 'v, 'b, 'm) Base.exp -> ('a, 'v, 'b, 'm) Base.exp) -> ('a, 'v, 'b, 'm) Base.exp -> ('a, 'v, 'b, 'm) Base.exp

val direct_exp_iter_exp : (('a, 'v, 'b, 'm) Base.exp -> unit) -> ('a, 'v, 'b, 'm) Base.exp -> unit

val direct_exp_fold_left_exp : ('a -> ('b, 'c, 'd, 'e) Base.exp -> 'a) -> 'a -> ('b, 'c, 'd, 'e) Base.exp -> 'a

val direct_exp_for_all_exp : (('a, 'b, 'c, 'd) Base.exp -> bool) -> ('a, 'b, 'c, 'd) Base.exp -> bool

val direct_exp_exists_exp : (('a, 'b, 'c, 'd) Base.exp -> bool) -> ('a, 'b, 'c, 'd) Base.exp -> bool

Recursive maps and iters

This section is filled on demand.

a_map_b take a function b -> b and applies it to all the b in a, and do that recursively on all b that appear directly or indirectly in a

a_iter_b take a function b -> unit and applies it to all the b in a, and do that recursively

Doing this when a = b is not well defined, and can be easily done using the direct version from previous section.

val exp_iter_var : ('v -> unit) -> ('a, 'v, 'b, 'm) Base.exp -> unit

Iterate a function on all the variable of an expression

val exp_map_var : 'va 'vb 'a 'b 'm. ('va -> 'vb) -> ('a, 'va, 'b, 'm) Base.exp -> ('a, 'vb, 'b, 'm) Base.exp

val exp_iter_annot : ('a -> unit) -> ('a, 'v, 'b, 'm) Base.exp -> unit

Iterate a function on all the annotations of an expression

Variable type conversion

All of those function convert the underlying variable type through the AST. They cannot be the usual map function because they change the type

val exp_conv_var : ('a -> 'b) -> ('c, 'a, 'd, 'e) Base.exp -> ('c, 'b, 'd, 'e) Base.exp

Old alias to make conversion explicit

val exp_var_subst : 'va 'a 'vb 'b 'm. ('va -> 'a -> ('a, 'vb, 'b, 'm) Base.exp) -> ('a, 'va, 'b, 'm) Base.exp -> ('a, 'vb, 'b, 'm) Base.exp

Substitute variable with expression according to substitution function

Bound variables and let-bindings

This section allow to go from expression without let-bindings to expression with them and vice-versa.

val allow_lets : ('a, 'v, Base.no, 'm) Base.exp -> ('a, 'v, 'b, 'm) Base.exp

Allow bound variables and lets in an expression. This operation is a no-op and has no runtime cost, it's just a type change.

val smt_allow_lets : ('a, 'v, Base.no, 'm) Base.smt -> ('a, 'v, 'b, 'm) Base.smt

Same as allow_lets but for the smt type

val unfold_lets : 'b1 'a 'v 'b2 'm. ?context:('b1, ('a, 'v, 'b2, 'm) Base.exp) Stdlib.Hashtbl.t -> ('a, 'v, 'b1, 'm) Base.exp -> ('a, 'v, 'b2, 'm) Base.exp

Unfold all lets. There are no remaining lets in the output, Therefore the output type of let binding can be anything including Ast.no. In particular doing allow_lets after this function is useless

Memory operation

This section allow to go from expression without memory operations to expression with them and vice-versa.

val allow_mem : ('a, 'v, 'b, Base.no) Base.exp -> ('a, 'v, 'b, 'm) Base.exp

Allow memory operations in an expression.

val smt_allow_mem : ('a, 'v, 'b, Base.no) Base.smt -> ('a, 'v, 'b, 'm) Base.smt

Same as allow_mem but for the smt type

val ty_allow_mem : Base.no Base.ty -> 'm Base.ty

Same as allow_mem but for the ty type

val check_no_mem : ('a, 'v, 'b, 'm) Base.exp -> bool

Check that not memory operation take place in that expression. Return true if that's the case and false otherwise.

This is not resilient to change of type: If a new memory constructor is added, then this function will be wrong

val expect_no_mem : ?handler:(unit -> ('a, 'v, 'b, 'm2) Base.exp) -> ('a, 'v, 'b, 'm1) Base.exp -> ('a, 'v, 'b, 'm2) Base.exp

Expect that an exp has no memory constructor, and then return it with memory removed from the type. Throws Failure if the value had memory constructors.

This is not resilient to change of type, If a new memory constructor is added, then this function will be unsound.

TODO: Find a way to make it resilient

val all_subterms : ('a, 'b, 'c, 'd) Base.exp -> ('a, 'b, 'c, 'd) Base.exp list