Module AstGen.Def

include module type of struct include Ott end

type bvar = string

syntax

type flag = string

type enum = int * int

type bvmanyarith = Ott.bvmanyarith =

| Bvand

| Bvor

| Bvxor

| Bvadd

| Bvmul

type 'm binmem = 'm Ott.binmem =

| Select of 'm

| Store of 'm

type bvcomp = Ott.bvcomp =

| Bvult

| Bvslt

| Bvule

| Bvsle

| Bvuge

| Bvsge

| Bvugt

| Bvsgt

type bvarith = Ott.bvarith =

| Bvuremi

| Bvsremi

| Bvsmodi

| Bvnand

| Bvnor

| Bvxnor

| Bvsub

| Bvudiv

| Bvudivi

| Bvsdiv

| Bvsdivi

| Bvurem

| Bvsrem

| Bvsmod

| Bvshl

| Bvlshr

| Bvashr

type manyop = Ott.manyop =

| And

| Or

| Bvmanyarith of bvmanyarith

| Concat

type 'm binop = 'm Ott.binop =

| Binmem of 'm binmem

| Eq

| Bvarith of bvarith

| Bvcomp of bvcomp

type unop = Ott.unop =

| Not

| Bvnot

| Bvredand

| Bvredor

| Bvneg

| Extract of int * int

| ZeroExtend of int

| SignExtend of int

type ('a, 'v, 'b, 'm) exp = ('a, 'v, 'b, 'm) Ott.exp =

| Var of 'v * 'a

| Bound of 'b * 'a

| Bits of Utils.BitVec.t * 'a

| Bool of bool * 'a

| Enum of enum * 'a

| Unop of unop * ('a, 'v, 'b, 'm) exp * 'a

| Binop of 'm binop * ('a, 'v, 'b, 'm) exp * ('a, 'v, 'b, 'm) exp * 'a

| Manyop of manyop * ('a, 'v, 'b, 'm) exp list * 'a

| Vec of ('a, 'v, 'b, 'm) exp list * 'a

| Ite of ('a, 'v, 'b, 'm) exp * ('a, 'v, 'b, 'm) exp * ('a, 'v, 'b, 'm) exp * 'a

| Let of 'b * ('a, 'v, 'b, 'm) exp * ('b * ('a, 'v, 'b, 'm) exp) list * ('a, 'v, 'b, 'm) exp * 'a

type 'm ty = 'm Ott.ty =

| Ty_Mem of 'm

| Ty_Bool

| Ty_BitVec of int

| Ty_Enum of int

| Ty_Array of 'm ty * 'm ty

type ('a, 'v, 'b, 'm) smt = ('a, 'v, 'b, 'm) Ott.smt =

| DeclareConst of 'v * 'm ty

| DefineConst of 'v * ('a, 'v, 'b, 'm) exp

| Assert of ('a, 'v, 'b, 'm) exp

| Simplify of ('a, 'v, 'b, 'm) exp * (string * bool) list

| Push

| Pop

| GetVersion

| CheckSat

| Exit

type ('a, 'v, 'b, 'm) smt_ans = ('a, 'v, 'b, 'm) Ott.smt_ans =

| Error of string

| Version of string

| Sat

| Unsat

| Unknown

| Unsupported

| Exp of ('a, 'v, 'b, 'm) exp

auxiliary functions on the new list types

library functions

subrules

auxiliary functions

free variables

substitutions

context application

definitions

type no = |

Generic empty type. This kind of type is explained here. In particular, when this type appear in a match case, the match case can contain a refutation pattern . to indicate that this impossible.

A refutation pattern also work with higher-level constructors, for example in this case:

type a = A of int | B of no * int
let f : a -> int = function
  | A i -> i
  | B _ -> .

It also work in product types:

 let f : no * int -> unit = function _ -> . 

However it doesn't work for sum types: the no type need to appear directly in the pattern:

type complex_empty = A of no | B of no
let f : complex_empty = function _ -> . (* does not compile *)
let f : complex_empty = function A _ | B _ -> . (* compiles *)

In case this behavior is needed, there Destructors in Ast. See this section to see how they are used.

type loc = Stdlib.Lexing.position

Lexing.position

type lrng = Isla_lang.AST.lrng

The type of an expression range. This imported from Isla to avoid having two incompatible types.

This represent a range in a source file, so generally this is a pair of loc, but it may also be Unknown.

module Size : sig ... end

type rexp = (lrng, string, string, Size.t) exp

Raw expression coming out of the parser

type rty = Size.t ty

Raw type coming out of the parser

type rsmt = (lrng, string, string, Size.t) smt

Raw SMT command coming out of the parser

type rsmt_ans = (lrng, string, string, Size.t) smt_ans

Raw SMT answer coming out of the parser