Module Def.Size

type t =

| B8

| B16

| B32

| B64

The possible sizes for memory accesses. It may be necessary to add B128 at some point

val of_bytes : int -> t

Create a size value from a valid size in byte

val of_bits : int -> t

Create a size value from a valid size in bits

val to_bytes : t -> int

Get the byte size corresponding to that value

val to_bits : t -> int

Get the bits size corresponding to that value

val equal : 'a -> 'a -> bool

val pp_bytes : t -> Utils.Pp.document

Pretty-print a size as just the byte number

val pp_bits : t -> PPrintEngine.document

Pretty print a size at "16bits" for example