Module Ast__.Base

include AstGen.Def

include AstGen.Ott

type bvar = string

syntax

type flag = string

type enum = int * int

type 'm binmem =

| Select of 'm

| Store of 'm

type bvarith =

| Bvuremi

| Bvsremi

| Bvsmodi

| Bvnand

| Bvnor

| Bvxnor

| Bvsub

| Bvudiv

| Bvudivi

| Bvsdiv

| Bvsdivi

| Bvurem

| Bvsrem

| Bvsmod

| Bvshl

| Bvlshr

| Bvashr

type bvcomp =

| Bvult

| Bvslt

| Bvule

| Bvsle

| Bvuge

| Bvsge

| Bvugt

| Bvsgt

type bvmanyarith =

| Bvand

| Bvor

| Bvxor

| Bvadd

| Bvmul

type unop =

| Not

| Bvnot

| Bvredand

| Bvredor

| Bvneg

| Extract of int * int

| ZeroExtend of int

| SignExtend of int

type 'm binop =

| Binmem of 'm binmem

| Eq

| Bvarith of bvarith

| Bvcomp of bvcomp

type manyop =

| And

| Or

| Bvmanyarith of bvmanyarith

| Concat

type ('a, 'v, 'b, 'm) exp =

| Var of 'v * 'a

| Bound of 'b * 'a

| Bits of Utils.BitVec.t * 'a

| Bool of bool * 'a

| Enum of enum * 'a

| Unop of unop * ('a, 'v, 'b, 'm) exp * 'a

| Binop of 'm binop * ('a, 'v, 'b, 'm) exp * ('a, 'v, 'b, 'm) exp * 'a

| Manyop of manyop * ('a, 'v, 'b, 'm) exp list * 'a

| Ite of ('a, 'v, 'b, 'm) exp * ('a, 'v, 'b, 'm) exp * ('a, 'v, 'b, 'm) exp * 'a

| Let of 'b * ('a, 'v, 'b, 'm) exp * ('b * ('a, 'v, 'b, 'm) exp) list * ('a, 'v, 'b, 'm) exp * 'a

type 'm ty =

| Ty_Mem of 'm

| Ty_Bool

| Ty_BitVec of int

| Ty_Enum of int

| Ty_Array of 'm ty * 'm ty

type ('a, 'v, 'b, 'm) smt =

| DeclareConst of 'v * 'm ty

| DefineConst of 'v * ('a, 'v, 'b, 'm) exp

| Assert of ('a, 'v, 'b, 'm) exp

| Simplify of ('a, 'v, 'b, 'm) exp * (string * bool) list

| Push

| Pop

| GetVersion

| CheckSat

| Exit

type ('a, 'v, 'b, 'm) smt_ans =

| Error of string

| Version of string

| Sat

| Unsat

| Unknown

| Unsupported

| Exp of ('a, 'v, 'b, 'm) exp

type no =

Generic empty type. This kind of type is explained here. In particular, when this type appear in a match case, the match case can contain a refutation pattern . to indicate that this impossible.

A refutation pattern also work with higher-level constructors, for example in this case:

type a = A of int | B of no * int
let f : a -> int = function
  | A i -> i
  | B _ -> .

It also work in product types:

let f : no * int -> unit = function _ -> . 

However it doesn't work for sum types: the no type need to appear directly in the pattern:

type complex_empty = A of no | B of no
let f : complex_empty = function _ -> . (* does not compile *)
let f : complex_empty = function A _ | B _ -> . (* compiles *)

In case this behavior is needed, there Destructors in Ast. See this section to see how they are used.

type loc = Stdlib.Lexing.position

Lexing.position

type lrng = Isla_lang.AST.lrng

The type of an expression range. This imported from Isla to avoid having two incompatible types.

This represent a range in a source file, so generally this is a pair of loc, but it may also be Unknown.

module Size = AstGen.Def.Size

type rexp = (lrng, string, string, Size.t) exp

Raw expression coming out of the parser

type rty = Size.t ty

Raw type coming out of the parser

type rsmt = (lrng, string, string, Size.t) smt

Raw SMT command coming out of the parser

type rsmt_ans = (lrng, string, string, Size.t) smt_ans

Raw SMT answer coming out of the parser

Size

module Size : sig ... end

Parsing

module Parser = AstGen.Parser

module Lexer = AstGen.Lexer

val unknown : Isla_lang.AST.lrng

val loc : Stdlib.Lexing.position -> PPrint.document

exception ParseError of loc * string

Exception that represent an Isla parsing error

exception LexError of loc * string

Exception that represent an Isla lexing error

type lexbuf = Stdlib.Lexing.lexbuf

type lexer = lexbuf -> Parser.token

type 'a parser = lexer -> lexbuf -> 'a

val parse : 'a parser -> ('b -> lexbuf) -> ?⁠filename:string -> 'b -> 'a

Parse a single Isla instruction output from a Lexing.lexbuf

val from_string : string -> Stdlib.Lexing.lexbuf

val from_channel : Stdlib.in_channel -> Stdlib.Lexing.lexbuf

val parse_exp_string : ?⁠filename:string -> string -> (AstGen.Def.lrng, string, string, AstGen.Def.Size.t) AstGen.Ott.exp

Parse a single Isla expression from a string

val parse_exp_channel : ?⁠filename:string -> Stdlib.in_channel -> (AstGen.Def.lrng, string, string, AstGen.Def.Size.t) AstGen.Ott.exp

Parse a single Isla expression from a channel

val parse_smt_ans_string : ?⁠filename:string -> string -> AstGen.Def.rsmt_ans

Parse a single Isla expression from a string

val parse_smt_ans_channel : ?⁠filename:string -> Stdlib.in_channel -> AstGen.Def.rsmt_ans

Parse a single Isla expression from a channel

include AstGen.Parser_pp

val pp_raw_bvar : string -> PPrintEngine.document

val pp_raw_bvf : Utils.BitVec.t -> PPrintEngine.document

val pp_raw_flag : string -> PPrintEngine.document

val pp_raw_vvar : int -> PPrintEngine.document

val pp_raw_name : string -> PPrintEngine.document

val pp_raw_enum : AstGen.Ott.enum -> PPrintEngine.document

val pp_raw_enum_ty : int -> PPrintEngine.document

val pp_raw_int : int -> PPrintEngine.document

val pp_raw_bvi : int -> PPrintEngine.document

val pp_raw_bv : string -> PPrintEngine.document

val pp_raw_str : string -> PPrintEngine.document

val pp_raw_mem_size : AstGen.Def.Size.t -> Utils.Pp.document

val pp_raw_binmem : AstGen.Def.Size.t AstGen.Ott.binmem -> PPrintEngine.document

val pp_raw_var : ('a -> PPrintEngine.document) -> 'a -> PPrintEngine.document

val pp_raw_bbvar : string -> PPrintEngine.document

val pp_raw_bind : ('a -> PPrintEngine.document) -> (string * ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.exp) -> PPrintEngine.document

val pp_raw_exp : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.exp -> PPrintEngine.document

val pp_raw_fbool : (string * bool) -> PPrintEngine.document

val pp_raw_smt : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt -> PPrintEngine.document

val pp_raw_smts : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt list -> PPrintEngine.document

val pp_raw_smt_ans : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt_ans -> PPrintEngine.document

val pp_raw_ty : AstGen.Def.Size.t AstGen.Ott.ty -> PPrintEngine.document

val pp_raw_bool : bool -> PPrintEngine.document

val pp_raw_unop : AstGen.Ott.unop -> PPrintEngine.document

val pp_raw_bvarith : AstGen.Ott.bvarith -> PPrintEngine.document

val pp_raw_bvcomp : AstGen.Ott.bvcomp -> PPrintEngine.document

val pp_raw_binop : AstGen.Def.Size.t AstGen.Ott.binop -> PPrintEngine.document

val pp_raw_bvmanyarith : AstGen.Ott.bvmanyarith -> PPrintEngine.document

val pp_raw_manyop : AstGen.Ott.manyop -> PPrintEngine.document

val pp_bvar : string -> PPrintEngine.document

val pp_bvf : Utils.BitVec.t -> PPrintEngine.document

val pp_flag : string -> PPrintEngine.document

val pp_vvar : int -> PPrintEngine.document

val pp_name : string -> PPrintEngine.document

val pp_enum : AstGen.Ott.enum -> PPrintEngine.document

val pp_enum_ty : int -> PPrintEngine.document

val pp_int : int -> PPrintEngine.document

val pp_bvi : int -> PPrintEngine.document

val pp_bv : string -> PPrintEngine.document

val pp_str : string -> PPrintEngine.document

val pp_j : int -> string

val pp_mem_size : AstGen.Def.Size.t -> Utils.Pp.document

val pp_binmem : AstGen.Def.Size.t AstGen.Ott.binmem -> PPrintEngine.document

val pp_var : ('a -> PPrintEngine.document) -> 'a -> PPrintEngine.document

val pp_bbvar : string -> PPrintEngine.document

val pp_bind : ('a -> PPrintEngine.document) -> (string * ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.exp) -> PPrintEngine.document

val pp_exp : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.exp -> PPrintEngine.document

val pp_fbool : (string * bool) -> PPrintEngine.document

val pp_smt : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt -> PPrintEngine.document

val pp_smts : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt list -> PPrintEngine.document

val pp_smt_ans : ('a -> PPrintEngine.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt_ans -> PPrintEngine.document

val pp_ty : AstGen.Def.Size.t AstGen.Ott.ty -> PPrintEngine.document

val pp_bool : bool -> PPrintEngine.document

val pp_unop : AstGen.Ott.unop -> PPrintEngine.document

val pp_bvarith : AstGen.Ott.bvarith -> PPrintEngine.document

val pp_bvcomp : AstGen.Ott.bvcomp -> PPrintEngine.document

val pp_binop : AstGen.Def.Size.t AstGen.Ott.binop -> PPrintEngine.document

val pp_bvmanyarith : AstGen.Ott.bvmanyarith -> PPrintEngine.document

val pp_manyop : AstGen.Ott.manyop -> PPrintEngine.document

val pp_loc : Stdlib.Lexing.position -> PPrint.document

Prints a lexing location

val pp_lrng : Isla_lang.AST.lrng -> PPrint.document

Prints a lexing range

Analysers

All function that start with is_

val is_atomic : ('a, 'b, 'c, 'd) exp -> bool

Destructors

Aparently I overestimated ocaml type-system in it's handling of empty types.

Here are some function to destroy empty types.

val destr_binmem : no binmem -> 'a

Expectors

Functions to assert that a specific constructor is used and get the value

val expect_bits : ('a, 'b, 'c, 'd) exp -> Utils.BitVec.t

val ty_expect_bv : 'a ty -> int

Construtors

val assert_smt : ('a, 'b, 'c, 'd) exp -> ('a, 'b, 'c, 'd) smt

val simplify_smt : ?⁠flags:(string * bool) list -> ('a, 'b, 'c, 'd) exp -> ('a, 'b, 'c, 'd) smt

Comparisons

val equal_exp : ?⁠annot:('a -> 'a -> bool) -> var:('b -> 'b -> bool) -> ?⁠bnd:('c -> 'c -> bool) -> ('a, 'b, 'c, 'd) exp -> ('a, 'b, 'c, 'd) exp -> bool

Equality for expression. Polymorphic equality will fail. I love boilerplate code and love Ocaml!