Module BranchTable

This module is designed to symbolically evaluate all the branch tables in an ELF file. Known bugs:

How to handle the following for live = x9 ? 40812c: 2a0803e9 mov w9, w8 408130: 2a0903e8 mov w8, w9 408134: 71000d08 subs w8, w8, #0x3

Multiple starts

module Analyses = BranchTable__.InstrsAnalyses

module Reg = State.Reg

module Instr = Trace.Instr

module Runner = Run.Runner

module Block = Run.Block_lib

module ConcreteEval = Exp.ConcreteEval

module Value = Exp.Value

val pp_regs : Reg.t list -> Stdlib.out_channel -> unit

val pp_ptrs : int list -> Stdlib.out_channel -> unit

type and_t =

| AndS of Reg.t * Utils.BitVec.t

| Mask of Reg.t * Reg.t * Utils.BitVec.t

val looks_like_ands : Instr.trace_meta -> Instr.trace_meta -> and_t option

val looks_like_and_low_mask : Reg.t -> ('a, Trace.Var.t, 'b, 'c) Ast.exp -> and_t option

val looks_like_and : Instr.trace_meta list -> and_t option

val looks_like_csel : Instr.trace_meta list -> bool

type cmp_case =

| RangeFrom0

| Eq_Neq

val fixup_written_read : (int -> Runner.slot option) -> int -> Instr.t -> (cmp_case * (State.Reg.t * Utils.BitVec.t)) option * Instr.Reg.t list * State.Reg.t list

Two things are 'fixed' for the reads and writes:

Reads of the PC and SP are removed since they are always live

Comparisons of a register with a constant are considered a write, because we set the value of the register with a range of constants before simulating.

type dataflow_context = {

instr_at : int -> Runner.slot option;

comes_before : int -> (int * int list) option;

spills_for : int -> int list;

depth_limit : int;

relevant_instrs : (int, unit) Stdlib.Hashtbl.t;

relevant_spills : (int, bool) Stdlib.Hashtbl.t;

live : Reg.t list;

cmp_operands : (cmp_case * (Reg.t * Utils.BitVec.t)) option;

}

exception DeadEnd

exception MultipleStarts

exception IrrelevantEarlyRange of cmp_case

val update_live : dataflow_context -> int -> int * (int, unit) Stdlib.Hashtbl.t * (cmp_case * (Reg.t * Utils.BitVec.t)) option

A standard backwards-dataflow analysis with following modifications:

comparisons with constants are treated as writes

irrelevant instructions (those which do not write to any live registers) are ignored

relevant instructions are tracked and returned

the operands of the first relevant comparison with a constant (which is followed by a conditional branch) are also returned, if it exists

val step_backwards : dataflow_context -> int -> int * (int, unit) Stdlib.Hashtbl.t * (cmp_case * (Reg.t * Utils.BitVec.t)) option

val reverse_dataflow : (int, Runner.slot) Stdlib.Hashtbl.t -> (int, int * int list) Stdlib.Hashtbl.t -> (int, int) Stdlib.Hashtbl.t Stdlib.Lazy.t -> Analyses.reg_branches -> (int * (int * (int, unit) Stdlib.Hashtbl.t * (cmp_case * (Reg.t * Utils.BitVec.t)) option)) list

Work backwards through the dataflow of the registers to find the earliest instruction from which point to run a straight-line simluation.

This function returns a tuple containing:

the address of the start of the branch-register target computation

the set of instruction addresses which are relevant between the start (inclusive) and the branch-register instruction (exclusive)

the register that must be set with a value before simulating the relevant instructions

the inclusive maximum value which that register should be set to

val make_block : Run.Runner.t -> (int * (int * (int, unit) Stdlib.Hashtbl.t * 'a)) -> int * ((unit -> Block.t) * (int, unit) Stdlib.Hashtbl.t * 'a)

type pc_exp_guess =

| Sym of Block.label State.Tree.t

| Ints of Block.label State.Tree.t array

val symeval_range_for_trees : Elf.File.t -> State.t -> ((unit -> Block.t) * (int, 'a) Stdlib.Hashtbl.t * (cmp_case * (State.Reg.t * Utils.BitVec.t)) option) -> pc_exp_guess

type 'a target =

| Sym of State.exp * (State.var * State.exp) list

| Conc of 'a list

val map_target : ('a -> 'b) -> 'a target -> 'b target

val exp_to_int : ?⁠ctxt:'a ConcreteEval.context -> ('b, 'a, Ast.no, Ast.no) Ast.exp -> int

val target_to_ints : int target -> int list

val int_or_sym : (State.t * (Ast.no Ast.ty, State.Exp.var, Ast.no, Ast.no) Ast.exp) -> int target

val pp_target : int target -> PPrintEngine.document

val extract_branch_targets : State.Reg.t -> pc_exp_guess -> int target

val process_sym : Utils.String.t Utils.List.t -> Dw.t -> State.t -> (int * string * int target) list

val get_branches : bool -> Utils.String.t Utils.List.t -> string -> (int * string * int target) list

val symeval_branch_table : string -> Utils.String.t Utils.List.t -> unit

val elf : string Cmdliner.Term.t

val funcs : string list Cmdliner.Term.t

val term : unit Cmdliner.Term.t

val info : Cmdliner.Term.info

val command : unit Cmdliner.Term.t * Cmdliner.Term.info