Module BranchTable__.InstrsAnalyses

type 'a branch =

| Simple of int * 'a

| Reg of State.Reg.t list

val looks_like_a_branch : Trace.Instr.t -> int -> bool branch option

An instruction looks like a branch if it is not a return and writes to the PC

type instructions = (int, Run.Runner.slot) Stdlib.Hashtbl.t

type simple_branches = private (int * bool * int) list

Triples of address of the branch, whether it is conditional (i.e. does it fall-through to the next instruction) and the target address

type reg_branches = private (int * State.Reg.t list) list

Pairs of the address of the branch and the register whose computed values is branched on.

val find_branches : instructions -> simple_branches * reg_branches

Find and split simple branches from register branches. If it is simple, we record whether a branch is conditional (and so can fall-through to the next instruction) or unconditional and its target address to help construct the (simple) comes-before relation later.

val compute_comes_before : instructions -> simple_branches -> (int, int * int list) Stdlib.Hashtbl.t

Compute (the one b to many as) comes-before relation. Address a comes-before address b iff a falls-through and/or branches to b. NOTE: The graph of this relation can be be cyclic. NOTE: It ignores the fact that BLR instructions effectively fall-through.

val track_spills : instructions -> simple_branches -> (int, int) Stdlib.Hashtbl.t

Track register spilling and reloading. Returns a map (int, int) Hashtbl.t where the key is the address of instructions that load from a stack-offset that was spilled-to earlier and the value is the address of that earlier spill. NOTE: This is an over-approximation: it will track all register spills but may also include standard stack/local variable manipulation. NOTE: Does NOT do/use any sort of pointer-aliasing analysis, just syntactic pattern-matching.