Module Config

module type S = sig ... end

This is the compile-time configuration module. It is pulled from the root config.ml file or will fallback to the default_config.ml file

include S

val config_file : string

Default config file

val z3 : string

Default Z3 command. (Can be overidden with Z3_PATH and --z3)

val isla_client : string

Default Isla command. (Can be overidden with ISLA_CLIENT and --isla)

val arch : string

The default architecture to pick when none is specified

val enable_backtrace : bool

Whether to enable backtraces or not

val enable_tests : bool

Whether to enable internal unit test and the test test sub-command

module Arch : sig ... end

This module provides an enumeration of architecture for internal identification

module File : sig ... end

This module is to handle the configuration file of the program.

module CommonOpt : sig ... end

This module provide support for common command line option to be used across multiple subcomands.