Module Config.CommonOpt

This module provide support for common command line option to be used across multiple subcomands.

It also provide some utilities on the command line.

Exit codes

val exits : Cmdliner.Cmd.Exit.info list

Description of read-dwarf exit codes. May need to be updated

Config options

This section is to manage the configuration file. See Config.File to see how the configuration file works

Passing --config=FILE on the CLI will setup the Config.File module to load that file as the configuration file.

For test executables, you will have to call File.ensure_loaded directly.

val config : unit Cmdliner.Term.t

val arch_val : 'a -> Arch.t option -> Arch.t

val arch_opt : Arch.t option Cmdliner.Term.t

val arch : Arch.t Cmdliner.Term.t

Isla options

val isla_client_ref : string Stdlib.ref

The isla_client path

val isla_client : unit Cmdliner.Term.t

Z3 options

val z3_ref : string Stdlib.ref

The z3 command

val z3 : unit Cmdliner.Term.t

The z3 option

Common option list

val comopts : unit Cmdliner.Term.t list

The list of common options. Almost all sub-commands should use this.

val command_tuple_to_cmd_t : ('a Cmdliner.Term.t * Cmdliner.Cmd.info) -> 'a Cmdliner.Cmd.t

val grouped_exe : (unit Cmdliner.Term.t * Cmdliner.Cmd.info) list -> (unit Cmdliner.Term.t * Cmdliner.Cmd.info) -> 'a

val quick_exe : name:string -> doc:string -> (unit -> unit) -> 'a