Module Config.File

This module is to handle the configuration file of the program.

For now I expect a TOML format and use the toml library from opam. The data structure themselves are format agnostic so if we change the TOML library of change of format, only this file must be modified. For example to JSON or YAML.

The top level usage is to first set the file reference which will be done by CommonOpt.config using ensure_loaded. Then the config can be accessed with the various Accessors

Configuration structure

All field of records in this section have a matching line in the checked-in config.toml. If something is optional here, and should be disabled in the default config.toml, put a commented line there, so that the correspondence can be immediately seen.

module ArchConf : sig ... end

module Z3 : sig ... end

type archs_type = (Arch.t, ArchConf.t) Stdlib.Hashtbl.t

Each architecture specific configuration is specified by a TOML section represented by a Arch.t record. They are stored in this type

type t = {

arch : Arch.t option;
(*
The default architecture to be choosen when no ELF file is specified on the CLI. This is optional, but if supplied, it will override Config.arch
*)

archs : archs_type;
(*
All the architecture specific configurations
*)

z3 : Z3.t;
(*
The Z3 configuration
*)

}

File loading

This section is about loading the configuration from a TOML file

val load : string -> unit

Load the configuration file from the specified file name

val ensure_loaded : string -> unit

If the configuration file is not already loaded, load it from the specified file.

Accessors

The section provide various accessor to access directly part of the configuration

The config must be loaded otherwise UnloadedConfig will be thrown

exception UnloadedConfig

This exception will be raised if any of the accessor is called while the config is not loaded.

val get_config : unit -> t

Get all the configuration information

val get_arch_name : unit -> Arch.t

Get the default architecture name to be used if no architecture is implicitly or explicitly specified on the CLI

val get_arch_config : Arch.t -> ArchConf.t

Get the architecture configuration for a specific architecture. To get the architecture configuration of the currently enabled architecture, Call Arch.get_config

val get_isla_config : Arch.t -> ArchConf.Isla.t

Get the isla configuration for a specific architecture. To get the isla configuration of the currently enabled architecture, Call Arch.get_isla_config

val get_z3_config : unit -> Z3.t

Get the Z3 configuration