Module ArchConf.Isla

This module provides the current isla configuration.

type t = {

ignored_regs : string list;
(*
The list of register to be ignored
*)

arch_file : string;
(*
The name of the architecture.ir file to use. This is the file that was compiled with isla-sail from the sail source. Alternatively this file can be found in the isla-snapshots respository.
*)

arch_toml : string;
(*
The name of the architecture.toml file to use with Isla.
*)

arch_file_digest : string;
(*
The digest of the file in arch_file. This will be used instead of arch_file by digest.
*)

linearize : string list;
(*
List of sail function to be linearized. That means that if there is a control-flow branching in that function, instead of generating multiple traces, isla will generate a single trace with it-then-else expression
*)

other_opts : string list;
(*
List of other option to pass to isla. Ideally if an option is to be easily used, a new field of t should be created, but for quick and dirty testing, this field can be used.
*)

}

Isla configuration option. Everything in here is salient for cache coherency which means that if any option if changed here, the whole Isla.Cache is invalidated. This is checked with digest.

val digest : t -> string

Produces a digest of the Isla configuration for invalidating the Isla.Cache when some configuration parameter changes