Module type Config.S

This is the compile-time configuration module. It is pulled from the root config.ml file or will fallback to the default_config.ml file

All compile-time configuration options are specified here. For runtime configuration, look at Config.File.

val config_file : string

Default config file

val z3 : string

Default Z3 command. (Can be overidden with Z3_PATH and --z3)

val isla_client : string

Default Isla command. (Can be overidden with ISLA_CLIENT and --isla)

val arch : string

The default architecture to pick when none is specified

val enable_backtrace : bool

Whether to enable backtraces or not

val enable_tests : bool

Whether to enable internal unit test and the test test sub-command