Module Ctype

This module provides the internal C-like type system. This type system is slightly different than the normal C type system. This module only provides the Ocaml datastructure to represent those types. The typing rules are implemented in Trace.Typer, where they are applied dire

These types follow the normal C type structure except for pointers that are more complex. To handle the fact that the C compiler knows perfectly the ABI and is "allowed" to used it, we have to make pointer types resist manual adjusting of pointers to point to the field of a struct. However the new pointer cannot just have type field_type* because one could want to get back a pointer to the whole structure by subtracting an offset from the field pointer. Thus a pointer must never forget information about its surroundings. Those surroundings are called a fragment and represent all the type information of the fragment of memory in which a pointer lies. The pointer is thus represented as a fragment of memory and an offset. Since the pointer type is more complex and packs more information, the surface syntax has changed. A pointer type is written between braces, so A* becomes {A}, but in more complex cases, all informations fit between the braces.

Furthermore, the fragment part of the pointer does not record any information about aliasing: two different type fragments are perfectly allowed to alias. To handle non-aliasing properties, like the stack not aliasing the heap or restrict pointers, pointers also have a provenance field. See State.Mem.

There is another problem: The C language does not define a C type system for the whole program, contrary to C++. It defines only a type system per compilation unit. This limitation is too annoying to work with so the module implements some kind of linking of types, similar to C++ rules. See C type linking: From Linksem.

DWARF constants

val vDW_ATE_address : int

val vDW_ATE_boolean : int

val vDW_ATE_signed : int

val vDW_ATE_signed_char : int

val vDW_ATE_unsigned : int

val vDW_ATE_unsigned_char : int

Types

type provenance =

| Restricted of int

| Main

This is the provenance of the pointer. This tells to which symbolic memory block a pointer points to. To get the full explanation, go to State.Mem

val pp_provenance : provenance -> Utils.Pp.document

type unqualified =

| Machine of int
(*
Size in bytes for now
*)

| Cint of {

name : string;

signed : bool;

size : int;

ischar : bool;

}

| Cbool

| Ptr of {

fragment : fragment;

offset : offset;

provenance : provenance;

}
(*
fragment is about a type fragment. offset is the position in that type fragment. provenance is about a runtime symbolic block (see State.Mem).
*)

| Struct of {

name : string;

size : int;

id : int;

}
(*
See env for what the id refers to. The int is the size
*)

| Array of {

elem : t;

dims : int option list;

}

| Enum of {

name : string;

id : int;

}
(*
See env for what the id refers to
*)

| FuncPtr
(*
Hack to accommodate PKVM
*)

| Missing
(*
Hack to accommodate PKVM
*)

| Bits
(*
Hack to prevent losing type information when processing bitvectors with non-whole-byte sizes
*)

The unqualified part of the C type without const volatile, ...

and t = {

unqualified : unqualified;

const : bool;

volatile : bool;

restrict : bool;

constexpr : bool;

}

The internal representation of generalized C types

and fragment =

| Unknown
(*
Unknown type, But without possibility of learning
*)

| Single of t
(*
Single object: Only when accessing of a global variable
*)

| DynArray of t
(*
Generic C pointer, may point to multiple element of that type
*)

| DynFragment of int
(*
Writable fragment for memory whose type is changing dynamically
*)

| Global of string
(*
The Global fragment that contains all the fixed ELF section .text, .data, .rodata, ...
*)

The type of a memory fragment

and offset =

| Const of int
(*
Constant offset
*)

| Somewhere

The type of an offset in a fragment

type linksem_t = Dwarf.c_type

The type of C types in Linksem

type field = {

fname : string option;

offset : int;

typ : t;

size : int;

}

A field in a structure

type linksem_field = linksem_t Dwarf.struct_union_member

module FieldMap : Utils.RngMap.S with type obj = field

A range map over field to represent a structure layout

type struc = {

layout : FieldMap.t;

name : string;

size : int;

complete : bool;

}

The type of a C structure.

A structure can be complete or incomplete but due to some internal hackery this is a need for a subtle difference with C: Even incomplete structure have a size, they just don't have any field.

A incomplete struct can and will often be complete later as the interpretation of DWARF information advances.

The name field is the linking name of the struct.

type enum = {

name : string;

labels : (int, string option) Stdlib.Hashtbl.t;

}

The representation type of a C enumeration

type cupdie_id = int * int

The identifier for a linksem_cupdie. See ids_of_cupdie

type linksem_env = linksem_t list

The type of environement linksem gives us.

Only structs, enums and unions can appear. A type can appear multiple times and also be forward-declared with missing data like size. env_of_linksem hopefully deals with all those problems

type linksem_indexed_env = (cupdie_id, linksem_t) Stdlib.Hashtbl.t

An version of the linksem environement indexed by cupdie_id

type env = {

structs : (string, struc) Utils.IdMap.t;

enums : (string, enum) Utils.IdMap.t;

lenv : linksem_indexed_env;

}

The type environment that contain mapping from linking name and a generated id

to the actual content of structs and enumerations.

Linking names can be:

A plain name for a struct/enum declared with a tag.

typedef.name for an unnamed struct declared in a typedef

outer.member for unnamed struct used as the type of a member of a struct with linking name outer.

As an unnamed struct can be declared in the linksem environement but used with the typedef only after the initial environement setup of env_of_linksem, the original linksem type must be kept alive in lenv

val ptr_size : int

The size of pointer. This will be configurable later

val enum_size : int

The size of enums. This will be configurable later

Comparisons

val equal : t -> t -> bool

Convenience manipulation

val is_struct : t -> bool

val is_array : t -> bool

val is_ptr : t -> bool

val is_scalar : t -> bool

val is_composite : t -> bool

val is_constexpr : t -> bool

val ptr : t -> unqualified

Make a simple pointer from a type

val voidstar : unqualified

A void* pointer

val qual : ?const:bool -> ?volatile:bool -> ?restrict:bool -> ?constexpr:bool -> unqualified -> t

Create a qualified type from an unqualified type with the specified qualifiers

val add_qual : ?const:bool -> ?volatile:bool -> ?restrict:bool -> ?constexpr:bool -> t -> t

update specified qualifiers. Other qualifier are kept

val machine : ?constexpr:bool -> int -> t

Create a machine type of that size without qualifiers

val of_frag : ?provenance:provenance -> ?offset:int -> ?constexpr:bool -> ?restrict:bool -> fragment -> t

Create a pointer to fragment with specified offset (0 by default)

val of_frag_somewhere : ?provenance:provenance -> ?constexpr:bool -> ?restrict:bool -> fragment -> t

Create a pointer to fragment Somewhere

val incomplete_struct : string -> int -> struc

Build an incomplete struct with a linking name and a size

val make_env : linksem_indexed_env -> env

Makes an empty environement from and indexed linksem environement

val offset_update : offset -> int -> offset

Update an offset with an integer update

val ptr_update : t -> int -> t

Update an pointer with an integer update

val ptr_set : t -> int -> t

Set a arbitrary new offset in a pointer

val ptr_forget : t -> t

Make a pointer forget it's offset

Sizeof

This section give implementation of sizeof function.

Dynamic array have size 0 until we are able to deal with C99 last member dynamic arrays. This will mess up with type_at. TODO fix it.

val sizeof_unqualified : unqualified -> int

Give the size of an unqualified type. Need the environement.

val sizeof : t -> int

Give the size of an type. Need the environement.

val len : t -> int

For being used in RngMap.LenObject

C type linking: From Linksem

This section contain the whole hierarchy of function used to convert type from DWARF representation to the internal type system.

During this conversion, C type linking happens. This means that structure with the same name from different compilation unit are identified as being the same. If they do not have the same layout, an error is raised. This would mean that either the C program does very weird thing with it's type that we don't want to verify or more likely that it is ill-formed. As we have to deal with anonymous struct, name are a bit more complex than plain struct tags. See env for a description.

The top-level interface for types is of_linksem and the top-level interface for environement is env_of_linksem. Those will be called by Dw at DWARF loading time to generate a coherent type system.

type conversion_context = {

env : env;

potential_link_name : string option;

}

This type is a conversion context.

Its role is to contain all the things that all the functions in this section will need to convert types.

val ids_of_cupdie : Dwarf.cupdie -> cupdie_id

Get the id of a linksem cupdie

val pp_decl : Dwarf.decl -> Utils.Pp.document

Pretty print the dwarf decl type

TODO: Move that in the appropriate place

exception LinkError

This exception is raised when the type we are trying to reach must came from another translation unit or later in the current one.

The information is incomplete at this moment to create is.

Normally this exception should only happen during the initial env_of_linksem. If it happens elsewhere, either the code used an anonymous struct that do not have C++-like linkage or the compiler did not do its job.

val expect_some_link : 'a option -> 'a

val base_type_of_linksem : ?size:Sym_ocaml.Num.t -> encoding:Sym_ocaml.Num.t -> string -> unqualified

Convert a base type with a name and encoding and maybe a size to its inner representation

Only integers, chars and bools supported. No floating points

val field_of_linksem : cc:conversion_context -> linksem_field -> FieldMap.obj

val field_map_of_linksem : string -> cc:conversion_context -> linksem_field list -> FieldMap.t

val struc_of_linksem : cc:conversion_context -> string -> int -> linksem_field list -> struc

val struct_type_of_linksem : ?force_complete:bool -> cc:conversion_context -> cupdie:Dwarf.cupdie -> mname:string option -> decl:Dwarf.decl -> unqualified

Build a struct from it's cupdie and name. If force_complete is true and the struct is incomplete. It will try to complete is using cupdie and throw LinkError if it fails.

val enum_of_linksem : cc:conversion_context -> string -> Dwarf.enumeration_member list -> enum

val enum_type_of_linksem : cc:conversion_context -> cupdie:Dwarf.cupdie -> mname:string option -> decl:Dwarf.decl -> unqualified

val unqualified_of_linksem : ?force_complete:bool -> cc:conversion_context -> linksem_t -> unqualified

Convert an unqualified type. Union are just Machine n where n is their size

val of_linksem_none : unit -> t

Placeholder for whatever the right thing to do is when linksem found no type - TODO: fix

val of_linksem_cc : ?force_complete:bool -> cc:conversion_context -> ?const:bool -> ?volatile:bool -> ?restrict:bool -> linksem_t -> t

The main of_linksem that take a full conversion_context and qualifiers.

The user friendly version is of_linksem

See struct_type_of_linksem for the explanation of force_complete.

All the qualifier passed as parameter are added to the resulting type.

val of_linksem : env:env -> linksem_t -> t

The user friendly interface that convert a type using the environment. This useful when there is no additional linking information to pass on

val env_of_linksem : linksem_env -> env

The main environment conversion function.

This the function that deal with all the type linking process, forward declaration an all that stuff.

First it create the indexed linksem environment (member lenv of env),

Then it registers all named structs as incomplete in the environment (to deal with self recursion, only named structs can self-recurse).

Finally it run of_linksem_cc on all the types with force_complete on. During this phase it ignore all LinkError that arise. It assumes that if some thing was incomplete at that point, it will become complete later.

Then it can return the freshly built environment.

If some structs are still incomplete, but are actually used, the LinkError will be raised at the time of use.

Pretty printing

val pp_signed : bool -> Utils.Pp.document

val pp : t -> Utils.Pp.document

Pretty print a type. If an environement is provided, structs and enums will be printed with a name, otherwise they will just have a number

val pp_unqualified : unqualified -> Utils.Pp.document

val pp_fragment : fragment -> Utils.Pp.document

val pp_offset : offset -> Utils.Pp.document

val pp_arr : t -> int option list -> Utils.Pp.document

val pp_arr_dim : int option -> Utils.Pp.document

val pp_field : field -> Utils.Pp.document

val pp_struct : struc -> Utils.Pp.document

val pp_enums : enum -> Utils.Pp.document

val pp_env : env -> Utils.Pp.document

Print the whole environement (not the linksem indexed environment)

Type at

This section is about getting the type at a specific offset in things.

val struct_at : env:env -> size:int -> struc -> int -> t Utils.Option.t

Same as type_at but for structs

val type_at : env:env -> size:int -> t -> int -> t Utils.Option.t

Get the type of size size at the provided offset in another type