Module Dw

This module provides the specifically interpreted DWARF information needed for read-dwarf operations

I would have called this module dwarf but linksem decided that is was a good idea to dump all its modules in the global namespace.

module Var : sig ... end

This module contain all the definition to handle local and global variables as defined in the DWARF information of the target file

module Func : sig ... end

This module contain all the definition to handle functions as defined in the DWARF information of the target file

module Loc : sig ... end

This module represent architectural locations. Locations in DWARF information are represented as a little stack language (represented as dwop list) which describes how to compute the target value from the current concrete state. This computation can be arbitrarily complex, for example take half of the value from the high bits of a register and then the other half from memory at a specific address coming from another register.

type t = {

elf : Elf.File.t;

ldwarf : Dwarf.dwarf;

ldwarf_sdt : Dwarf.sdt_dwarf;

funcs : (string, Func.t) Stdlib.Hashtbl.t;

vars : (string, Var.t) Stdlib.Hashtbl.t;

tenv : Ctype.env;

}

The type that represent a elf-dwarf binary whose information has been fully interpreted

exception DwarfError of string

Error on Dwarf parsing

val dwarferror : ('a, unit, string, 'b) Stdlib.format4 -> 'a

Throw a DwarfError

val of_elf : Elf.File.t -> t

Get Dwarf information from an Elf file.

May raise a DwarfError if a problem occurs.

val of_file : string -> t

Get Dwarf information from an Elf file by name. Use Elf.File.of_file

val get_func_opt : name:string -> t -> Func.t option

Get a function by name

val pp_raw : t -> Utils.Pp.document

Pretty print dwarf data as an ocaml structure