Module Dw.Func

This module contain all the definition to handle functions as defined in the DWARF information of the target file

type func = {

name : string;

scope : scope;

ret : Ctype.t option;

}

Type of a dwarf function that may or may not be inlined

If this type stand on its own, then it is inlined. If it is inside a t then it's a top level function. There is no separate type for inline functions because they do not have any special data that a top level function may not have

and scope = {

vars : Var.t list;

funcs : func list;

scopes : scope list;

}

Type of a dwarf scope that may contain recursively other data. The lists here have the semantic meaning of sets: the order is irrelevant.

type linksem_func = Dwarf.sdt_subroutine

This is the type a dwarf function in linksem

type linksem_scope = Dwarf.sdt_lexical_block

This is the type a dwarf scope in linksem

val func_of_linksem : Elf.File.t -> Ctype.env -> linksem_func -> func

Create a Dwarf function from its linksem counterpart

val scope_of_linksem : Elf.File.t -> Ctype.env -> linksem_scope -> scope

Create and Dwarf scope from its linksem counterpart

val func_get_api : func -> Arch.func_api

Get the API of the function

val pp_raw_func : func -> Utils.Pp.document

val pp_raw_scope : scope -> Utils.Pp.document

type t = {

sym : Elf.Symbol.t option;

func : func;

}

This the type of a top-level function. It may have an associated elf symbol

This type will contain all necessary indexes for function-wide fast access to relevant dwarf information.

type linksem_t = Dwarf.sdt_subroutine

This is the type a dwarf top_level function in linksem

val of_linksem : Elf.File.t -> Ctype.env -> linksem_t -> t

Create a dwarf top level function from its Linksem counterpart. The ELF file is to get a potential matching symbol. For now the matching is made with name and then code address, but perhaps the code using this should be rephrased in terms of addresses instead of symbols, to be more resilient

val get_api : t -> Arch.func_api

Get the API of a top level function

val pp_raw : t -> Utils.Pp.document

Pretty print a raw top level function