Module Dw.Loc

This module represent architectural locations. Locations in DWARF information are represented as a little stack language (represented as dwop list) which describes how to compute the target value from the current concrete state. This computation can be arbitrarily complex, for example take half of the value from the high bits of a register and then the other half from memory at a specific address coming from another register.

Those expression are not directly usable to guide the type inference, as we would much more like simple direct information like "This value is in that register". To solve this, there is a custom type t that interpret simple static locations directly and leaves more complex location uninterpreted.

This interpretation is currently very basic. It remains to be seen if it actually need to be improved.

type dwop = Dwarf.operation

The type of a dwarf location stack operation

type t =

| Register of State.Reg.t
(*
In the register
*)

| RegisterOffset of State.Reg.t * int
(*
At register + offset address
*)

| StackFrame of int
(*
On the stackFrame with offset
*)

| Global of Elf.SymTable.sym_offset
(*
Global variable with an offset
*)

| Const of Utils.Sym.t

| Dwarf of dwop list
(*
Uninterpreted dwarf location
*)

The type of a location, as static as possible

type linksem_t = dwop list

The type of a location in linksem format

val vDW_OP_addr : int

The integer value of the DW_OP_addr constant in DWARF standard

TODO this should come from LinkSem's dwarf

val vDW_OP_reg0 : int

The integer value of the DW_OP_reg0 constant in DWARF standard

val vDW_OP_breg0 : int

The integer value of the DW_OP_breg0 constant in DWARF standard

val of_linksem : ?amap:Arch.dwarf_reg_map -> Elf.File.t -> linksem_t -> t

Convert a linksem location description into a Loc.t

Very naive for now : If the list has a single element that we can translate directly, we do. Otherwise, we dump it into the t.Dwarf constructor

val to_string : t -> string

Convert the location to a string. This is not reversible

val compare : t -> t -> int

Compare two location. Loc.t is not compatible with polymorphic compare

val pp : t -> Utils.Pp.document

Pretty-print the location