Module Dw.Var

This module contain all the definition to handle local and global variables as defined in the DWARF information of the target file

type range = Elf.Address.t * Elf.Address.t option

type t = {

name : string;

param : bool;

ctype : Ctype.t;

locs : (range * Loc.t) list;

locs_frame_base : (range * Loc.t) list;

}

Type of a DWARF variable

type linksem_t = Dwarf.sdt_variable_or_formal_parameter

Type of a DWARF variable in linksem

val loc_merge : (('a * 'a option) * Loc.t) list -> (('a * 'a option) * Loc.t) list

Merge contiguous location lists

val end_addr_of_sym : Utils.Sym.t -> Elf.Address.t option

val of_linksem : Elf.File.t -> Ctype.env -> linksem_t -> t

Create a DWARF variable from its linksem counterpart

val pp_raw : t -> Utils.Pp.document

Pretty print a variable