Module Dw__.Loc

type dwop = Dwarf.operation

The type of a dwarf location stack operation

type t =

| Register of State.Reg.t
In the register

| RegisterOffset of State.Reg.t * int
At register + offset address

| StackFrame of int
On the stackFrame with offset

| Global of Elf.SymTable.sym_offset
Global variable with an offset

| Dwarf of dwop list
Uninterpreted dwarf location

The type of a location, as static as possible

type linksem_t = dwop list

The type of a location in linksem format

val vDW_OP_addr : int

The integer value of the DW_OP_addr constant in DWARF standard

TODO this should come from LinkSem's dwarf

val vDW_OP_reg0 : int

The integer value of the DW_OP_reg0 constant in DWARF standard

val vDW_OP_breg0 : int

The integer value of the DW_OP_breg0 constant in DWARF standard

val of_linksem : ?⁠amap:Arch.dwarf_reg_map -> Elf.File.t -> linksem_t -> t

Convert a linksem location description into a Loc.t

Very naive for now : If the list has a single element that we can translate directly, we do. Otherwise, we dump it into the t.Dwarf constructor

val to_string : t -> string

Convert the location to a string. This is not reversible

val compare : t -> t -> int

Compare two location. Loc.t is not compatible with polymorphic compare

val pp : t -> Utils.Pp.document

Pretty-print the location