Module Elf

module Address : sig ... end

module File : sig ... end

This module represent an ELF 64 file. We do not deal with 32 bit ELF files for now The main interesting information is the symbol table.

module LinksemRelocatable : sig ... end

module Relocations : sig ... end

module Segment : sig ... end

The goal of this module is to represent a segment as loaded in memory. In particular, all information about file layout is intentionally lost I use basic ints for speed. It it fails for some reason, I'll move to int64s.

module SymTable : sig ... end

The module provide a type to represent a symbol table.

module Symbol : sig ... end

This module represent an Elf symbol. One important difference with linksem symbols is that the symbols of this module always have the corresponding data (code or initial value). That's why function like of_linksem_with_data exist.