Module Elf.File

This module represent an ELF 64 file. We do not deal with 32 bit ELF files for now The main interesting information is the symbol table.

module SymTbl = SymTable

type machine =

| Supp of Config.Arch.t

| Other of int

The machine type of the ELF file. It can be a known architecture, or one that is not in Arch

val machine_of_linksem : Nat_big_num.num -> machine

Convert the linksem machine number to machine

val machine_to_string : machine -> string

Convert a machine to a human readable string

val pp_machine : machine -> Utils.Pp.document

Pretty prints a machine

module SMap : sig ... end

type section = {

name : string;

size : int;

align : int;

}

type t = {

filename : string;
(*
The name on the file system. Useful for error messages
*)

symbols : SymTbl.t;
(*
The symbol table
*)

entry : int;
(*
The address of the entry point; only used in dumpSym.ml
*)

machine : machine;
(*
The target architecture of the file; only used in arch.ml, dumpSym.ml, dw.ml
*)

linksem : Elf_file.elf_file;
(*
The original linksem structure for the file; only used in dw.ml
*)

rodata : Segment.t SMap.t;
(*
The read-only data sections
*)

sections : section list;

}

The type containing all the information about an ELF file

exception ElfError of string

Error on Elf parsing

val elferror : ('a, unit, string, 'b) Stdlib.format4 -> 'a

Throw an ElfError

val of_file : string -> t

Parse an ELF file to create an Elf.File.t using Linksem.

May raise an ElfError