Module Elf__Segment

The goal of this module is to represent a segment as loaded in memory. In particular, all information about file layout is intentionally lost I use basic ints for speed. It it fails for some reason, I'll move to int64s.

This is basically a BytesSeq with metadata.

type t = {

data : Utils.BytesSeq.t;

addr : int;
The actual start address of the BytesSeq

size : int;
redundant with BytesSeq.length data

read : bool;

write : bool;

execute : bool;

}

The type of a segment

val of_linksem : Elf_interpreted_segment.elf64_interpreted_segment -> t

Loads a t using a linksem interpreted segment.

val is_in : t -> int -> bool

Check if an address is inside a segment

val get_addr : (Utils.BytesSeq.t -> int -> 'a) -> t -> int -> 'a

Get a value at an address which is in this segment using the getter provided

val get_addr_list_opt : (Utils.BytesSeq.t -> int -> 'a) -> t list -> int -> 'a option

Get a value at an address which is one of the segment of this list. It must be entirely in one of the segment

val get_containing : t list -> int -> t option

Get the segment containing an address, among a list of them or None