Module Elf__SymTable

The module provide a type to represent a symbol table.

The interesting operations provided are fetching symbol by name and knowing which symbol owns a specific address

of_position_string provides a convenient way of describing a position in the ELF file from a human text input like the CLI.

type sym = Elf.Symbol.t

type linksem_sym = Elf.Symbol.linksem_t

type sym_offset = sym * int

The type of a symbol with offset

type linksem_t = Elf_file.global_symbol_init_info

type t

The type of a symbol table.

val empty : t

The empty symbol table

val add : t -> sym -> t

Return a new table with the symbol added If there already exists a symbol covering the same area, merge them with Symbol.other_names

val of_name : t -> string -> sym

Get a symbol by name. Raise Not_found if no name matches

val of_name_opt : t -> string -> sym option

Get a symbol by name, None if no name matches

val of_addr : t -> int -> sym

Get the symbol owning that address. Not_found is raised if no symbol own that address.data See of_addr_opt

val of_addr_opt : t -> int -> sym option

Get the symbol owning that address. None if no symbol own that address. See of_addr

val of_addr_with_offset : t -> int -> sym_offset

Get a symbol with the offset that correspond to that address

val of_addr_with_offset_opt : t -> int -> sym_offset option

Get a symbol with the offset that correspond to that address

val to_addr_offset : sym_offset -> int

Get back the raw address from a symbol+offset value

val string_of_sym_offset : sym_offset -> string

Transform a symbol + offset into the corresponding string

val sym_offset_of_string : t -> string -> sym_offset

Transform a symbol + offset string into the actual symbol and the integer offset

val of_position_string : t -> string -> sym_offset

Convert a position string to a symbol + offset

A position string is a string describing a position in an ELF file. Two format are accepted for now:

A raw address of the form "0x40cafe"

A symbol name with optional offset like "sym" or "sym+4" or "sym+0x4"

val of_linksem : Elf.Segment.t list -> linksem_t -> t

Extract the symbol from the linksem symbol representation.

Need the segments for filling the missing symbol data

val pp_raw : t -> Utils.Pp.document

Pretty print the table as a raw ocaml value

val iter : t -> (sym -> unit) -> unit

Iterate through all the symbols in the table.

val fold : t -> 'a -> (sym -> 'a -> 'a) -> 'a

Fold over all the symbols in the table.