Module Elf__Symbol

This module represent an Elf symbol. One important difference with linksem symbols is that the symbols of this module always have the corresponding data (code or initial value). That's why function like of_linksem_with_data exist.

For now addresses are in ints and assume the top bit is sign extended. It may become Int64.t if required

type typ =

| NOTYPE

| OBJECT

| FUNC

| SECTION

| FILE

| UNKNOWN

The type of the ELF symbol

type linksem_typ = Z.t

type t = {

name : string;

other_names : string list;

typ : typ;

addr : int;

size : int;

writable : bool;

data : Utils.BytesSeq.t;

}

The ELF symbol. This type guarantee the data exists contrary to linksem symbols (it may be all zeros though)

type linksem_t = string * (Z.t * Z.t * Z.t * Utils.BytesSeq.t option * Z.t)

The type of an ELF symbol in linksem. See of_linksem

val push_name : string -> t -> t

Add a name to the other names list

val is_in : t -> int -> bool

Check if an address is in a symbol

val len : t -> int

For conformance with the RngMap.LenObject module type

val typ_of_linksem : linksem_typ -> typ

Convert the integer type into typ

val linksem_typ : linksem_t -> linksem_typ

Get the type from the linksem symbol type

exception LoadingError of string * int

LoadingError(name,addr) means that symbol name at addr could not be loaded.

val of_linksem : Elf.Segment.t list -> linksem_t -> t

Convert a symbol from linksem to read-dwarf representation using the segment data

May raise LoadingError when the symbol has no data and the data cannot be found in the segments

val is_interesting : typ -> bool

Tell if a symbol type is interesting for readDwarf purposes

val is_interesting_linksem : linksem_t -> bool

Tell if a linksem symbol is interesting for readDwarf purposes

val sub : t -> int -> int -> Utils.BytesSeq.t

Take the BytesSeq.t corresponding to the offset and length

val compare : t -> t -> int

Starting address comparison

val pp_typ : typ -> Utils.Pp.document

Pretty prints a symbol type

val pp_raw : t -> Utils.Pp.document

Raw pretty printing of a symbol