Module Exp.PpExp

This module provides a human readable pretty printing for Ast expressions

If you don't want to bother with details, use pp_exp and don't read the rest.

The precedence are the ones from C/C++ and Ocaml with some tweaks. In particular the precedence between bitwise operation and arithmetic operation are separated, so parenthesis will always be required between them.

The order is:

Extraction and extension

Concatenation

unary minus, unary bitwise negation, unary reduction

multiplications, divisions, separately shift

additions, substractions, separately bitwise operation

comparisons

equality

and

or

ifs

Unary operator cannot linebreak (but their content can)

Examples:

-a[1-3].2a:6[z+32] which is: (bvneg (concat ((_ extract 3 1) a) ((_ zero_extend 32) #b101010)))

Precedence

type prec =

| IF

| OR

| AND

| EQ

| COMP

| ADD

| MUL

| BITS

| SHIFTS

| UNARY

| CONCAT

| EXTRACT

| PARENS

The operators possible precedence

val compat : outer:prec -> inner:prec -> bool

Figure out if an expression of precedence inner in an expression of precedence outer needs parentheses

val prec_unop : Ast.unop -> prec

val prec_bvarith : Ast.bvarith -> prec

val prec_binop : Ast.no Ast.binop -> prec

val prec_bvmanyarith : Ast.bvmanyarith -> prec

val prec_manyop : Ast.manyop -> prec

Actual Pretty printing

val pp_bits : Utils.BitVec.t -> Utils.Pp.document

val ppnot : Utils.Pp.document

val pp_unop : Ast.unop -> Utils.Pp.document -> Utils.Pp.document

val sym_bvarith : Ast.bvarith -> Utils.Pp.document

val sym_bvcomp : Ast.bvcomp -> Utils.Pp.document

val pp_binop : Ast.no Ast.binop -> Utils.Pp.document -> Utils.Pp.document -> Utils.Pp.document

val sym_bvmanyarith : Ast.bvmanyarith -> Utils.Pp.document

val pp_manyop : Ast.manyop -> Utils.Pp.document list -> Utils.Pp.document

val pp_if : Utils.Pp.document -> Utils.Pp.document -> Utils.Pp.document -> Utils.Pp.document

val pp_exp_prec : ('v -> Utils.Pp.document) -> ('a, 'v, Ast.no, Ast.no) Ast.exp -> Utils.Pp.document * prec

Pretty print an expression and return its precedence

val pp_exp : ('a -> Utils.Pp.document) -> ('b, 'a, Ast.no, Ast.no) Ast.exp -> Utils.Pp.document

The main function for pretty printing an expression