Module Exp.Typed

This module provide operation on typed expressions i.e expressions whose annotations are their SMT type (Ast.ty).

type ('v, 'm) t = ('m Ast.ty, 'v, Ast.no, 'm) Ast.exp

The type for a typed expression.

The 'v type is the type of expression variable.

The 'm type should be either Ast.no if memory operation are disabled or Ast.Size.t if they are enabled.

The let bindings are always disabled.

Generic operation on types

val get_type : ('v, 'm) t -> 'm Ast.ty

Get the type of a typed expression. Specialized version of Ast.Manip.annot_exp

val is_bool : 'a Ast.ty -> bool

val is_bv : 'a Ast.ty -> bool

val is_enum : 'a Ast.ty -> bool

val expect_bool : 'a Ast.ty -> unit

val expect_bv : 'a Ast.ty -> int

val expect_enum : 'a Ast.ty -> int

Generic constructors

This constuctors build a new expression as required, and compute the new type. They assume the operation is well typed (They assert it).

val var : typ:'m Ast.ty -> 'v -> ('v, 'm) t

val bits : Utils.BitVec.t -> ('v, 'm) t

val bool : bool -> ('v, 'm) t

val enum : Ast.enum -> ('v, 'm) t

val vec_idx_type : 'a Ast.ty

This is completely arbitrary, because it is not currently used.

val vec : ('a Ast.ty, 'b, Ast.no, 'a) Ast.exp list -> ('a Ast.ty, 'b, Ast.no, 'a) Ast.exp

val unop : Ast.unop -> ('v, 'm) t -> ('v, 'm) t

val binop : 'm Ast.binop -> ('v, 'm) t -> ('v, 'm) t -> ('v, 'm) t

val manyop : AstGen.Ott.manyop -> ('v, 'm) t list -> ('v, 'm) t

In addition to well-typedness requirement, this function will throw Invalid_argument if the list is empty. If the list has a single element, it will just return that element instead of building the symbolic operation.

val ite : cond:('v, 'm) t -> ('v, 'm) t -> ('v, 'm) t -> ('v, 'm) t

Specific constructors

val bits_int : size:int -> int -> ('a, 'b) t

val bits_smt : string -> ('a, 'b) t

val zero : size:int -> ('a, 'b) t

val true_ : ('a, 'b) t

val false_ : ('a, 'b) t

val (+) : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

val sum : ('a, 'b) t list -> ('a, 'b) t

val sub : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

val (-) : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

val (*) : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

val prod : ('a, 'b) t list -> ('a, 'b) t

val sdiv : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

val not : ('a, 'b) t -> ('a, 'b) t

val neg : ('a, 'b) t -> ('a, 'b) t

val extract : first:int -> last:int -> ('a, 'b) t -> ('a, 'b) t

val eq : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

val (=) : ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

val concat : ('a, 'b) t list -> ('a, 'b) t

val comp : Ast.bvcomp -> ('a, 'b) t -> ('a, 'b) t -> ('a, 'b) t

Add types

val add_type : 'a 'v 'm. ty_of_var:('a -> 'v -> 'm Ast.ty) -> ('a, 'v, Ast.no, 'm) Ast.exp -> ('v, 'm) t

Replace the annotation of expression by the SMT types. The expression must be already well typed (you can trust the SMT solver on this one)

It will still assert that an expression is well typed as a side effect.

val is_well_typed : ('v, 'm) t -> bool

Check if an expression is well typed