Module Exp.Value

This module provide a type to represent concrete values.

There is not concrete value for memory yet, but maybe there should be at some point to be able to fully support concrete evaluation without external tool

type t =

| Bool of bool

| Enum of Ast.enum

| Bv of Utils.BitVec.t

| Vec of t list

The type of concrete values

val to_string : t -> string

String representation of concrete values

val pp : t -> Utils.Pp.document

Pretty printer for concrete values

val bool : bool -> t

Bool constructor

val enum : Ast.enum -> t

Enum constructor

val bv : Utils.BitVec.t -> t

Bv constructor

val vec : t list -> t

Vec constructor

val expect_bool : t -> bool

Extract a boolean or fail

val expect_enum : t -> Ast.enum

Extract an enumeration or fail

val expect_bv : ?size:int -> t -> Utils.BitVec.t

Extract a bit vector or fail. If size is specified, then it fail if the size don't match

val to_exp : t -> ('v, 'm) Typed.t

Convert to a constant expression