Module Isla

include module type of struct include Base end

Aliases

include module type of struct include Isla_lang.AST end

type enum = int * int

type lrng = Isla_lang__Isla_lang_ast.lrng =

| UnknownRng

| Generated of lrng

| Range of Stdlib.Lexing.position * Stdlib.Lexing.position

exception Parse_error_locn of lrng * string

val pp_lpos : Stdlib.Lexing.position -> PPrint.document

val pp_lrng : lrng -> PPrint.document

type accessor = Isla_lang__Isla_lang_ast.accessor =

| Field of string

type bvarith = Isla_lang__Isla_lang_ast.bvarith =

| Bvnand

| Bvnor

| Bvxnor

| Bvsub

| Bvudiv

| Bvudivi

| Bvsdiv

| Bvsdivi

| Bvurem

| Bvsrem

| Bvsmod

| Bvshl

| Bvlshr

| Bvashr

type bvcomp = Isla_lang__Isla_lang_ast.bvcomp =

| Bvult

| Bvslt

| Bvule

| Bvsle

| Bvuge

| Bvsge

| Bvugt

| Bvsgt

type bvmanyarith = Isla_lang__Isla_lang_ast.bvmanyarith =

| Bvand

| Bvor

| Bvxor

| Bvadd

| Bvmul

type accessor_list = Isla_lang__Isla_lang_ast.accessor_list =

| Nil

| Cons of accessor list

type unop = Isla_lang__Isla_lang_ast.unop =

| Not

| Bvnot

| Bvredand

| Bvredor

| Bvneg

| Extract of int * int

| ZeroExtend of int

| SignExtend of int

type binop = Isla_lang__Isla_lang_ast.binop =

| Eq

| Bvarith of bvarith

| Bvcomp of bvcomp

type manyop = Isla_lang__Isla_lang_ast.manyop =

| And

| Or

| Bvmanyarith of bvmanyarith

| Concat

type base_val = Isla_lang__Isla_lang_ast.base_val =

| Val_Symbolic of int

| Val_Bool of bool

| Val_Bits of string

| Val_Enum of enum

type assume_val = Isla_lang__Isla_lang_ast.assume_val =

| AVal_Var of string * accessor_list

| AVal_Bool of bool

| AVal_Bits of string

| AVal_Enum of enum

type ty = Isla_lang__Isla_lang_ast.ty =

| Ty_Bool

| Ty_BitVec of int

| Ty_Enum of int

| Ty_Array of ty * ty

type !'a exp = 'a Isla_lang__Isla_lang_ast.exp =

| Val of base_val * 'a

| Unop of unop * 'a exp * 'a

| Binop of binop * 'a exp * 'a exp * 'a

| Manyop of manyop * 'a exp list * 'a

| Ite of 'a exp * 'a exp * 'a exp * 'a

type valu = Isla_lang__Isla_lang_ast.valu =

| RegVal_Base of base_val

| RegVal_I of int * int

| RegVal_String of string

| RegVal_Unit

| RegVal_Vector of valu list

| RegVal_List of valu list

| RegVal_Struct of (string * valu) list

| RegVal_Constructor of string * valu

| RegVal_Poison

type !'a a_exp = 'a Isla_lang__Isla_lang_ast.a_exp =

| AExp_Val of assume_val * 'a

| AExp_Unop of unop * 'a a_exp * 'a

| AExp_Binop of binop * 'a a_exp * 'a a_exp * 'a

| AExp_Manyop of manyop * 'a a_exp list * 'a

| AExp_Ite of 'a a_exp * 'a a_exp * 'a a_exp * 'a

type !'a smt = 'a Isla_lang__Isla_lang_ast.smt =

| DeclareConst of int * ty

| DefineConst of int * 'a exp

| Assert of 'a exp

| DefineEnum of int

type arg_list = Isla_lang__Isla_lang_ast.arg_list =

| NilArgs

| ListArgs of valu list

type tag_value = valu option

type segment = Isla_lang__Isla_lang_ast.segment =

| Segment of string * int * int

type !'a event = 'a Isla_lang__Isla_lang_ast.event =

| Smt of 'a smt * 'a

| Branch of int * string * 'a

| ReadReg of string * accessor_list * valu * 'a

| WriteReg of string * accessor_list * valu * 'a

| ReadMem of valu * valu * valu * int * tag_value * 'a

| WriteMem of valu * valu * valu * valu * int * tag_value * 'a

| BranchAddress of valu * 'a

| Barrier of valu * 'a

| CacheOp of valu * valu * 'a

| MarkReg of string * string * 'a

| Cycle of 'a

| Instr of valu * 'a

| Sleeping of int * 'a

| WakeRequest of 'a

| SleepRequest of 'a

| AssumeReg of string * accessor_list * valu * 'a

| Assume of 'a a_exp * 'a

| FunAssume of string * valu * arg_list * 'a

| UseFunAssume of string * valu * arg_list * 'a

| AbstractCall of string * valu * arg_list * 'a

| AbstractPrimop of string * valu * arg_list * 'a

type instruction_segments = Isla_lang__Isla_lang_ast.instruction_segments =

| Segments of segment list

type !'a trc = 'a Isla_lang__Isla_lang_ast.trc =

| Trace of 'a event list

type !'a maybe_fork = 'a Isla_lang__Isla_lang_ast.maybe_fork =

| Cases of string * 'a tree_trc list

| End

and !'a tree_trc = 'a Isla_lang__Isla_lang_ast.tree_trc =

| TreeTrace of 'a event list * 'a maybe_fork

type !'a trcs = 'a Isla_lang__Isla_lang_ast.trcs =

| Traces of 'a trc list

| TracesWithSegments of instruction_segments * 'a trc list

type !'a whole_tree = 'a Isla_lang__Isla_lang_ast.whole_tree =

| BareTree of 'a tree_trc

| TreeWithSegments of instruction_segments * 'a tree_trc

val subst_val_base_val : base_val -> int -> base_val -> base_val

val subst_val_exp : base_val -> int -> 'a exp -> 'a exp

val subst_val_valu : base_val -> int -> valu -> valu

val subst_val_smt : base_val -> int -> 'a smt -> 'a smt

val subst_val_arg_list : base_val -> int -> arg_list -> arg_list

val subst_val_event : base_val -> int -> 'a event -> 'a event

val subst_val_tree_trc : base_val -> int -> 'a tree_trc -> 'a tree_trc

val subst_val_maybe_fork : base_val -> int -> 'a maybe_fork -> 'a maybe_fork

val subst_val_trc : base_val -> int -> 'a trc -> 'a trc

val subst_val_whole_tree : base_val -> int -> 'a whole_tree -> 'a whole_tree

val subst_val_trcs : base_val -> int -> 'a trcs -> 'a trcs

module Lexer = Base.Lexer

module Parser = Base.Parser

type loc = Stdlib.Lexing.position

type rtrc = lrng trc

The type of raw traces out of the parser

type revent = lrng event

The type of raw events out of the parser

type rsmt = lrng smt

The type of raw SMT declaration out of the parser

type rexp = lrng exp

The type of raw expressions out of the parser

type rtrcs = lrng trcs

IslaTrace parsing

exception ParseError of loc * string

Exception that represent an Isla parsing error

exception LexError of loc * string

Exception that represent an Isla lexing error

type lexer = Stdlib.Lexing.lexbuf -> Parser.token

type 'a parser = lexer -> Stdlib.Lexing.lexbuf -> 'a

val parse : 'a parser -> ?filename:string -> Stdlib.Lexing.lexbuf -> 'a

Parse a single Isla instruction output from a Lexing.lexbuf

val parse_exp : ?filename:string -> Stdlib.Lexing.lexbuf -> rexp

Parse a single Isla expression from a Lexing.lexbuf

val parse_exp_string : ?filename:string -> string -> rexp

Parse a single Isla expression from a string

val parse_exp_channel : ?filename:string -> Stdlib.in_channel -> rexp

Parse a single Isla expression from a channel

val parse_trc : ?filename:string -> Stdlib.Lexing.lexbuf -> rtrc

Parse an Isla trace from a Lexing.lexbuf

val parse_trc_string : ?filename:string -> string -> rtrc

Parse an Isla trace from a string

val parse_trc_channel : ?filename:string -> Stdlib.in_channel -> rtrc

Parse an Isla trace from a channel

val parse_trcs : ?filename:string -> Stdlib.Lexing.lexbuf -> Isla_lang__.Isla_lang_ast.lrng Isla_lang__.Isla_lang_ast.trcs

val parse_trcs_string : ?filename:string -> string -> rtrcs

val parse_trcs_channel : ?filename:string -> Stdlib.in_channel -> rtrcs

val parse_segments : ?filename:string -> Stdlib.Lexing.lexbuf -> instruction_segments

val parse_segments_string : ?filename:string -> string -> instruction_segments

val parse_segments_channel : ?filename:string -> Stdlib.in_channel -> instruction_segments

IslaTrace pretty printing

include module type of struct include Isla_lang.PP end

val pp_raw_vvar : int -> PPrint.document

val pp_raw_name : string -> PPrint.document

val pp_raw_enum_ty : int -> PPrint.document

val pp_raw_enum : Isla_lang__.Isla_lang_ast.enum -> PPrint.document

val pp_raw_nat : int -> PPrint.document

val pp_raw_bvi : int -> PPrint.document

val pp_raw_bv : string -> PPrint.document

val pp_raw_str : string -> PPrint.document

val pp_raw_int : int -> PPrint.document

val pp_raw_ty : Isla_lang__.Isla_lang_ast.ty -> PPrint.document

val pp_raw_bool : bool -> PPrint.document

val pp_raw_unop : Isla_lang__.Isla_lang_ast.unop -> PPrint.document

val pp_raw_bvarith : Isla_lang__.Isla_lang_ast.bvarith -> PPrint.document

val pp_raw_bvcomp : Isla_lang__.Isla_lang_ast.bvcomp -> PPrint.document

val pp_raw_binop : Isla_lang__.Isla_lang_ast.binop -> PPrint.document

val pp_raw_bvmanyarith : Isla_lang__.Isla_lang_ast.bvmanyarith -> PPrint.document

val pp_raw_manyop : Isla_lang__.Isla_lang_ast.manyop -> PPrint.document

val pp_raw_base_val : Isla_lang__.Isla_lang_ast.base_val -> PPrint.document

val pp_raw_assume_val : Isla_lang__.Isla_lang_ast.assume_val -> PPrint.document

val pp_raw_exp : 'a Isla_lang__.Isla_lang_ast.exp -> PPrint.document

val pp_raw_a_exp : 'a Isla_lang__.Isla_lang_ast.a_exp -> PPrint.document

val pp_raw_smt : 'a Isla_lang__.Isla_lang_ast.smt -> PPrint.document

val pp_raw_valu : Isla_lang__.Isla_lang_ast.valu -> PPrint.document

val pp_raw_selem : (string * Isla_lang__.Isla_lang_ast.valu) -> PPrint.document

val pp_raw_accessor : Isla_lang__.Isla_lang_ast.accessor -> PPrint.document

val pp_raw_accessor_list : Isla_lang__.Isla_lang_ast.accessor_list -> PPrint.document

val pp_raw_arg_list : Isla_lang__.Isla_lang_ast.arg_list -> PPrint.document

val pp_raw_tag_value : Isla_lang__.Isla_lang_ast.tag_value -> PPrint.document

val pp_raw_event : 'a Isla_lang__.Isla_lang_ast.event -> PPrint.document

val pp_raw_segment : Isla_lang__.Isla_lang_ast.segment -> PPrint.document

val pp_raw_instruction_segments : Isla_lang__.Isla_lang_ast.instruction_segments -> PPrint.document

val pp_raw_trc : 'a Isla_lang__.Isla_lang_ast.trc -> PPrint.document

val pp_raw_trcs : 'a Isla_lang__.Isla_lang_ast.trcs -> PPrint.document

val pp_raw_maybe_fork : 'a Isla_lang__.Isla_lang_ast.maybe_fork -> PPrint.document

val pp_raw_tree_trc : 'a Isla_lang__.Isla_lang_ast.tree_trc -> PPrint.document

val pp_raw_whole_tree : 'a Isla_lang__.Isla_lang_ast.whole_tree -> PPrint.document

val pp_vvar : int -> PPrint.document

val pp_name : string -> PPrint.document

val pp_enum_ty : int -> PPrint.document

val pp_enum : Isla_lang__.Isla_lang_ast.enum -> PPrint.document

val pp_nat : int -> PPrint.document

val pp_bvi : int -> PPrint.document

val pp_bv : string -> PPrint.document

val pp_str : string -> PPrint.document

val pp_j : int -> string

val pp_int : int -> PPrint.document

val pp_ty : Isla_lang__.Isla_lang_ast.ty -> PPrint.document

val pp_bool : bool -> PPrint.document

val pp_unop : Isla_lang__.Isla_lang_ast.unop -> PPrint.document

val pp_bvarith : Isla_lang__.Isla_lang_ast.bvarith -> PPrint.document

val pp_bvcomp : Isla_lang__.Isla_lang_ast.bvcomp -> PPrint.document

val pp_binop : Isla_lang__.Isla_lang_ast.binop -> PPrint.document

val pp_bvmanyarith : Isla_lang__.Isla_lang_ast.bvmanyarith -> PPrint.document

val pp_manyop : Isla_lang__.Isla_lang_ast.manyop -> PPrint.document

val pp_base_val : Isla_lang__.Isla_lang_ast.base_val -> PPrint.document

val pp_assume_val : Isla_lang__.Isla_lang_ast.assume_val -> PPrint.document

val pp_exp : 'a Isla_lang__.Isla_lang_ast.exp -> PPrint.document

val pp_a_exp : 'a Isla_lang__.Isla_lang_ast.a_exp -> PPrint.document

val pp_smt : 'a Isla_lang__.Isla_lang_ast.smt -> PPrint.document

val pp_valu : Isla_lang__.Isla_lang_ast.valu -> PPrint.document

val pp_selem : (string * Isla_lang__.Isla_lang_ast.valu) -> PPrint.document

val pp_accessor : Isla_lang__.Isla_lang_ast.accessor -> PPrint.document

val pp_accessor_list : Isla_lang__.Isla_lang_ast.accessor_list -> PPrint.document

val pp_arg_list : Isla_lang__.Isla_lang_ast.arg_list -> PPrint.document

val pp_tag_value : Isla_lang__.Isla_lang_ast.tag_value -> PPrint.document

val pp_event : 'a Isla_lang__.Isla_lang_ast.event -> PPrint.document

val pp_segment : Isla_lang__.Isla_lang_ast.segment -> PPrint.document

val pp_instruction_segments : Isla_lang__.Isla_lang_ast.instruction_segments -> PPrint.document

val pp_trc : 'a Isla_lang__.Isla_lang_ast.trc -> PPrint.document

val pp_trcs : 'a Isla_lang__.Isla_lang_ast.trcs -> PPrint.document

val pp_maybe_fork : 'a Isla_lang__.Isla_lang_ast.maybe_fork -> PPrint.document

val pp_tree_trc : 'a Isla_lang__.Isla_lang_ast.tree_trc -> PPrint.document

val pp_whole_tree : 'a Isla_lang__.Isla_lang_ast.whole_tree -> PPrint.document

module Base : sig ... end

This module wraps all isla-lang functionality. No other module should directly touch the Isla_lang module.

module Cache : sig ... end

This module provide a caching system for isla trace on top of Server.

module Conv : sig ... end

module Manip : sig ... end

This module provide generic manipulation function of isla ast

module Preprocess : sig ... end

This module is about preprocessing isla traces. This includes:

module Relocation : sig ... end

module Run : sig ... end

This module provide facility to run Isla trace over states

module Server : sig ... end

This module is about launching isla as a background server and using it

module Test : sig ... end

This file about testing interaction with Isla for single instructions

module Type : sig ... end

This module is about type isla trace and register discovery.