Module Isla.Cache

This module provide a caching system for isla trace on top of Server.

It uses the cache named "isla" of Cache

Call start to start and stop. Do not interact directly with the Server if you use the cache.

You can use ensure_started to force the Server to start but you probably shouldn't do that. By the default the Server is only started if the traces of an instruction are required and not in the cache.

type config = Server.config

The type of Isla configuration

module Opcode : sig ... end

Implementation of Cache.Key for opcodes.

module TraceList : sig ... end

Representation of trace lists on disk.

val epoch : int

An epoch independant of the isla version, bump if you change the representation of the traces on disk.

Reset (or not) when bumping isla version (Server.required_version)

The Epoch also include the digest of the Isla configuration. Any change of configuration will wipe out the cache.

module Epoch : sig ... end

module IC : sig ... end

The isla cache module

val cache : (IC.t * config) option Stdlib.ref

This varaible stores the cache RAM representation

val configr : config option Stdlib.ref

If this is set and cache is also set, then the server should be started with the architecture inside this variable when required

val start : config -> unit

Start the caching system. Do not yet start the server

val stop : unit -> unit

Stop the caching system, stop the server if it was started

val ensure_started : unit -> unit

Start the server if not already started

val get_cache : unit -> IC.t * config

Get the cache and fails if the cache wasn't started

val get_traces : Server.opcode -> Base.rtrcs

Get the traces of the opcode given. Use Server if the value is not in the cache

val get_nop : unit -> Base.rtrc

Get the traces of the nop opcode (The initialization code). Use Server if the value is not in the cache