Module Isla.Conv

val ty : Base.ty -> 'a Ast.ty

val unop : Base.unop -> Ast.unop

val bvarith : Base.bvarith -> Ast.bvarith

val bvcomp : Base.bvcomp -> Ast.bvcomp

val bvmanyarith : Base.bvmanyarith -> Ast.bvmanyarith

val binop : Base.binop -> 'm Ast.binop

val manyop : Base.manyop -> Ast.manyop

val direct_exp_no_var : ('a Base.exp -> ('a, 'v, 'b, 'm) Ast.exp) -> 'a Base.exp -> ('a, 'v, 'b, 'm) Ast.exp

val exp_var_conv : (int -> 'v) -> 'a Base.exp -> ('a, 'v, 'b, 'm) Ast.exp

val exp : 'a Base.exp -> ('a, int, 'b, 'c) Ast.exp

val exp_var_subst : (int -> 'a -> ('a, 'v, 'b, 'm) Ast.exp) -> 'a Base.exp -> ('a, 'v, 'b, 'm) Ast.exp

Convert an expression from isla to Ast but using a var-to-exp conversion function

val exp_add_type_var_subst : (int -> 'a -> ('v, 'm) Exp.Typed.t) -> 'a Base.exp -> ('v, 'm) Exp.Typed.t

Convert directly from an untyped isla expression to an Exp.Typed by substituing isla variables with already typed expressions

val smt_var_conv : (int -> 'v) -> 'a Base.smt -> ('a, 'v, 'b, 'm) Ast.smt

val smt : 'a Base.smt -> ('a, int, 'b, 'c) Ast.smt