Module Isla.Type

This module is about type isla trace and register discovery.

The actual goal is dicovering the existence and type of registers from Isla traces, but this require type the trace with ty and thus full type checking of traces. We expect isla to be correct, so a type error would be very surprising.

type type_context = Base.ty Utils.HashVector.t

A context that associate Isla types to Isla variables

type lty = Base.lrng * Base.ty

A group of the source range and type for an expression

exception TypeError of Base.lrng * string

Exception that represent an Isla typing error

val tassert : Base.lrng -> string -> bool -> unit

Assert some properties for type correctness. Requires a lrng and a string error message

val expect_bool : string -> (Base.lrng * Base.ty) -> unit

val expect_bv : string -> (Base.lrng * Base.ty) -> int

val expect_enum : string -> (Base.lrng * Base.ty) -> int

val type_unop : Base.lrng -> Base.unop -> lty -> Base.ty

val type_binop : Base.lrng -> Base.binop -> (Base.lrng * Base.ty) -> (Base.lrng * Base.ty) -> Base.ty

val type_manyop : Base.lrng -> Base.manyop -> lty list -> Base.ty

val type_valu : Base.lrng -> type_context -> Base.valu -> (State.Reg.Path.t * State.Reg.ty) list

Take an Isla value and a context and give the list of field that correspond to that value. Those fields would need to be prefixed with the top register name before being added in State.Reg

val ltype_expr : type_context -> Base.lrng Base.exp -> lty

val type_expr : type_context -> Base.lrng Base.exp -> Base.ty

val type_trc : ?tc:Base.ty Utils.HashVector.t -> Base.rtrc -> Base.ty Utils.HashVector.t

Add the new register found in the trace and returns the type context for free variables

val pp_tcontext : Isla_lang__.Isla_lang_ast.ty Utils.HashVector.t -> Utils.Pp.document

Print a type context for debugging