Module Isla__.Manip

Trace Properties

val is_linear : 'a Isla__.Base.trc -> bool

Check if a trace is linear (has no branches)

Get annotations

val annot_exp : 'a Isla__.Base.exp -> 'a

Get the annotation of an expression

val annot_event : 'a Isla__.Base.event -> 'a

Get the annotation of an event

Non-recursive maps and iters

This section is filled on demand.

direct_a_map_b take a function b -> b and applies it to all b in a, non-recursively. Then a new a with the same structure is formed.

direct_a_iter_b take a function b -> unit and applies it to all b in a, non-recursively.

val direct_exp_map_exp : ('a Isla__.Base.exp -> 'a Isla__.Base.exp) -> 'a Isla__.Base.exp -> 'a Isla__.Base.exp

val direct_exp_iter_exp : ('a Isla__.Base.exp -> unit) -> 'a Isla__.Base.exp -> unit

val direct_smt_map_exp : ('a Isla__.Base.exp -> 'a Isla__.Base.exp) -> 'a Isla__.Base.smt -> 'a Isla__.Base.smt

val direct_event_map_exp : ('a Isla__.Base.exp -> 'a Isla__.Base.exp) -> 'a Isla__.Base.event -> 'a Isla__.Base.event

val direct_event_iter_valu : (Isla__.Base.valu -> unit) -> 'a Isla__.Base.event -> unit

val direct_event_map_valu : (Isla__.Base.valu -> Isla__.Base.valu) -> 'a Isla__.Base.event -> 'a Isla__.Base.event

val direct_valu_iter_valu : (Isla__.Base.valu -> unit) -> Isla__.Base.valu -> unit

val direct_valu_map_valu : (Isla__.Base.valu -> Isla__.Base.valu) -> Isla__.Base.valu -> Isla__.Base.valu

Recursive maps and iters

This section is filled on demand.

a_map_b take a function b -> b and applies it to all the b in a, and do that recursively on all b that appear directly or indirectly in a

a_iter_b take a function b -> unit and applies it to all the b in a, and do that recursively

Doing this when a = b is not well defined, and can be easily done using the direct version from previous section.

In case where a type is not recusive like event, both direct and recursive versions are the same.

val exp_iter_var : (int -> unit) -> 'a Isla__.Base.exp -> unit

iterate a function on all the variable of an expression

val event_iter_valu : (Isla__.Base.valu -> unit) -> 'a Isla__.Base.event -> unit

Variable substitution

val var_subst : (int -> 'a -> 'a Isla__.Base.exp) -> 'a Isla__.Base.exp -> 'a Isla__.Base.exp

Substitute variable with expression according to subst function

Accessor list conversion

val accessor_of_string : string -> Isla__.Base.accessor

val string_of_accessor : Isla__.Base.accessor -> string

val accessor_of_string_list : string list -> Isla__.Base.accessor_list

val string_of_accessor_list : Isla__.Base.accessor_list -> string list

Valu string path access

val valu_get : Isla__.Base.valu -> string list -> Isla__.Base.valu

Follow the path in a value like A.B.C in (struct (|B| (struct (|C| ...))))

Trace Filtering

This is some basic trace filtering that remove unwanted item from the trace in various situations

val split_cycle : 'a Isla__.Base.trc -> 'a Isla__.Base.trc * 'a Isla__.Base.trc

Split the trace between before and after the "cycle" event

val remove_init : 'a Isla__.Base.trc -> 'a Isla__.Base.trc

Remove all events before the "cycle" event, keep the SMT statements

val remove_ignored : string list -> 'a Isla__.Base.trc -> 'a Isla__.Base.trc

Remove all the events related to ignored registers