Module Isla__.Server

module ConfigFile = Config.File

module CommonOpt = Config.CommonOpt

module Config = Config.File.ArchConf.Isla

module Server = Utils.Cmd.SocketServer

type config = Config.t

The configuration record type

type trcs = (bool * Isla.Base.rtrc) list

The raw output of the server for an instruction.

It is a list of traces, each with a flag telling if they are normal traces (no processor exception/fault) or not

val required_version : string

Bump when updating isla. TODO: move the version checking to allow a range of version. Also, right now the cache invalidation is based on this and not on the actual isla version, which may be dangerous.

val req_num : int Stdlib.ref

val server : Server.t option Stdlib.ref

This instance of socket server for isla

val get_server : unit -> Server.t

Assume the server is started and get it out of the reference

val cmd_of_config : config -> string -> string array

Compute the isla-client command line from the isla configuration

val raw_start : config -> unit

Start the server with the specified architecture, do not attempt any checks

val raw_stop : unit -> unit

Stop the server by cutting the connection.

type basic_answer =

| Error

| Version of string

| StartTraces

| Trace of bool * string

| EndTraces

This should match exactly with the Answer type in isla-client code

val read_basic_answer : unit -> basic_answer

Read an answer from isla-client. This must match exactly write_answer in client.rs in isla

type answer =

| Version of string

| Traces of (bool * string) list

The interpreted answer. If the protocol is followed, then one request lead to exactly one answer of that type

val expect_version : answer -> string

Expect a version answer and fails if it is not the case

val expect_traces : answer -> (bool * string) list

Expect isla traces and fails if it is not the case

val expect_parsed_traces : answer -> trcs

Expect isla traces and fails if it is not the case, additionally parse them

exception IslaError

When isla encounter a non fatal error with that specific request. This error is recoverable and the sever can accept other requests

val read_answer : unit -> answer

Read the answer from isla, block until full answer

val pp_answer : answer -> PPrintEngine.document

Answer pretty printer

type request =

| TEXT_ASM of string

| ASM of Utils.BytesSeq.t

| VERSION

| STOP

The type of a request to isla

val string_of_request : request -> string

Convert a request into the string message expected by isla-client This should match the protocol

val send_string_request : string -> unit

Send a string request to the server, and do not wait for any answer

val string_request : string -> answer

Same as request but takes the request directly as a string

val request : request -> answer

Send a request and wait for answer

val request_bin_parsed : Utils.BytesSeq.t -> trcs

Request the traces of a binary instruction and parse the result.

This is the main entry point of this module.

val send_request : request -> unit

Send a request without expecting any answer

val stop : unit -> unit

Stop isla client by sending a stop request

val start : config -> unit

Start isla and check version

module Cmd : sig ... end

This module provide a CLI subcommand to test isla directly. All isla output is reported as raw text