Module Isla__.Type

type type_context = Isla__.Base.ty Utils.HashVector.t

A context that associate Isla types to Isla variables

type lty = Isla__.Base.lrng * Isla__.Base.ty

A group of the source range and type for an expression

exception TypeError of Isla__.Base.lrng * string

Exception that represent an Isla typing error

val tassert : Isla__.Base.lrng -> string -> bool -> unit

Assert some properties for type correctness. Requires a lrng and a string error message

val expect_bool : string -> (Isla__.Base.lrng * Isla__.Base.ty) -> unit

val expect_bv : string -> (Isla__.Base.lrng * Isla__.Base.ty) -> int

val expect_enum : string -> (Isla__.Base.lrng * Isla__.Base.ty) -> int

val type_unop : Isla__.Base.lrng -> Isla__.Base.unop -> lty -> Isla__.Base.ty

val type_binop : Isla__.Base.lrng -> Isla__.Base.binop -> (Isla__.Base.lrng * Isla__.Base.ty) -> (Isla__.Base.lrng * Isla__.Base.ty) -> Isla__.Base.ty

val type_manyop : Isla__.Base.lrng -> Isla__.Base.manyop -> lty list -> Isla__.Base.ty

val type_valu : Isla__.Base.lrng -> type_context -> Isla__.Base.valu -> (State.Reg.Path.t * State.Reg.ty) list

Take an Isla value and a context and give the list of field that correspond to that value. Those fields would need to be prefixed with the top register name before being added in State.Reg

val ltype_expr : type_context -> Isla__.Base.lrng Isla__.Base.exp -> lty

val type_expr : type_context -> Isla__.Base.lrng Isla__.Base.exp -> Isla__.Base.ty

val type_trc : ?⁠tc:Isla__.Base.ty Utils.HashVector.t -> Isla.Base.rtrc -> Isla__.Base.ty Utils.HashVector.t

Add the new register found in the trace and returns the type context for free variables

val pp_tcontext : Isla_lang__.Isla_lang_ast.ty Utils.HashVector.t -> Utils.Pp.document

Print a type context for debugging