Module Isla__Conv

val ty : Isla__.Base.ty -> 'a Ast.ty

val unop : Isla__.Base.unop -> Ast.unop

val bvarith : Isla__.Base.bvarith -> Ast.bvarith

val bvcomp : Isla__.Base.bvcomp -> Ast.bvcomp

val bvmanyarith : Isla__.Base.bvmanyarith -> Ast.bvmanyarith

val binop : Isla__.Base.binop -> 'm Ast.binop

val manyop : Isla__.Base.manyop -> Ast.manyop

val direct_exp_no_var : ('a Isla__.Base.exp -> ('a, 'v, 'b, 'm) Ast.exp) -> 'a Isla__.Base.exp -> ('a, 'v, 'b, 'm) Ast.exp

val exp_var_conv : (int -> 'v) -> 'a Isla__.Base.exp -> ('a, 'v, 'b, 'm) Ast.exp

val exp : 'a Isla__.Base.exp -> ('a, int, 'b, 'c) Ast.exp

val exp_var_subst : (int -> 'a -> ('a, 'v, 'b, 'm) Ast.exp) -> 'a Isla__.Base.exp -> ('a, 'v, 'b, 'm) Ast.exp

Convert an expression from isla to Ast but using a var-to-exp conversion function

val exp_add_type_var_subst : (int -> 'a -> ('v, 'm) Exp.Typed.t) -> 'a Isla__.Base.exp -> ('v, 'm) Exp.Typed.t

Convert directly from an untyped isla expression to an Exp.Typed by substituing isla variables with already typed expressions

val smt_var_conv : (int -> 'v) -> 'a Isla__.Base.smt -> ('a, 'v, 'b, 'm) Ast.smt

val smt : 'a Isla__.Base.smt -> ('a, int, 'b, 'c) Ast.smt