Module Isla__Test

This file about testing interaction with Isla for single instructions

module SMT = Z3

type pmode =

| DUMP
Dump isla output on standard output

| PARSE
parse isla output and prettyprint it

| TYPE
Type isla output and dump var types. Also dump the deduced register file

| RUN
Run the the isla output on a test state and print all branches and states

| SIMP
Run the the isla output on a test state and also print a simplified version

The type of processing requested

type isla_mode =

| RAW
Do not call isla and take the input test as if it was isla output

| ASM
Call isla with text assembly

| HEX
Call isla with text hexadecimal as binary encoding of instruction

| BIN
Call isla with actually binary (for now 4 bytes)

The input syntax

type imode =

| CMD
Read the input as the main command line argument

| FILE
Read the input in a file

| ELF of string
Read the input from an elf symbol + offset. implies BIN for isla_mode

The way input is taken

val arg : string Cmdliner.Term.t

val direct : bool Cmdliner.Term.t

val bin : bool Cmdliner.Term.t

val hex : bool Cmdliner.Term.t

val noparse : bool Cmdliner.Term.t

val preprocess : bool Cmdliner.Term.t

val typer : bool Cmdliner.Term.t

val run : bool Cmdliner.Term.t

val simp : bool Cmdliner.Term.t

val file : bool Cmdliner.Term.t

val sym : string option Cmdliner.Term.t

val input_f2m : bool -> string option -> imode Cmdliner.Term.ret

Input flag to mode conversion

val imode_term : imode Cmdliner.Term.t

val input : imode -> string -> (string * string) Cmdliner.Term.ret

Input takes the imode and the main argument and returns the filename and input string

val input_term : (string * string) Cmdliner.Term.t

val isla_f2m : bool -> bool -> bool -> 'a option -> isla_mode Cmdliner.Term.ret

Convert various flag describe the mode of operation into the mode of operation If sym is activated, then the default mode is BIN and not ASM

val isla_mode_term : isla_mode Cmdliner.Term.t

val isla_mode_to_request : isla_mode -> string -> Isla.Server.request

val isla_run : isla_mode -> Config.Arch.t -> (string * string) -> string * string * Isla.Server.config

Run isla and return a text trace with a filename (if mode is RAW than just return the trace and filename without isla)

If isla return multiple traces, just silently pick the first non-exceptional one

val isla_term : (string * string * Isla.Server.config) Cmdliner.Term.t

val processing_f2m : bool -> bool -> bool -> bool -> pmode

How far into the processing pipeline we go. We just pick the deepest option chosen

val pmode_term : pmode Cmdliner.Term.t

val processing : bool -> pmode -> (string * string * Isla.Server.config) -> unit

Does the actual processing of the trace

val term : unit Cmdliner.Term.t

val info : Cmdliner.Term.info

val command : unit Cmdliner.Term.t * Cmdliner.Term.info