Module Run

module BB : sig ... end

This module allow to do a test run of all the machinery for a single basic block

module Bb_lib : sig ... end

This module provide code to manipulate basic block and run them.

module Block : sig ... end

This module add the run-block sub command.

module Block_lib : sig ... end

This module provides a representation of a complex block of code.

module Func : sig ... end

module FuncRD : sig ... end

This module is a merge of rd sub command and run-func --loop=1. It will run the equivalent of run-func --loop=1, and print the instruction like rd but will print the state in a light form (State.pp_partial) between instructions.

module Init : sig ... end

This module handle architecture initialization. For now it trust default isla initialization but this should change soon (TODO)

module Instr : sig ... end

This module add the run-instr sub command.

module ReadDwarf : sig ... end

This module implements the rd subcommand. This was to original meaning of "read-dwarf", and the first command.

module RelProg : sig ... end

module Runner : sig ... end

This module provide the program runner that caches all the information required to make a state transition

module TestRelProg : sig ... end