Module Run.BB

This module allow to do a test run of all the machinery for a single basic block

module SMT = Z3

val dump : bool Cmdliner.Term.t

val reg_types : bool Cmdliner.Term.t

val no_run : bool Cmdliner.Term.t

val simp_trace : bool Cmdliner.Term.t

val simp_state : bool Cmdliner.Term.t

val simp : bool Cmdliner.Term.t

val elf : string Cmdliner.Term.t

val sym : string Cmdliner.Term.t

val len : int option Cmdliner.Term.t

val get_code : string -> string -> int option -> Utils.BytesSeq.t

val code_term : Utils.BytesSeq.t Cmdliner.Term.t

val simp_trace_term : bool Cmdliner.Term.t

val simp_state_term : bool Cmdliner.Term.t

val get_bb : bool -> bool -> bool -> Utils.BytesSeq.t -> Bb_lib.t

val bb_term : Bb_lib.t Cmdliner.Term.t

val run_bb : bool -> bool -> Bb_lib.t -> unit

val term : unit Cmdliner.Term.t

val info : Cmdliner.Cmd.info

val command : unit Cmdliner.Term.t * Cmdliner.Cmd.info