Module Run.Block_lib

This module provides a representation of a complex block of code.

The end of the block is decided by endpred, an arbitrary predicate on the pc. In particular if the PC is symbolic the execution is stopped anyway. This means that we either reached the top level function return or an unresolved branch table.

To generate easily end predicates, there is gen_endpred.

type t = {

runner : Runner.t;

start : Elf.Address.t;

endpred : State.exp -> string option;

}

endpred pc_exp gives when to stop

val make : runner:Runner.t -> start:Elf.Address.t -> endpred:(State.exp -> string option) -> t

Build a complex block starting from start in sym and ending when endpred says so. endpred is a predicate on the symbolic PC expression

type label =

| Start
(*
Root node of the tree
*)

| End of string
(*
Lead node of the tree, the string describe which end condition has be triggered
*)

| BranchAt of Elf.Address.t
(*
A Branching node at a given PC
*)

| NormalAt of Elf.Address.t
(*
A normal instruction at PC. Exists only if every_instruction is true
*)

The labels on tree node at the output of run

val label_to_string : label -> string

val pp_label : label -> Utils.Pp.document

val run : ?every_instruction:bool -> ?relevant:(Elf.Address.t, 'a) Stdlib.Hashtbl.t -> t -> State.t -> label State.Tree.t

Run the block an return a state tree indexed by the addresses of the branches.

When every_instruction is true, It will make a snapshot of the state i.e a tree node at each instruction. By default it will only make a Tree node on branching points.

The output is a tree because state merging is not implemented so if we are going twice on the same PC, the whole thing will be run twice separately in two separate tree branches.

val gen_endpred : ?min:Elf.Address.t -> ?max:Elf.Address.t -> ?loop:int -> ?brks:Elf.Address.t list -> unit -> State.exp -> string option

Generic end predicate. Will stop if:

pc below min

pc above max

pc is one of brks

pc has be seen more than loop