Module Run.FuncRD

This module is a merge of rd sub command and run-func --loop=1. It will run the equivalent of run-func --loop=1, and print the instruction like rd but will print the state in a light form (State.pp_partial) between instructions.

val run_func_rd : string -> string -> string -> string option -> string list -> unit

val elf : string Cmdliner.Term.t

val func : string Cmdliner.Term.t

val objdump_d : string Cmdliner.Term.t

val branch_table : string option Cmdliner.Term.t

val breakpoints : string list Cmdliner.Term.t

val term : unit Cmdliner.Term.t

val info : Cmdliner.Cmd.info

val command : unit Cmdliner.Term.t * Cmdliner.Cmd.info