Module Run.Instr

This module add the run-instr sub command.

This subcommand is about testing the various operations that can happen in an instruction processing. This is not isla-test, the instructions will always be fetched from the cache and use Init. If this fail at an earlier point, use isla-test to debug the Isla pipeline

Additionally run-instr supports branching instructions.

type traces =

| IslaTraces of Isla.rtrc list

| Traces of Trace.t list

val instr : string Cmdliner.Term.t

val dump_trace : bool Cmdliner.Term.t

val dump_isla : bool Cmdliner.Term.t

val no_run : bool Cmdliner.Term.t

val isla_run : bool Cmdliner.Term.t

val simp_trace : bool Cmdliner.Term.t

val simp_state : bool Cmdliner.Term.t

val simp : bool Cmdliner.Term.t

val reg_types : bool Cmdliner.Term.t

val init : bool Cmdliner.Term.t

val elf : string option Cmdliner.Term.t

val get_instr : Config.Arch.t -> string -> string option -> Utils.BytesSeq.t

val instr_term : Utils.BytesSeq.t Cmdliner.Term.t

val simp_trace_term : bool Cmdliner.Term.t

val simp_state_term : bool Cmdliner.Term.t

val get_traces : 'a -> 'b -> 'c -> traces

val pre_traces_term : traces Cmdliner.Term.t

val simp_traces : bool -> traces -> traces

val dump_traces : bool -> traces -> traces

val traces_term : traces Cmdliner.Term.t

val run_instr : bool -> bool -> bool -> traces -> unit

val term : unit Cmdliner.Term.t

val info : Cmdliner.Cmd.info

val command : unit Cmdliner.Term.t * Cmdliner.Cmd.info