Module Run__Block

This module add the run-block sub command.

This subcommand is about running a complex block of execution It start at a specific offset in a block and can terminate on various condition. If not enough condition are met, it may crash by reaching an invalid instruction

val no_run : bool Cmdliner.Term.t

val reg_types : bool Cmdliner.Term.t

val len : int option Cmdliner.Term.t

val breakpoints : string list Cmdliner.Term.t

val ensure_linear : bool Cmdliner.Term.t

val elf : string Cmdliner.Term.t

val start : string Cmdliner.Term.t

val get_elf_start : string -> string -> Elf.File.t * Elf.SymTable.sym_offset

val elf_term : (Elf.File.t * Elf.SymTable.sym_offset) Cmdliner.Term.t

val gen_block : (Elf.File.t * Elf.SymTable.sym_offset) -> int option -> string list -> Elf.File.t * Run.Block_lib.t

val elfblock_term : (Elf.File.t * Run.Block_lib.t) Cmdliner.Term.t

val prune_paths : (State.Base.t -> bool) -> 'a State.Tree.t -> 'a State.Tree.t option

val has_assert_false : State.t -> bool

val run_block : (Elf.File.t * Run.Block_lib.t) -> bool -> bool -> bool -> unit

val term : unit Cmdliner.Term.t

val info : Cmdliner.Term.info

val command : unit Cmdliner.Term.t * Cmdliner.Term.info