Module Run__Runner

This module provide the program runner that caches all the information required to make a state transition

Later this module will handle inlining in a transparent way, and may also compress basic block traces (We'll need to check this is okay with type inference)

For now this module load instructions on a per-symbol basis, and do not try to load instructions outside of function symbol. TODO: remove this restriction.

The final goal of this module is to encode all necessary information to perform a state transition. In particular, the run function should never require more argument: If extra information is required, it should be part of the runner if it is dependent on the whole program or in the state if it is specific to the state.

module Reg = State.Reg

type slot =

| Normal of Trace.Instr.t
The traces and the size of the instruction

| Special
Special instructions. Will be used to represent external events

| Nocode
The is no code at this address. Running it is UB. Also used if an address is in between instructions

| IslaFail
This means Isla pipeline failed on that instruction

Give the instruction descriptor at a given address

type t = {

elf : Elf.File.t;

dwarf : Dw.t option;

instrs : (int, slot) Stdlib.Hashtbl.t;
Instruction cache

pc : Reg.t;

funcs : int Utils.Vec.t;
Loaded functions by loading order

}

val of_elf : ?⁠dwarf:Dw.t -> Elf.File.t -> t

val of_dwarf : Dw.t -> t

val load_sym : t -> Elf.Symbol.t -> unit

Load a symbol into the runner. All instruction traces are fetched and cached.

TODO support variable length instructions.

val fetch : t -> int -> slot

Fetch an instruction, and return corresponding slot.

val execute_normal : ?⁠prelock:(State.t -> unit) -> pc:int -> t -> Trace.Instr.t -> State.t -> State.t list

Run the traces on the state.

If the state is unlocked and the instruction is single trace, then the function mutates the state and returns the same state.

Otherwise the state is locked (if not) and new states are returned

In any case the returned states are unlocked.

val run : ?⁠prelock:(State.t -> unit) -> t -> State.t -> State.t list

Do the whole fetch and execute cycle. Take the PC from the state, and fetch it's instruction and then run it. It return the list of possible behavior of that instruction. Normally the union of the set of concrete states represented by this list of symbolic state cover all the defined behaviors of the fetched instruction from the initial state.

If the state is unlocked and the instruction is single trace, then the function mutates the state and returns the same state.

Otherwise the state is locked (if not) and new unlocked states are returned.

In any case the returned states are unlocked.

val expect_normal : t -> int -> Trace.Instr.t

Return the Instr.t data of the instruction at address, and throw Not_found if the instruction was invalid

val get_normal_opt : t -> int -> Trace.Instr.t option

Return the Instr.t data of the instruction at address, and None if the instruction was invalid

val pp_slot : slot -> PPrintEngine.document

Pretty prints a instruction slot

val pp_instr : t -> Utils.Pp.document

Dump instruction table