Module State.Base

This module introduce a type to represent the state of the machine.

The symbolic state in this module do not mathematically represent a single state but a set of concrete state with all the symbolic variable over the whole range of their types. State also contain assertions, and so only represent the subset of concrete states, that satisfy all assertions.

Currently the state type only represent the register (including system register) and sequential memory part of an actual machine state. Any other architectural state is not represented.

Additionally, state contain C type information for the Ctype inference system. Those fields are not semantically part of the state and do not influence in any way which concrete states are represented by the symbolic state. Expect for provenance information: Pointers can be tagged with provenance information which mean that they are part of a specific restricted block of memory or the main block. See Mem for more information. Thus all the implicit non-aliasing assertion implied by those provenance field are to be considered as part of the group of assertion restricting the set of concrete state represented by a symbolic state.

The presence of C types is optional, which means that this state type can be used in a completely untyped context.

Concrete detail about how symbolic state are represented in Ocaml is in the documentation of t.

State id

module Id : sig ... end

The type of a state ID. for now it's an integer, but it may change later In particular whether a state belong to O0 or O2 may be part of the id at some point.

type id = Id.t

State variable management

module Var : sig ... end

This module provide state variables. Those are all symbolic variables that may appear in a state. If the same (in the Var.equal sense) variable appear in two state, that mean that when considered together, there is an implicit relation between them.

type var = Var.t

The type of variables

State expression and typed-value management

module Exp : sig ... end

Module for state expressions

type exp = Exp.t

module Tval : sig ... end

Module for optionally typed state expressions. Those are symbolic values which may or may not have a C type.

type tval = Tval.t

module Relocation : sig ... end

State memory management

module Mem : sig ... end

This module manages the memory part of the state.

State type

type t = private {

id : id;

base_state : t option;
(*
The immediate dominator state in the control flow graph
*)

mutable locked : bool;
(*
Tells if the state is locked
*)

mutable regs : Tval.t Reg.Map.t;
(*
The values and types of registers
*)

read_vars : Tval.t Utils.Vec.t;
(*
The results of reads made since base state
*)

mutable asserts : exp list;
(*
Only asserts since base_state
*)

mutable relocation_asserts : exp list;
(*
Only asserts since base_state
*)

mem : Mem.t;

elf : Elf.File.t option;
(*
Optionally an ELF file, this may be used when running instructions on the state to provide more concrete values in certain case (like when reading from .rodata). It will affect the execution behavior. However the symbolic execution should always be more concrete with it than without it
*)

fenv : Fragment.env;
(*
The memory type environment. See Fragment.env
*)

mutable last_pc : Elf.Address.t;
(*
The PC of the instruction that lead into this state. The state should be right after that instruction. This has no semantic meaning as part of the state. It's just for helping knowing what comes from where
*)

}

Represent the state of the machine.

State are represented by their id and may identified to their id, they may not be two different state (and I mean physical equality here) with same id. See State to id management.

A first remark must be made about mutability: The type itself has an imperative interface, a lot of implicitly or explicitly mutable fields. However, Sometime immutable version of the state are required, so the state has a "locking" mechanism. When the locked field, the state becomes immutable. This is unfortunately not enforced by the type system as that would require to have two different types. However all mutating functions assert that the state is unlocked before doing the mutation. The normal workflow with state is thus to create them unlocked, generaly by copying another state, then mutate is to make a new interesting state, and then lock it so that it can be passed around for it's mathematical pure meaning. To lock a state, use the lock function.

A second subtlety is that state are not represented in a standalone manner, They are represented as diff from a previous state, the base_state. In the idea, this state should be the immediate dominator of the current state in the control flow graph.However a state may not represent a full node of the control flow graph but only the part of that node that represent control-flow coming from specific paths. In that case the dominator notion is only about those paths.

Assertions (asserts) and memory (mem) are represented as diffs from the base state. In particular all assertion constraining the base state are still constraining the child state.

This also implies a restriction on state dependent variables like Var.t.Register. The id of such variables can only be the id of the current state or one of it's ancestor. Further more if a variable of type Var.t.ReadVar exists with and id and a number, then the read_vars array of the state with that id must contain that number and the sizes must match. Those restriction are not only about semantic meaning but also more practical Ocaml Gc consideration, see State to id management.

val equal : t -> t -> bool

State to id management

Each state has an id and the state can be refereed physically by id. This mean that there cannot be two different physical Ocaml state in the Gc memory that have the same id. Furthermore the id2state map is weak and so do not own the state. This means that possession of an id is the same as having a weak pointer to the state, except for two things:

The id can be serialized and read to external program and files without losing it's meaning.

The id allow to break cyclical type dependency.

This is in particular useful for variable that contain and id instead of a pointer to the state:

The variable can be serialized in text manner to a SMT solver and keep their meaning.

The Var.t type can be defined before the t type.

That means that in theory a state could be Garbage collected while a variable still point to it. However a variable is only allowed to exists in a state if the id it points to is among the ancestors via the base_state relationship of the containing state. Since base_state is an GC-owning pointer, this ensure that while the containing state is alive, the variable target state is also alive.

val id2state : (id, t) Utils.WeakMap.t

Global map of states to associate them with identifiers

val next_id : id Stdlib.ref

Next unused id

val of_id : id -> t

Get a state from its id

val to_id : t -> id

Get the id of a state

State management

val lock : t -> unit

Lock the state. Once a state is locked it should not be mutated anymore

val unsafe_unlock : t -> unit

Unlock the state. This is dangerous, do not use if you do not know how to use it, The only realistic use case, is for calling a simplifier and thus not changing the semantic meaning of the state in any way while mutating it.

This is deprecated and should disappear at some point.

deprecated Stop unlocking states

val is_locked : t -> bool

Tell if the state is locked, in which case it shouldn't be mutated

val is_possible : t -> bool

Tell is state is possible.

A state is impossible if it has a single assert that is false. This means that this symbolic state represent the empty set of concrete states.

StateSimplify.ctxfull will call the SMT solver and set the assertions to that if required so you should call that function before is_possible

val make : ?elf:Elf.File.t -> unit -> t

Makes a fresh state with all variable fresh and new. This fresh state is unlocked.

This should only be used by Init, all other state should be derived from Init.state.

val copy : ?elf:Elf.File.t -> t -> t

Do a deep copy of all the mutable part of the state, so it can be mutated without retro-action.

If the source state is locked, then new state is based on it (in the sense of t.base_state), otherwise it is a literal copy of each field.

The returned state is always unlocked

val copy_if_locked : ?elf:Elf.File.t -> t -> t

Copy the state with copy if and only if it is locked. The returned state is always unlocked

val init_sections : sp:(unit -> Reg.t) -> addr_size:int -> t -> t

val init_sections_symbolic : sp:(unit -> Reg.t) -> addr_size:int -> t -> t

Assigns all sections with global objects to Main fragment

State convenience manipulation

val push_assert : t -> exp -> unit

Add an assertion to a state

val push_relocation_assert : t -> exp -> unit

Add an assertion to a state

val set_impossible : t -> unit

Set a state to be impossible (single false assert).

val set_asserts : t -> exp list -> unit

Set a state's asserts

val map_mut_exp : (exp -> exp) -> t -> unit

Map a function on all the expressions of a state by mutating. This function, must preserve the semantic meaning of expression (like a simplification function) otherwise state invariants may be broken.

val iter_exp : (exp -> unit) -> t -> unit

Iterates a function on all the expressions of a state

val iter_var : (var -> unit) -> t -> unit

Iterates a function on all the variables of a state

State memory accessors

val make_read : t -> ?ctyp:Ctype.t -> Ast.Size.t -> var

Create a new Var.t.ReadVar by mutating the state

val set_read : t -> int -> exp -> unit

Set a Var.t.ReadVar to a specific value in t.read_vars

val read_from_rodata : t -> addr:exp -> size:Ast.Size.t -> exp option

Read memory from rodata

val read : provenance:Ctype.provenance -> ?ctyp:Ctype.t -> t -> addr:exp -> size:Ast.Size.t -> exp

Read the block designated by addr and size from the state and return an expression read. This will mutate the state to bind the read result to the newly created read variable.

The ctyp parameter may give a type to the read variable. This type is fully trusted and not checked in any way.

The expression could be either:

An actual expression if the read could be resolved.

Just the symbolic read variable if the read couldn't be resolved

This function is for case with provenance information is known.

val read_noprov : ?ctyp:Ctype.t -> t -> addr:exp -> size:Ast.Size.t -> exp

A wrapper around read for use when there is no provenance information. It may able to still perform the read under certain condition and otherwise will fail.

val write : provenance:Ctype.provenance -> t -> addr:exp -> size:Ast.Size.t -> exp -> unit

Write the provided value in the block. Mutate the state.

val write_noprov : t -> addr:exp -> size:Ast.Size.t -> exp -> unit

A wrapper around write for use when there is no provenance information. It may able to still perform the write under certain condition and otherwise will fail.

State register accessors

val reset_reg : t -> ?ctyp:Ctype.t -> Reg.t -> unit

Reset the register to a symbolic value, and resets the type to the provided type (or no type if not provided)

val set_reg : t -> Reg.t -> tval -> unit

Sets the content of register

val set_reg_type : t -> Reg.t -> Ctype.t -> unit

Sets the type of the register, leaves the value unchanged

val get_reg : t -> Reg.t -> tval

Get the content of the register with it's type

val get_reg_exp : t -> Reg.t -> exp

Get the content of the register without it's type

val update_reg_exp : t -> Reg.t -> (exp -> exp) -> unit

Apply a function to a register. Leave the type intact

Pc manipulation

val set_pc : pc:Reg.t -> t -> int -> unit

Set the PC to a concrete value and keep its type appropriate

val set_pc_sym : pc:Reg.t -> t -> Elf.Address.t -> unit

val bump_pc : pc:Reg.t -> t -> int -> unit

Bump a concrete PC by a concrete bump (generally the size of a non-branching instruction

val concretize_pc : pc:Reg.t -> t -> unit

Try to evaluate the PC if it is concrete

val set_last_pc : t -> Elf.Address.t -> unit

Set the last_pc of the state

Pretty printing

val pp : t -> Utils.Pp.document

val pp_partial : regs:Reg.t list -> t -> Utils.Pp.document

Print only the mentioned regs and the memory and asserts since the base_state. Until a better solution is found, the fenv will be printed entirely all the time