Module Base.Exp

Module for state expressions

include Exp.S with type var = var

type var = var

The type of variable provided in the functor

type t = (var, Ast.no) Exp.Typed.t

The type of expression on which this module works

val equal : t -> t -> bool

Test syntactic equality. a + b and b + a would test different under this predicate

val pp : t -> Utils.Pp.document

Pretty print the expression using PpExp

val pp_smt : t -> Utils.Pp.document

Pretty print the expression in SMTLIB language

val of_var : var -> t

Create an expression from a variable

val add_type : ('a, var, Ast.no, Ast.no) Ast.exp -> t

Convert a similar but untyped expression to an expression of type t

val of_reg : id -> Reg.t -> t

Create an expression from an register and a state id

val expect_sym_address : t -> Elf.Address.t

val of_section : size:int -> string -> t

val of_address : size:int -> Elf.Address.t -> t