Module Fragment.Block

This module provide the concept of memory block, as used by a fragment.

The block represent a memory area that contain a single memory expression.

It may represent a symbolic or concrete address. In the fist case, It may also be concretely bounded.

type t = private {

base : exp option;
(*
The symbolic base. If None the block is concrete
*)

offset : int;
(*
The concrete part of the address
*)

size : Size.t;

bounds : (int * int) option;
(*
Optional bounds: (min, max) means [min:max)
*)

}

The type for representing memory blocks

val make_split : ?bounds:(int * int) -> exp -> Size.t -> t

Make the block from an address and a size. The address is automatically split between a symbolic and concrete part.