Module Base.Var

This module provide state variables. Those are all symbolic variables that may appear in a state. If the same (in the Var.equal sense) variable appear in two state, that mean that when considered together, there is an implicit relation between them.

This means that the set of pair of concrete states represented by a pair of symbolic state could be a strict subset of the Cartesian product of the sets of concrete state represented by each symbolic state individually.

type t =

| Register of id * Reg.t
(*
The value of this register in this state
*)

| ReadVar of id * int * Ast.Size.t
(*
The result of a certain read in a certain state. The size part is not semantically important: Two ReadVar with same id and same number may no have different sizes
*)

| Arg of int
(*
A function argument
*)

| RetArg
(*
The address to which the return value should be written. This is used only in certain calling conventions
*)

| RetAddr
(*
The return address: The address to which a "return" instruction would jump.
*)

| NonDet of int * Ast.Size.t
(*
Variable representing non-determinism in the spec. Can only be a bit-vector for now.
*)

| Section of string
(*
Symbolic base address of ELF section. Assume 64bit for now.
*)

The type of a variable in the state

val to_string : t -> string

Convert the variable to the string encoding. For parsing infrastructure reason, the encoding must always contain at least one :.

val expect_register : t -> Reg.t

Expect a register variable and return the corresponding register. Throw otherwise.

val expect_readvar : t -> int

Expect a read variable and return the corresponding index. Throw otherwise.

val of_string : string -> t

The opposite of to_string. Will raise Invalid_argument when the string don't match

val of_reg : id -> Reg.t -> t

Create a register variable bound to the provided state id and register

val equal : t -> t -> bool

val hash : t -> int

Hashing function for variable. For now it's polymorphic but this may stop at any time

val pp : t -> Utils.Pp.document

Basically to_string in pp mode

val pp_bar : t -> Utils.Pp.document

Pretty prints but with bars around

val ty : t -> Reg.ty

Get the type of a variable

val new_nondet : Ast.Size.t -> t

Get a fresh NonDet variable