Module State.Fragment

This module define a memory fragment to be used by C types

Fragments

obj is just Ctype.t, but a bug in odoc hides that information. Fragment.obj actually do not exists.

The odoc PR is #349 on github, we just have to wait for it to be merged.

include Utils.RngMap.S with type obj := Ctype.t

type obj_off = Ctype.t * int

The type of an object with an offset

type t

The type of the map from address ranges to obj

val empty : t

An empty RngMap

val is_in : objaddr:int -> Ctype.t -> int -> bool

Test if an address is inside the object at address objaddr

val at : t -> int -> Ctype.t

Get the object containing the address. Throw Not_found if no object contains the address

val at_opt : t -> int -> Ctype.t option

Get the object containing the address. None if no object contains the address

val at_off : t -> int -> obj_off

Get the object containing the address and the offset of the address inside the object

at_off map addr = (obj, off) 

means:

   |                      |           |         |
 map 0                 obj start    point    obj end
   |<--------------addr-------------->|
                          |<---off--->|
                          |<------len obj------>|

In other words, at_off allow a change of coordinate from the map frame to the object frame.

Throw Not_found if no object contains the address

val at_off_opt : t -> int -> obj_off option

Get the object containing the address and the offset of the address inside the object. See at_off for more explanation.

None if no object contains the address

val update : (Ctype.t -> Ctype.t) -> t -> int -> t

Update the binding containing the provided address. If no binding contained the address, this is a no-op

val map : (Ctype.t -> Ctype.t) -> t -> t

Map a function over all the objects

val mapi : (int -> Ctype.t -> Ctype.t) -> t -> t

Map a function over all the objects with their address

val iter : (Ctype.t -> unit) -> t -> unit

Iter a function over all the objects

val iteri : (int -> Ctype.t -> unit) -> t -> unit

Iter a function over all the objects with their address

val clear_at : t -> int -> t

Clear the object containing the address if any

val clear : t -> pos:int -> len:int -> t

Clear an area of the RngMap.

If an object is partially in the specified block. It will be removed entirely.

See clear_crop for a different behavior. See clear_bounds to allow some bounds to be infinity.

val clear_crop : t -> pos:int -> len:int -> crop:(pos:int -> len:int -> Ctype.t -> Ctype.t) -> t

Clear an area of the RngMap.

If a block is partially in the specified block, It will be cropped by using the provided crop function.

crop ~pos ~len obj is supposed to crop the object obj and keep only the segment [pos:pos +len) of it (in the object coordinate frame).

val clear_bounds : ?start:int -> ?endp:int -> t -> t

Same as clear but if a bound is missing, then we erase until infinity in that direction. The target interval is [start:endp).

In particular clear_bounds map = empty.

val add : t -> int -> Ctype.t -> t

Add an object at a specific address. The whole range of addresses covered by the object must be free

val addp : t -> obj_off -> t

Add an object at a specific address. The whole range of addresses covered by the object must be free

val bindings : t -> (int * Ctype.t) list

Give the list of bindings

val to_seq : ?start:int -> ?endp:int -> t -> (int * Ctype.t) Utils.Seq.t

Return a sequence of all the object overlapping the range [start:endp). The first and last element may not be entierly contained in the ranged. If any bound is unspecified, it goes to infinity in that direction.

In particular to_seq map will iterate the entiere RngMap

val pp : t -> Utils.Pp.document

Environment

module Env : sig ... end

type env = Env.t