Module State.Reg

This module handle the register abstraction.

A register is defined by a Path and a type ty. The path is a representation of dot separated list of identifiers.

Registers are not part of the Arch module because they are discovered dynamically. This module keeps a global index of all register of the current architecture (in a Utils.IdMap). This map also fix the types of registers.

This allow to represent registers as integer everywhere.

The module also provides Map and PMap a respectively full and partial maps over registers. They need special support (especially the full map) because new registers may be added at any time after the creation of the map.

TODO: Support sail vectors (A path will be of type (string + int) list)

Paths

Representation of register path. The string reprensentation is with dots. A name may not contain dots, but this is not checked.

module Path : sig ... end

Registers

Global register properties and accessors

type t = private int

The type representing a register. The module invariant is that this type always contains a value bound in the global index and so this is always a valid register id.

type ty = Ast.no Ast.ty

The type of a register. This is isomorphic to Isla.ty. Use IslaConv.ty to convert

val of_int : int -> t option

Convert an integer into the corresponding register.

val mem_path : Path.t -> bool

Check if register is declared with that path

val mem_string : string -> bool

Check if a register is declared with that name. Same as Path.of_string |> mem_path

val of_path : Path.t -> t

Give the register corresponding to that path

val to_path : t -> Path.t

Give the path of a register

val of_string : string -> t

Give the register corresponding to a register name

val to_string : t -> string

Give the name of a register

val num : unit -> int

Give the current number of registers

val reg_type : t -> ty

Give the type of a register

val path_type : Path.t -> ty

Give the type of register path. Throw Not_found, it that path is not a declared register

val add : Path.t -> ty -> t

Add a new register to the global index. Return it's representation

val ensure_add : Path.t -> ty -> t

Ensure that a register with that path exists with that type, by adding it or checking it already exists with that type. Return the corresponding register

val adds : Path.t -> ty -> unit

Same as add but returns unit

val ensure_adds : Path.t -> ty -> unit

Ensure that a register with that path exists with that type, by adding it or checking it already exists with that type.

val iter : (Path.t -> t -> ty -> unit) -> unit

Run a function over all registers

val seq_all : unit -> t Utils.Seq.t

Returns a sequence of all registers

val (=) : t -> t -> bool

Equality predicate

val (<>) : t -> t -> bool

Inequality predicate

val compare : t -> t -> int

Compare according to an implementation-defined total order.

val pp : t -> Utils.Pp.document

Pretty prints a register (Just use to_string)

val pp_ty : ty -> Utils.Pp.document

Pretty prints a register type

val pp_index : unit -> Utils.Pp.document

Prints the register index (mainly for debugging I suppose)

Register map

To achieve a partial map on register, one may just used a plain Hashtbl. However as register is a finite type one may want to have a map where all the register are bound and thus access to a bound value cannot fail. This is complicated by the fact that new registers can be added after the creation of the map. To handle all those subtleties, there is the Utils.Map module.

module Map : sig ... end

This module provide a full map over register in the same way than Utils.FullVec provide a map of integers. It still need a generator to generate the value bound to not-yet-added registers.