Module Reg.Map

This module provide a full map over register in the same way than Utils.FullVec provide a map of integers. It still need a generator to generate the value bound to not-yet-added registers.

Because the domain of registers is finite, some extra function are available like iter and iteri that are not possible in Utils.FullVec.

If a register is added with add, it is automatically and implicitly added to the Utils.Map and the generator must accept this new value. The generator will never be called on invalid register values (i.e. when the generator is called on a register, the former can get the latter's type and name with reg_type and to_string)

type reg = t

type 'a t

The type of the complete map

val init : (reg -> 'a) -> 'a t

Initialize the map with a generator

val reinit : 'a t -> (reg -> 'a) -> unit

Clear the map and restart with this generator

val copy : 'a t -> 'a t

Make a copy of the map

val set : 'a t -> reg -> 'a -> unit

Set the value of a register

val get : 'a t -> reg -> 'a

Get the value of a register

val map : ('a -> 'b) -> 'a t -> 'b t

Map the function all the registers (including future, not yet added ones)

val mapi : (reg -> 'a -> 'b) -> 'a t -> 'b t

Same as map but with the index

val map_mut : ('a -> 'a) -> 'a t -> unit

Map the function on all the register by mutation (including future ones)

val map_mut_current : ('a -> 'a) -> 'a t -> unit

Map the function on all current register. Future registers are unchanged

val iter : ('a -> unit) -> 'a t -> unit

Iterate over all the value of all currently present registers

val iteri : (reg -> 'a -> unit) -> 'a t -> unit

Same as iter but also with the register index

val bindings : 'a t -> (reg * 'a) list

Give all the registers bindings

val pp : ('a -> Utils.Pp.document) -> 'a t -> Utils.Pp.document

Contrary to Utils.FullVector.pp, this one will print the binding of all registers, and may call the generator to do that