Module State.Simplify

This module provide utility to simplify states

module Z3St : sig ... end

val ctxfree : Base.t -> unit

Context free simplify, just expression by expression. Do it mutably

module ContextFull : sig ... end

Do a context aware simplify, for now, it just does a context free simplify and then remove assertion that proven true or false by Z3

val ctxfull : Base.t -> unit

Do a context aware simplify, for now, it just does a context free simplify and then remove assertion that proven true or false by Z3.

If a state has a false assertion anywhere then all assertions will collapse in a single false.