Module State.SymbolicBytes

This module provide high-level support for a symbolic array of bytes.

The main differences between this and a big concat expression is that:

It allows some bytes to be undefined

It allows fast (logarithmic or better) access to any byte or group of bytes.

It provides dedicated functions for efficiently extracting an expression from a range of bytes (reading) or blitting (writing) an expression to a range of bytes.

The indexing can start at any point (negative value can be mapped) without any costs

This data structure deliberately do not have any infrastructure to read or write bytes at symbolic positions. See SymbolicFragment for that.

This data structure do not have a concept of beginning or an end (But it has the concept of first defined byte and last defined byte). In particular addresses can be negative.

It is functorized of the type of variables (Var) to get variable equality and pretty-printing

Currently this has a pure interface.

module type S = sig ... end

module Make : functor (Var : Exp.Var) -> S with type var = Var.t