Module SymbolicBytes.Make

Parameters

Var : Exp.Var

Signature

type var = Var.t

The type of variables used

type exp = (var, Ast.no) Exp.Typed.t

type t

The type of symbolic bytes.

val empty : t

The empty symbolic byte sequence with no bytes defined

val sub : pos:int -> len:int -> t -> exp option

Extract an expression in [pos:pos+len). If any bytes in the range is undefined, Then None is returned, otherwise a expression is returned

val blit_exp : exp -> pos:int -> len:int -> t -> t

Write the expresion on the interval [pos:pos +len) of the bytes. The expression must be a bitvector of size exactly 8 * len.

TODO check it when values are type annotated

val clear_bounds : ?⁠start:int -> ?⁠endp:int -> t -> t

Clear a range of the symbolic bytes, making all those bytes undefined again. If a bound is missing, it means up to infinity in that direction

val map_exp : (exp -> exp) -> t -> t

Map a function over all the contained expressions

val iter_exp : (exp -> unit) -> t -> unit

Iter a function over all the contained expressions

val pp : t -> Utils.Pp.document

Pretty prints the symbolic bytes