Module Base.Mem

This module manages the memory part of the state.

The symbolic memory bounds certain symbolic address to symbolic values, but most addresses are not bound. This basically correspond to a concrete instantiation of a SymbolicFragment.

However in most case, we have some information that some symbolic value do not alias other symbolic values.For example in no address involving the stack pointer may alias any address no involving the stack pointer except in specific case of escaping which are explicitly not supported (yet). It is intended to support escaping later. There can be other kind of non-aliasing information in case of explicit restrict annotation or when implicitly passing or returning value by pointer (which some ABI do when such value are too big).

In the absence of escaping, such problem can be solved by representing the memory with multiple SymbolicFragment. One for the main memory, and one for each "restricted block". The module encapsulate all those different fragment.

To manage such a system, we use the C type system to carry around provenance information in the type Ctype.provenance. When doing a read or a write, this provenance is used to route the read or write to the right symbolic block.

This means that the provenance-tracking part of the C type system must be part of the TCB of read-dwarf as the soundness of the symbolic memory model depend on it.

Finally, the current implementation is only suitable for sequential execution, a new theoretical model and implementation must be developed for concurrent shared memory.

Implementation detail: This type has an imperative interface even if the underlying SymbolicFragment has a pure interface.

type t

val new_frag : t -> exp -> Ctype.provenance

Add a new fragment with the specified base

module Fragment : State__.SymbolicFragment.S with type var = Var.t

val get_main : t -> Fragment.t

Get the main fragment of memory