Module Fragment.Env

type frag = t

type t = {

frags : frag Utils.Vec.t;

}

val make : unit -> t

val copy : t -> t

val add_typ : addr:int -> Ctype.t -> t -> id:int -> unit

Add the provided type at addr into a fragment indexed by id in the environment

val adds_frag : ?⁠frag:frag -> t -> unit

Add a new fragment to the environment. empty by default.

val add_frag : ?⁠frag:frag -> t -> int

Add a new fragment to the environment, returns the id of the new fragment. frag is empty by default.

val get : t -> int -> frag

val set : t -> int -> frag -> unit

val pp : t -> Utils.Pp.document