Module State__Base.Tval

Module for optionally typed state expressions. Those are symbolic values which may or may not have a C type.

type t = {

ctyp : Ctype.t option;

exp : exp;

}

val make : ?⁠ctyp:Ctype.t -> exp -> t

Make a new typed value with optionally a Ctype

val of_exp : ?⁠ctyp:Ctype.t -> exp -> t

val of_var : ?⁠ctyp:Ctype.t -> var -> t

val of_reg : ?⁠ctyp:Ctype.t -> id -> State.Reg.t -> t

val map_exp : (exp -> exp) -> t -> t

val iter_exp : (exp -> 'a) -> t -> 'a

val exp : t -> exp

val ctyp : t -> Ctype.t option

val equal : t -> t -> bool

val pp : t -> PPrintEngine.document