Module State__SymbolicFragment.Make

A functor to make a memory fragment module from a variable module Var

Parameters

Var : Exp.Var

Signature

type var = Var.t

The type of variables used

module Size = Ast.Size

type exp = (var, Ast.no) Exp.Typed.t

The type of expressions stored in the fragment

module Block : sig ... end

This module provide the concept of memory block, as used by a fragment.

type t

The type of a memory fragment

val empty : t

The empty memory fragment. Any read will be symbolic

val from : t -> t

Start a new memory fragment based on the previous one

val try_read : t -> Block.t -> exp option

Try to read a expression in a block. If one can provide a symbolic expression representing the content of the block then Some is returned, otherwise None is returned.

val try_read_naive : t -> Block.t -> exp option

Same semantic as try_read but ignores the caches. Is supposed to be slower.

The required property is that if try_read return Some value then try_read_naive must return the same value. It's possible that try_read_naive give a result when try_read don't

val read_sym : t -> Block.t -> var -> t

Read a symbolic variable from a block. This bound this symbolic variable to the The content of the block in the current memory state.contents

val write : t -> Block.t -> exp -> t

Write a symbolic expression at a block

val map_exp : (exp -> exp) -> t -> t

Map a function over all contained expressions. This function must not change the semantic meaning of symbolic expressions in any way. It is intended to be used with simplifying functions and the like

val iter_exp : (exp -> unit) -> t -> unit

Iter a function over all contained expression. Expression may appear more or less than anticipated because of various caching.

val check_cache : t -> unit

Check cache integrity. Throw if something is wrong

val is_empty : t -> bool

Tells if the memory is empty since it's initial base. (No new memory operation were added

val pp_raw : t -> Utils.Pp.document

Pretty prints the raw internals fragment. TODO: A nice pretty printer