Module Trace.Base

This module defines a new simplified kind of trace to replace Isla traces in the later stages of the instruction semantics processing.

The traces are even simpler and more easily typable. The possible events are in the type event and traces (t) are just list of them.

Compared to Isla, the concept of reading a register do not exist anymore. Nor the concept of pure symbolic variable or Sail structured values. Instead expression can contain only registers and results of previous memory read as decribe in the type Var.t. All writing event directly write an entire expression. There are no intermediary variable definitions.

This raise a problem that if an isla trace reads a register after having written to it, then this is ambiguous to represent.

Thus a partially monadic representation has been chosen:

The effect of writes on registers are delayed to the end, which means that registers variable value in every expression is the value of that register at the beginning of the trace. In particular, if there are two write to the register, only the last one have any effect, the other can be deleted.

The memory operation behave in normal monadic way, in particular a memory read can read the value written by a previous write in the same instruction, even if this is very unlikely.

In the end, both assertion and register write to different registers can be reordered at will.

TODO: To make it more clean, getting register writes and assertions out of the trace would make sense like:

type mem_event = Read of ... | Write of ...
type t = { asserts : exp list; reg_writes : (Reg.t * exp) list; mem: mem_event list }

For all those reason, concatenating two trace semantically is very different that concatenating list of event, and is not implemented yet.

The important functions are of_isla to convert and Isla traces and simplify for simplify traces.

module Var : sig ... end

This module contains variable used in traces

module ExpPp = Exp.Pp

A trace expression. No let bindings, no memory operations

module Typed = Exp.Typed

module Exp : sig ... end

type exp = Exp.t

type event =

| WriteReg of {

reg : State.Reg.t;

value : exp;

}

| ReadMem of {

addr : exp;

value : int;

size : Ast.Size.t;

}

| WriteMem of {

addr : exp;

value : exp;

size : Ast.Size.t;

}

| Assert of exp

The event type. See the module description for more details

type t = event list

val iter_var : (Exp.var -> unit) -> event list -> unit

Pretty printing

val pp_exp : ('a, Var.t, Ast.no, Ast.no) Ast.exp -> Utils.Pp.document

Pretty print an expression

val pp_event : event -> Utils.Pp.document

Pretty print an event

val pp : event list -> Utils.Pp.document

Pretty print a trace

Isla to Trace conversion

This section perform the conversion from Isla trace to the traces of this module.

The conversion is generrally obvious, however there is subtlety: If the Isla trace reads a register after having written it, then the read produce the written expression instead of just the symbolic value of that register. That why there is a written_registers parameter to some function of this section.

exception OfIslaError

Throw an error in case of local conversion error. Normally a type-checked Isla trace should not fail in this section

type value_context = exp Utils.HashVector.t

The context mapping Isla variable numbers to trace expression

val get_var : 'a Utils.HashVector.t -> int -> 'a

Get the exression of the variable at the index. Throw OfIslaError if the variable is not bound

val exp_conv_subst : value_context -> Isla.rexp -> exp

Convert an Isla expression to a Trace expression by replacing all Isla variable by their value in the context. Throw OfIslaError if the substitution fails

val exp_of_valu : Isla_lang.AST.lrng -> exp Utils.HashVector.t -> Isla.valu -> exp

Convert an Isla.valu in a expression

val write_to_valu : 'a Utils.HashVector.t -> Isla.valu -> 'a -> unit

Write an expression to an Isla.valu

val events_of_isla : segments_map:(string * int) Utils.HashVector.t -> written_registers:(State.Reg.t, exp) Stdlib.Hashtbl.t -> read_counter:Utils.Counter.t -> vc:value_context -> Isla.revent -> event list

Convert an isla event to Trace events, most events are deleted

val of_isla : Isla.segment list -> Isla.rtrc -> t

Top level function to convert an isla trace to one of this module

Trace simplification

module SimpContext : sig ... end

A instance of Z3.ContextCounter.

module Z3Tr : sig ... end

module VarTbl : sig ... end

val declare_non_det : Z3.server -> event list -> unit

val simplify : event list -> event list

Simplify a trace by using Z3. Perform both local expression simplification and global assertion removal (when an assertion is always true)