Module Base.Var

This module contains variable used in traces

module Reg = State.Reg

type t =

| Register of Reg.t
(*
The value of the register at the beginning of the trace
*)

| Read of int * Ast.Size.t
(*
The result of that memory reading operation
*)

| NonDet of int * Ast.Size.t
(*
Variable representing non-determinism in the spec
*)

| Segment of string * int
(*
Variable representing symbolic segment in the opcode
*)

A trace variable

val to_string : t -> string

Convert the variable to the string encoding. For parsing infractructure reason, the encoding must always contain at least one :.

val of_string : string -> t

Inverse of to_string

val pp : t -> Utils.Pp.document

Pretty prints the variable

val equal : 'a -> 'a -> bool

val hash : 'a -> int

val ty : t -> Reg.ty

val of_reg : Reg.t -> t