Module Cache.Traces

Store element of type Base.tlist on disk.

First a hashmap from register numbers to path and type is stored. Then the actual tracelist is marshaled.

When loading the register numbering may be different, and some may not exist. So we create all missing register, then we replace the old number with the new ones.

type t = Base.t list

module Reg = State.Reg

type regs = (Reg.t, Reg.Path.t * Reg.ty) Stdlib.Hashtbl.t

val to_file : string -> t -> unit

val of_file : string -> t