Module Trace.Context

This module provide the type for a context to run a trace

Any information that should be required to run a trace but is not part of the state itself should be added here

module SMap : sig ... end

type t = {

reg_writes : (State.Reg.t * State.tval) Utils.Vec.t;
(*
Stores the delayed register writes
*)

mem_reads : State.tval Utils.HashVector.t;
(*
Stores the result of memory reads
*)

nondets : State.var Utils.HashVector.t;
(*
Stores the mapping of nondet variables
*)

state : State.t;

segments : State.exp SMap.t;

asserts : State.exp list;

dwarf : Dw.t option;
(*
Optionally DWARF information. If present, typing is enabled
*)

}

The context to run a trace

val make_context : ?dwarf:Dw.t -> ?relocation:Elf.Relocations.rel -> State.t -> t

Build a context from a state

val expand_var : ctxt:t -> Base.Var.t -> Ast.no Ast.ty -> State.exp

Expand a Trace variable to a State expression, using the context

val typing_enabled : ctxt:t -> bool

Tell if typing should enabled with this context

module Z3St = State.Simplify.Z3St

val simplify : ctxt:t -> State.exp -> State.exp