Module Trace.Run

This module is for running trace from Trace like Isla.Run runs Isla traces.

Due to the semantic of a register access being the register at the beginning of the trace, all register writes are not done immediately but delayed and stored in the context.

Typing is enabled if Context.typing_enabled returns true for functions that take a context. For other functions, typing is enabled if the dwarf optional argument is passed

module Ctxt = Context

type ctxt = Ctxt.t

val expand : ctxt:ctxt -> Base.exp -> State.exp

Expand a Trace expression to a State expression, using the context

val expand_simplify : ctxt:ctxt -> Base.exp -> State.exp

val expand_tval : ctxt:ctxt -> Base.exp -> State.tval

Expand a Trace expression to a typed State expression, using the context.

If the context enables typing, the expression will actually be typed, otherwise the type will be None

val event_mut : ctxt:ctxt -> Base.event -> unit

Run the event. The modified state is the one inside ctxt.

val trace_mut : ?dwarf:Dw.t -> ?relocation:Elf.Relocations.rel -> State.t -> Base.t -> unit

Run a trace on the provided state by mutation. Enable typing if dwarf is provided

val trace : ?dwarf:Dw.t -> ?relocation:Elf.Relocations.rel -> State.t -> Base.t -> State.t

Run a trace on the provided state by returning an updated copy.

val trace_pc_mut : ?dwarf:Dw.t -> ?relocation:Elf.Relocations.rel -> next:int -> State.t -> Base.t -> unit

Run a trace by mutating the provided state including it's PC. If the trace modified the PC then nothing is done otherwise next is added to it.

Thus this function automatically handle moving the PC for fall-through instruction