Module Trace__.Typer

module Value = Exp.Value

module ConcreteEval = Exp.ConcreteEval

module Typed = Exp.Typed

module Fragment = State.Fragment

Expressions

This section implements all the C type expression inference rules.

C Types are always optional and if any part of an expression is not typed, then the result is not typed.

type tval = {

ctyp : Ctype.t option;

exp : Trace.Base.exp;

}

A typed Base.exp

val get_ctypo : tval -> Ctype.t option

val get_ctyp : tval -> Ctype.t

val machine_of_size : ?⁠update:int -> Ctype.t -> Ctype.t

Output a Machine type of the same size at the original type. If update has a value, the size if modified by the update value.

val unop : Ast.unop -> tval -> Ctype.t option

val constexpr_to_int : ctxt:Trace.Context.t -> ('a, Trace.Base.Var.t, Ast.no, Ast.no) Ast.exp -> int

val binop : ctxt:Trace.Context.t -> Ast.no Ast.binop -> tval -> tval -> Ctype.t option

val manyop : ctxt:Trace.Context.t -> Ast.manyop -> tval list -> Ctype.t option

val expr : ctxt:Trace.Context.t -> Trace.Base.exp -> Ctype.t option

Stage 1 expression typer

val expr_tval : ctxt:Trace.Context.t -> (Ast.no Ast.ty, Trace.Base.Exp.var, Ast.no, Ast.no) Ast.exp -> tval

Same as expr but return the expression in a Tval

Memory Read

When reading memory there are two main cases:

The address has a pointer type, in which case we figure out the type of the read according to that pointer type.

The address do not have a pointer type, so we try to do an untyped read with State.read_noprov which may fail.

val fragment_at : dwarf:Dw.t -> fenv:Fragment.Env.t -> size:int -> Ctype.fragment -> int -> Ctype.t option

val ptr_deref : dwarf:Dw.t -> fenv:Fragment.Env.t -> size:int -> Ctype.fragment -> Ctype.offset -> Ctype.t option

val read : dwarf:Dw.t -> State.t -> ?⁠ptrtype:Ctype.t -> addr:State.exp -> size:Ast.Size.t -> State.tval

Does the same as State.read, but additionally take care of reading the type from a fragment and marking the type of the read variable.

Memory Write

val fragment_write_at : dwarf:Dw.t -> fenv:Fragment.Env.t -> ctyp:Ctype.t -> Ctype.fragment -> int -> unit

val ptr_write : dwarf:Dw.t -> fenv:Fragment.Env.t -> ctyp:Ctype.t -> Ctype.fragment -> Ctype.offset -> unit

val write : dwarf:Dw.t -> State.t -> ?⁠ptrtype:Ctype.t -> addr:State.exp -> size:Ast.Size.t -> State.tval -> unit

Does the same as State.write, but additionally take care of writing the type if the write is on a Ctype.FreeFragment.