Module Trace__Instr

This module provide the representation of an instruction. It only contains generic information about the opcode and not specific information about a place in the code.

module Reg = State.Reg

type trace_meta = {

trace : Trace.Base.t;

jump_target : Trace.Base.exp option;

read : Reg.t list;

written : Reg.t list;

}

A simple trace with its metadata

If there is a WriteReg {reg;value} event in trace where reg = Arch.pc (), jump = Some value, otherwise it is None. If there are multiple such events, it will store the value of the last one.

type t = {

traces : trace_meta list;

length : int;
Bytes length

read : Reg.t list;

written : Reg.t list;

opcode : Utils.BytesSeq.t;

}

A full instruction representation

val dedup_regs : State.Reg.t list -> State.Reg.t list

val footprint : t -> State.Reg.t list

val trace_meta_of_trace : Trace.Base.t -> trace_meta

Compute the metadata of trace

val of_traces : Utils.BytesSeq.t -> Trace.Base.t list -> t

Generate full instruction data from a list of traces

val pp : t -> Utils.Pp.document

Pretty print the representation of an instruction