Module Utils

module Array : sig ... end

This module is for extending the Array module of the standard library

module BitVec : sig ... end

This module provides an interface for a bit vector of dynamic size.

module Bits : sig ... end

Like bytes, but for bit level manipulation. The underlying type is still bytes and thus the size has to be a multiple of 8.

module BytesSeq : sig ... end

This module represent a byte sub view on a bytes object. Contrary to Bytes it is a non-owning immutable view. It do not prevent the original bytes from being modified, and the changes will be propagated in the view. It is additional sugar on top of Linksem's Byte_sequence_wrapper

module Cache : sig ... end

This module implement a caching system i.e a persistant structure stored on the disk.

module Cmd : sig ... end

This module provides high-level interaction with external processes.

module CmdlinerHelper : sig ... end

This module provide some Cmdliner helper functions.

module Counter : sig ... end

This module provide a small counter object which is just a int reference on which get can be called to get an identifier and increment the reference

module Files : sig ... end

This module provides simplified file management and some channel interaction function.

module FullVec : sig ... end

A full vector is a vector in which all non-negative integer are bound.

module Fun : sig ... end

More functional combinator. This module extends the base OCaml API of Fun.

module HashVector : sig ... end

An hash vector allow a vector to behave as hash map indexed by small integers. It is hash map with the identity hash function.

module IdMap : sig ... end

An IdMap is a map that associate an id to each key (and thus to each value).

module IntBits : sig ... end

Manipulate an int as bitfield of size 31 or 63.

module List : sig ... end

Extension of the List module of the standard library.

module Logs : sig ... end

This module provide a logging system for each module.

module Option : sig ... end

This module extends the base OCaml API of Option.

module Pair : sig ... end

This module contain random utility functions dealing with pairs

module Pp : sig ... end

This module provide all pretty printing functionality. It's main goal is not directly handle the output but to handle how to layout complex data structure in a text format.

module Protect : sig ... end

This module provide try-with-finally kind of exception handling.

module Raise : sig ... end

This module provide convenience facilities to raise exception or other exception management

module RngMap : sig ... end

A module giving a map indexing data with address ranges and providing access to quick access to data corresponding to any value in-between.

module Seq : sig ... end

This module is for extending the Seq module of the standard library

module String : sig ... end

Extension of the String module of the standard library.

module Sym : sig ... end

module Vec : sig ... end

This module provide a resizable array. It is a rename of Res.array with added features

module WeakMap : sig ... end

This module provide a simple way of associated key to value without keeping them alive The key is owned by the map but the value is weakly pointed to. When the value is deleted by the GC, the binding disappears.

module WeakPtr : sig ... end

This module provide a simple weak pointer that can be rendered invalid by the GC