Module Test.Single

type value = Value.t

type t = Single(Value).t

The type that represent the cache in RAM.

val make : ?⁠fake:bool -> string -> t

Build a new cache management object with a name If fake is set, the cache will not touch the disk and behave as a plain ref.

val get_opt : t -> value option

Get a value from the cache or None if no value is stored

val get : t -> value

Get a value from the cache or throws Not_found if no value is bound

val set : t -> value -> unit

Set the value in the cache

val clear : t -> unit

Clear the value in the cache