Parameter Make.Key

include Stdlib.Hashtbl.HashedType

type t

val equal : t -> t -> bool

val hash : t -> int

val to_file : string -> t -> unit

Write the necessary information to retrieve the key in a file. The filename supplied is without the key extension. The implementer must call to_keyfile on it

val of_file : int -> string -> t

Build back the key from the hash and a storage file. The filename supplied is without the key extension. The implementer must call to_keyfile on it.

If you start with a key and a file, then doing to_file file key and then of_file (hash key) file must return something equal to key.