Module type Cache.Key

The interface of keys.

Key are use to index values in the cache, however instead of using the key directly as a file name, the hexadecimal hash of the key is used trough int_to_file. See Make for more details.

There are two main way of managing keys:

The hash is injective, in which case there is no need for extra information.

The hash is not injective, in which case extra information required to disambiguate between different value with the same hash can be used. This information is stored in the file obtained with to_keyfile file. See to_file.

An implementation can do a mix of both i.e. Putting extra information only for value for which there is hash collision.

include Stdlib.Hashtbl.HashedType

type t

val equal : t -> t -> bool

val hash : t -> int

val to_file : string -> t -> unit

Write the necessary information to retrieve the key in a file. The filename supplied is without the key extension. The implementer must call to_keyfile on it

val of_file : int -> string -> t

Build back the key from the hash and a storage file. The filename supplied is without the key extension. The implementer must call to_keyfile on it.

If you start with a key and a file, then doing to_file file key and then of_file (hash key) file must return something equal to key.