Module Utils.Cmd

This module provides high-level interaction with external processes.

This provide a functionality similar to Bos), but this lib is still unstable.

There are two main mode of communication provided:

Interaction with pipes: capture the stdin or stdout of the command and use them

Interaction with socket: Create a socket on which the child program can connect to.

Programs can be launched in two modes

Calling: They are called like a function and we wait for them to give a result before contnuing

Server: The are launch as a background process that stay there and can be called.

Calling programs can be done with call* function like call, call_read, call_send and call_send_read and only support pipe interaction

Server like setups can be done in pipe mode with IOServer and in socket mode with SocketServer.

type cmd = string array

The type of a command to be sent. The program to call must be the item 0 of the array

exception Crash of cmd * Unix.process_status

If a program do not return with a 0 exit code, we throw that exception giving the command that failed and the invalid status it returned

Pipe calling

val call : cmd -> unit

Call the command without redirecting anything. Wait for completion before returning.

May throw Crash on error.

val call_send : cmd -> sender:(Stdlib.out_channel -> unit) -> unit

Call the command and then call sender to send it some data on it's stdin. Then wait for completion.

May throw Crash on error.

val call_send_string : cmd -> string -> unit

Call the command, send it the string on it's standard input and wait for completion.

May throw Crash on error.

val call_read : cmd -> reader:(Stdlib.in_channel -> 'a) -> 'a

Call the command and then call reader to parse what the command outputs on it's stdout. Then wait for completion and return the parsed value

May throw Crash on error.

val call_read_string : cmd -> string

Call the command, wait for completion, and return it's stdout in a string

val call_send_read : cmd -> sender:(Stdlib.out_channel -> unit) -> reader:(Stdlib.in_channel -> 'a) -> 'a

Call the command and then call sender to send the input data on it's stdin. Then call reader to parse an answer from stdout Then wait for completion and return the parsed value

May throw Crash on error.

Pipe server

module IOServer : sig ... end

This module provide functionality to run command in the background and communicate with it via redirection on it's standard input and output.

Socket server

module SocketServer : sig ... end

This module provide functionality for a socket server with which one can communicate on a sockets.