Module Cmd.IOServer

This module provide functionality to run command in the background and communicate with it via redirection on it's standard input and output.

An example of use is in Z3

type t = {

cmd : cmd;

input : Stdlib.in_channel;
(*
The output of the command from which answer can be read
*)

output : Stdlib.out_channel;
(*
The input of the server on which request can be sent
*)

}

The type of pipe IO server

val start : cmd -> t

Start the server using the specified command

val stop : t -> unit

Stop the server.

May throw Crash on error.