Module Utils.CmdlinerHelper

This module provide some Cmdliner helper functions.

val setter : 'a Stdlib.ref -> 'a Cmdliner.Term.t -> unit Cmdliner.Term.t

Return a unit Term that evaluates the input term and set the reference to the resulting value.

Due to the nature of Cmdliner, this may be evaluated multiple times, so setting the reference to anything else is dangerous.

val add_option : unit Cmdliner.Term.t -> 'a Cmdliner.Term.t -> 'a Cmdliner.Term.t

Add an unit term that need to be evaluated at the same time at the main term.

The order of evaluation is unspecified, but the option will be evaluated before the resulting term is returned.

val add_options : unit Cmdliner.Term.t list -> 'a Cmdliner.Term.t -> 'a Cmdliner.Term.t

Fold add_option on a list of option

val func_option : unit Cmdliner.Term.t -> 'a -> 'a Cmdliner.Term.t

Replaces Term.const but allow a unit terms (like the one generated by setter) to be evaluated before the function is called

val func_options : unit Cmdliner.Term.t list -> 'a -> 'a Cmdliner.Term.t

Same as func_option but with a list of unit terms