Module Utils.Files

This module provides simplified file management and some channel interaction function.

The functions read and write are about dealing with a whole file at once without caring about opening or closing it.

type 'a reader = Stdlib.in_channel -> 'a

The type of a reader that read an object from a channel. Function like input_* in Stdlib

type 'a writer = Stdlib.out_channel -> 'a -> unit

The type of a writer of 'a. Function named output_* in Stdlib

IO functions

val double_byte : bytes -> bytes

Double the size of a bytes object (technical function)

val input_bytes : Stdlib.in_channel -> bytes

Read the whole channel in a bytes

val input_string : Stdlib.in_channel -> string

Read the whole channel in a string

val input_sexp : Stdlib.in_channel -> string

Reads a S-expression from the input, line by line. When the sexp finishes, there should be nothing else on the line i.e. the last closing parenthesis should be followed by a new line.

val input_list : 'a reader -> Stdlib.in_channel -> 'a list

Try the reader until it fails with End_of_file and then build the list of all the successfully read objects in order.

val input_array : 'a reader -> Stdlib.in_channel -> 'a array

Try the reader until it fails with End_of_file and then build the array of all the successfully read objects in order.

val output_list : 'a writer -> Stdlib.out_channel -> 'a list -> unit

Output all the element of the list in order with the provided writer

Direct file IO

val read : 'a reader -> string -> 'a

Take a reader and a file and read an object from the file using the reader. Text mode

val read_bin : 'a reader -> string -> 'a

Take a reader and a file and read an object from the file using the reader. Binary mode

val write : 'a writer -> string -> 'a -> unit

Take a writer a file and object and write the object to the file using writer. Text mode

val write_bin : 'a writer -> string -> 'a -> unit

Take a writer fs a file and object and write the object to the file using writer. Binary mode

val read_string : string -> string

Return the content of specified file as a string

val write_string : string -> string -> unit

write_string file cont write cont in file which is overwritten if it exists

File management

val remove : string -> unit

Remove a file

val exists : string -> bool

Test if a files exists

val remove_at_exit : string -> unit

Remove a file at program exit

val add_to_relative : newp:string -> string -> string

Add newp in front of a relative path. If the path is not relative, then it is unchanged