Module Utils.HashVector

An hash vector allow a vector to behave as hash map indexed by small integers. It is hash map with the identity hash function.

If all goes well it has the same API as (int, 'a) Hashtbl.t for positive integers. Any attempt to use a negative integer will raise Invalid_argument.

type 'a t

The type of the hash vector

val mem : 'a t -> int -> bool

Check if an element is set

val copy : 'a t -> 'a t

Copy the HashVector

exception Exists

val set : 'a t -> int -> 'a -> unit

Set a value. Create a new binding if necessary

val clear : 'a t -> int -> unit

Clear a binding, do nothing is the value is not bound.

val add : 'a t -> int -> 'a -> unit

Create a new binding. Throws Exists if a binding already exists

val get_opt : 'a t -> int -> 'a option

Return Some v if v is bound to the integer in the hash vector. Return None if nothing is bound

val get : 'a t -> int -> 'a

Return the value bound to the integer. Throw Not_found if the integer is not bound

val empty : unit -> 'a t

Create an empty hashVector.t

val bindings : 'a t -> (int * 'a) list

Returns a list of bindings in the hash vector

val of_seq : (int * 'a) Seq.t -> 'a t

Fill the hashVector from a sequence

val pp : ('a -> Pp.document) -> 'a t -> Pp.document

Pretty print a hashVector