Module Utils.Logs

This module provide a logging system for each module.

The logging system is intended to print on stderr or stdout (and maybe in file later) messages of the form:

[Module][Level] Message here

For example:

[Z3][Info] Z3 server started with version 4.8.7

If you just want to use it to log things in your own module, just insert at the start of the module the following code:

open Logs.Logger (struct
  let str = __MODULE__
end)

You can then use all the function in Logger to print logging messages in your module.

The rest of this module is about manipulation log level in the different modules. You can dynamically interact with the logs level from the command line by using CommonOpt.logs

type 'a printf = ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

The type of normal logging function

type ('a, 'b) printf_fatal = ('a, Stdlib.out_channel, unit, 'b) Stdlib.format4 -> 'a

The type of a failing logging function, that will exit after printing it's message

type level =

| Base
(*
The actual output. The only thing printed in quiet mode. Should only appear in CLI modules
*)

| Err
(*
An error message
*)

| Warn
(*
An warning
*)

| Info
(*
Details on what happens that should be understandable if the person do not know the corresponding module
*)

| Debug
(*
Everything happening in detail
*)

The type of log level

val pp_level : level -> Pp.document

Pretty prints a level

val level_to_string : level -> string

Convert level to string

val level_to_header : level -> string

Convert level to string header like "[Error]". Base is the empty string

val set_default_level : level -> unit

Set a default level to all the modules. Erase local customized levels

val set_level : string -> level -> unit

Set level of a module by name

val level_conv : level Cmdliner.Arg.conv

Parser for log level on command line

val set_stdout_level : level -> unit

Set level below which the output goes to stdout

module type String = sig ... end

module Logger (S : String) : sig ... end

This declare a logger instance. The string parameter is the logger name. It should be the OCaml module name. The normal way of instantiating this module is:

val process_opts : bool -> bool list -> string list -> string list -> level -> unit

val term : unit Cmdliner.Term.t