Module Logs.Logger

This declare a logger instance. The string parameter is the logger name. It should be the OCaml module name. The normal way of instantiating this module is:

open Logs.Logger (struct
  let str = __MODULE__
end)

This module provide a lot of logging function, They only provide a format string interface but with Pp.top on may print efficiently but lazily arbitrary Pp.document to the logs.

Some examples:

 warn "This weird thing happened, I choose that default behavior for case %s" case 

 debug "Function ... received value %t" Pp.(top printer object) 

Parameters

module S : String

Signature

val set_level : level -> unit

Override the level of this logger. It may still be overridden by the command line

val get_level : unit -> level

Get the current level

val log : level -> 'a printf

Log a specific level using a format string

val log_fatal : code:int -> level -> ('a, 'b) printf_fatal

Log a fatal problem with format string then shutdown with code

val base : 'a printf

Base command line output asked for by the CLI. Level is Base

val fail : ('a, 'b) printf_fatal

Failure due to external circumstances, generally wrong user input, then exit with code 1. Level is Base

val err : 'a printf

Declare a non-fatal internal error. Level is Err

val fatal : ('a, 'b) printf_fatal

Declare a fatal internal error then exit with code 2. Level is Err

val warn : 'a printf

Raise a warning. Level is Warn

val info : 'a printf

Print general information about what's happening. Level is Info

val debug : 'a printf

Print debugging information about what's happening. Level is Debug

val has_debug : unit -> bool

Tell if debug logging is enabled