Module Utils.Option

This module extends the base OCaml API of Option.

In particular, it adds:

Monadic bindings

Option merging

Option lifting of list and pairs.

include module type of Stdlib.Option

type !'a t = 'a option =

| None

| Some of 'a

val none : 'a option

val some : 'a -> 'a option

val value : 'a option -> default:'a -> 'a

val get : 'a option -> 'a

val bind : 'a option -> ('a -> 'b option) -> 'b option

val join : 'a option option -> 'a option

val map : ('a -> 'b) -> 'a option -> 'b option

val fold : none:'a -> some:('b -> 'a) -> 'b option -> 'a

val iter : ('a -> unit) -> 'a option -> unit

val is_none : 'a option -> bool

val is_some : 'a option -> bool

val equal : ('a -> 'a -> bool) -> 'a option -> 'a option -> bool

val compare : ('a -> 'a -> int) -> 'a option -> 'a option -> int

val to_result : none:'e -> 'a option -> ('a, 'e) Stdlib.result

val to_list : 'a option -> 'a list

val to_seq : 'a option -> 'a Stdlib.Seq.t

Utility

val take_first : 'a option -> 'a option -> 'a option

Take the value in the first argument if there is one, otherwise take the value in the second argument, otherwise None

val (|||) : 'a option -> 'a option -> 'a option

take_first as an operator:

Behave like boolean or but keep the value of the first option that gave true. This is associative but obviously not commutative.

val take_first_list : 'a option list -> 'a option

Take the value of the first Some in the list. returns None if all the option were None

val take_all : 'a option -> 'b option -> ('a * 'b) option

If both option have values, give Some of the pair, otherwise None

val (&&&) : 'a option -> 'b option -> ('a * 'b) option

Take_all as an operator:

Behave like boolean and but keep all the value of the options This is not associative because at type level (a * b) * c is not a * (b * c). Using monadic bindings is recommended for more that 2 operands.

val value_fail : 'a option -> ('b, unit, string, 'a) Stdlib.format4 -> 'b

Expect the option to contain a value and fails (Failure) otherwise. The format string specify the content of the failure

val value_fun : 'a option -> default:(unit -> 'a) -> 'a

Like Stdlib.Option.value but the default is a called function, that can throw instead of giving a value

val of_bool : some:'a -> bool -> 'a option

Create an option from a bool, with the some value

val of_bool_fun : some:(unit -> 'a) -> bool -> 'a option

Create an option from a bool, with the some value as computed by the some function

val for_all : ('a -> bool) -> 'a option -> bool

for_all p o = fold ~none:true ~some:p

val exists : ('a -> bool) -> 'a option -> bool

exists p o = fold ~none:false ~some:p

val guard : bool -> 'a -> 'a option

Return the second argument if the first is true, otherwise None

val guardn : bool -> 'a -> 'a option

Return the second argument if the first is false, otherwise None

Monadic bindings

val (let+) : 'a option -> ('a -> 'b) -> 'b option

Applicative let.

let+ x = mx in e is Option.map (fun x -> e) mx

val (and+) : 'a option -> 'b option -> ('a * 'b) option

Applicative and.

let+ x = mx and+ y = my in e give Some e if both mx and my were Somes.

val (let+!) : 'a option -> ('a -> unit) -> unit

Iter applicative let.

let+! x = mx in e runs e if mx contained a value i.e Option.iter (fun x -> e) mx

val (let*) : 'a option -> ('a -> 'b option) -> 'b option

Monadic let: let* x = mx in e is Option.bind mx (fun x -> e)

val (and*) : 'a option -> 'b option -> ('a * 'b) option

Monadic and: let* x = mx and* y = my in e is let* x = mx in let* y = my in e

Lists

val lift : 'a option list -> 'a list option

Commute the list and the option. If the list contains one None then the result is None. If you want to keep all the Some value, use List.filter_map.

This can be condidered as a list-wide take_all. A list-wide take_first would be List.find_map

val map_lift : ('a -> 'b option) -> 'a list -> 'b list option

The same as a List.map and then a lift

Pairs

val lift_pair : ('a option * 'b option) -> ('a * 'b) option

Lift a pair of options to an option of pair. It is the same as take_all.

val unlift_pair : ('a * 'b) option -> 'a option * 'b option

Unlift an option of pair to a pair of options