Module Utils.Pp

This module provide all pretty printing functionality. It's main goal is not directly handle the output but to handle how to layout complex data structure in a text format.

The main idea of this library is to separate the layout description phase from the part where you actually lay out the thing to pretty print. Thus the two stage of pretty printing is first to generate a document object describing the layout from the object to print and then render the document to the string using one of the printing

This is a wrapper around the pprint library

include module type of struct include PPrint end

type document = PPrint.document

val empty : document

val char : char -> document

val string : string -> document

val substring : string -> int -> int -> document

val fancystring : string -> int -> document

val fancysubstring : string -> int -> int -> int -> document

val utf8string : string -> document

val utf8format : ('a, unit, string, document) Stdlib.format4 -> 'a

val hardline : document

val blank : int -> document

val break : int -> document

val (^^) : document -> document -> document

val group : document -> document

val ifflat : document -> document -> document

val nest : int -> document -> document

val align : document -> document

type point = int * int

type range = point * point

val range : (range -> unit) -> document -> document

val is_empty : document -> bool

module type RENDERER = PPrint.RENDERER

module ToChannel = PPrint.ToChannel

module ToBuffer = PPrint.ToBuffer

module ToFormatter = PPrint.ToFormatter

type requirement = int

val infinity : requirement

class type output = object ... end

type state = PPrint.state = {

width : int;

ribbon : int;

mutable last_indent : int;

mutable line : int;

mutable column : int;

}

class type custom = object ... end

val custom : custom -> document

val requirement : document -> requirement

val pretty : output -> state -> int -> bool -> document -> unit

val compact : output -> document -> unit

val lparen : document

val rparen : document

val langle : document

val rangle : document

val lbrace : document

val rbrace : document

val lbracket : document

val rbracket : document

val squote : document

val dquote : document

val bquote : document

val semi : document

val colon : document

val comma : document

val dot : document

val sharp : document

val slash : document

val backslash : document

val equals : document

val qmark : document

val tilde : document

val at : document

val percent : document

val dollar : document

val caret : document

val ampersand : document

val star : document

val plus : document

val minus : document

val underscore : document

val bang : document

val bar : document

val precede : document -> document -> document

val terminate : document -> document -> document

val enclose : document -> document -> document -> document

val squotes : document -> document

val dquotes : document -> document

val bquotes : document -> document

val braces : document -> document

val parens : document -> document

val angles : document -> document

val brackets : document -> document

val twice : document -> document

val repeat : int -> document -> document

val concat : document list -> document

val separate : document -> document list -> document

val concat_map : ('a -> document) -> 'a list -> document

val separate_map : document -> ('a -> document) -> 'a list -> document

val separate2 : document -> document -> document list -> document

val optional : ('a -> document) -> 'a option -> document

val lines : string -> document list

val arbitrary_string : string -> document

val words : string -> document list

val split : (char -> bool) -> string -> document list

val flow : document -> document list -> document

val flow_map : document -> ('a -> document) -> 'a list -> document

val url : string -> document

val hang : int -> document -> document

val prefix : int -> int -> document -> document -> document

val jump : int -> int -> document -> document

val infix : int -> int -> document -> document -> document -> document

val surround : int -> int -> document -> document -> document -> document

val soft_surround : int -> int -> document -> document -> document -> document

val surround_separate : int -> int -> document -> document -> document -> document -> document list -> document

val surround_separate_map : int -> int -> document -> document -> document -> document -> ('a -> document) -> 'a list -> document

val (!^) : string -> document

val (^/^) : document -> document -> document

val (^//^) : document -> document -> document

module OCaml = PPrint.OCaml

Rendering documents

This section provide function to render documents in a text output.

However unless you know what you are doing, you may prefer using the Logs module with the function of "Render documents in format strings", instead of calling those function directly.

val fprint : Stdlib.out_channel -> document -> unit

Render the document in a pretty manner on the provided out_channel

val fprintln : Stdlib.out_channel -> document -> unit

Render the document in a compact manner on the provided out_channel

val print : document -> unit

Print the document on stdout. The general way of

val println : document -> unit

Print the document on stdout, then a endline, then flush

val eprintln : document -> unit

Print the document on stderr, then a endline, then flush

val sprintc : ToBuffer.document -> string

Print the document in a string, in a compact manner

val sprint : ToBuffer.document -> string

Print the document in a string, in a pretty manner

Render documents in format strings

This sub section provide functions to embed a pprint in a usual printf-like format string.

The general usage is to use the %t format and then appropriate mapping function. For example:

Printf.printf "Documents: %t" Pp.(top printer object)

This is lazy which means that if you use this in disabled debug messages, the object will not even be read. The only cost will be to allocate the closure on the minor GC.

Wheter to use top or tos depends if the final output is an output stream or a plain string.

val top : ('a -> document) -> 'a -> Stdlib.out_channel -> unit

Convert a document printer to a Printf format string via %t for function outputing on an out_channel

val topi : ('a -> document) -> 'a -> Stdlib.out_channel -> unit

Same as top but add 4 space of indentation everywhere on the object. The general use is:

Printf.printf "Object:\n%t\n" Pp.(topi printer object)

and it will print:

Object:
object line 1
object line 2

val tos : ('a -> ToBuffer.document) -> 'a -> unit -> string

Convert a document printer to a Printf format string via %t for function outputing in a string

Combinators

This section provide function to create documents from various object as well as combinators to generate documents from more complex objects.

val dprintf : ('a, unit, string, document) Stdlib.format4 -> 'a

Printf like function that returns a document of the formatted string. In the end, it's just a string document, nothing more complex

val space : document

A breakable space

val nbspace : document

A non-breakable space (equivalent to char ' ')

val bool : bool -> document

Print a boolean as "true" or "false"

val int : int -> document

Print an int in decimal

val hex : int -> document

Print an int in hexadecimal as if it was unsigned

val shex : int -> document

Print an int in hexadecimal with a sign: -1 will be printed to -0x1

val ptr : int -> document

Print an int in hexadecimal with the 0x prefix

val byte : char -> document

Print a byte as 2 hexadecimal digit

val hex16 : int -> document

Print an unsigned 16 bit integer as 4 hexadecimal digit

val hex32 : int32 -> document

Print an unsigned 32 bit integer as 8 hexadecimal digit

val hex64 : int64 -> document

Print an unsigned 64 bit integer as 16 hexadecimal digit

val array : ('a -> PPrint.document) -> 'a array -> PPrint.document

Print an array when provided with an element printer

val list : ('a -> PPrint.document) -> 'a list -> PPrint.document

Print an list when provided with an element printer

val opt : ('a -> PPrint.document) -> 'a option -> PPrint.document

Print an option when provided with an element printer

val pair : ('a -> PPrint.document) -> ('b -> PPrint.document) -> ('a * 'b) -> PPrint.document

Print an pair when provided with both element printers

val tup3 : ('a -> PPrint.document) -> ('b -> PPrint.document) -> ('c -> PPrint.document) -> ('a * 'b * 'c) -> PPrint.document

Print an 3-sized tuple when provided with all three element printers

val qstring : string -> document

Print a quoted string

val erase : 'a -> document

Ignore the input and print nothing

val mapping : string -> (document * document) list -> document

Prints a mapping with this style:

name{
    key -> value;
    key -> value;
    key -> value;
}

val hashtbl : ?name:string -> ('a -> document) -> ('b -> document) -> ('a, 'b) Stdlib.Hashtbl.t -> document

Print a Hashtbl using mapping

val hashtbl_sorted : compare:('a -> 'a -> int) -> ?name:string -> ('a -> document) -> ('b -> document) -> ('a, 'b) Stdlib.Hashtbl.t -> document

Print a sorted Hashtbl using mapping

val record : OCaml.type_name -> (OCaml.record_field * PPrint.document) list -> document

Print a record in the format

name{
    field = value;
    field = value;
    field = value;
}

val separate_mapi : document -> (int -> 'a -> document) -> 'a list -> document

Like separate_map but with the index

val concat_array_map : ('a -> document) -> 'a array -> document

Concatenate the documents produced by the function over the array

val concat_array_mapi : (int -> 'a -> document) -> 'a array -> document

Concatenate the document produced by the function on the array. The function also gets the index of the element

Special printer

val status : (int -> document) -> Unix.process_status -> document

Prints a Unix.process_status with an integer printer

val statusi : Unix.process_status -> document

Prints a Unix.process_status with decimal integers

val statush : Unix.process_status -> document

Prints a Unix.process_status with hexadecimal integers