Module Utils.Raise

This module provide convenience facilities to raise exception or other exception management

val inv_arg : ('a, unit, string, 'b) Stdlib.format4 -> 'a

Printf like funtion that throws an Invalid_Argument with the formated string

val fail : ('a, unit, string, 'b) Stdlib.format4 -> 'a

Printf like funtion that throws a Failure with the formated string

val again : exn -> 'a

Raise again an exception without losing the backtrace.

exception Todo

val todo : unit -> 'a

Put that in unfinished places of the code that need to be completed

exception Unreachable

val unreachable : unit -> 'a

Put that in-place of dead code that is required by the typer