Module Utils.RngMap

A module giving a map indexing data with address ranges and providing access to quick access to data corresponding to any value in-between.

Each address is bound to at most one object so the ranges are not allowed to overlap.

In practice you dont specify the range. The size of the bound range is provided by the stored object itself that is assumed to have a length. The stored object must thus provide a len function as specfified by the LenObject signature. However a length can be added to any Object using PairLenObject

An important remark is that maps are perfectly allowed to have negative addresses, and will considered negative integer as negative addresses. An object of size 7 starting at -4 will end at 3

For now, this has a pure immutable interface.

module IMap : sig ... end

An integer map: Map.Make(Int)

module type Object = sig ... end

A module that represent a simple type with no operation

module type LenObject = sig ... end

A module for type that have a concept of length in the rngMap context

module PairLenObject (Obj : Object) : LenObject with type t = Obj.t * int

For types that do not have an inner length representation, we add it with a pair

module type S = sig ... end

The signature of the range map

module Make (Obj : LenObject) : S with type obj = Obj.t

How to make a RngMap from a LenObject