Module Utils.WeakPtr

This module provide a simple weak pointer that can be rendered invalid by the GC

type 'a t

The type of a weak pointer

val empty : unit -> 'a t

Make a new empty pointer

val make : 'a -> 'a t

Make a new pointer and set it with the given value

val seto : 'a t -> 'a option -> unit

Set or clear the value inside the pointer depending of the option

val set : 'a t -> 'a -> unit

Set the value inside the pointer

val reset : 'a t -> unit

Clear the value inside the pointer

val geto : 'a t -> 'a option

Retrieves the value and return None if the pointer is now empty

exception Deleted

Raised when accessing and empty ptr

val get : 'a t -> 'a

Retrieves the value. Will raise Deleted if the value was deleted by the GC

val check : 'a t -> bool

Check if the pointer is filled