Module Utils__Array

This module is for extending the Array module of the standard library

include Stdlib.Array

type 'a t = 'a array

val length : 'a array -> int

val get : 'a array -> int -> 'a

val set : 'a array -> int -> 'a -> unit

val make : int -> 'a -> 'a array

val create : int -> 'a -> 'a array

val create_float : int -> float array

val make_float : int -> float array

val init : int -> (int -> 'a) -> 'a array

val make_matrix : int -> int -> 'a -> 'a array array

val create_matrix : int -> int -> 'a -> 'a array array

val append : 'a array -> 'a array -> 'a array

val concat : 'a array list -> 'a array

val sub : 'a array -> int -> int -> 'a array

val copy : 'a array -> 'a array

val fill : 'a array -> int -> int -> 'a -> unit

val blit : 'a array -> int -> 'a array -> int -> int -> unit

val to_list : 'a array -> 'a list

val of_list : 'a list -> 'a array

val iter : ('a -> unit) -> 'a array -> unit

val iteri : (int -> 'a -> unit) -> 'a array -> unit

val map : ('a -> 'b) -> 'a array -> 'b array

val mapi : (int -> 'a -> 'b) -> 'a array -> 'b array

val fold_left : ('a -> 'b -> 'a) -> 'a -> 'b array -> 'a

val fold_right : ('b -> 'a -> 'a) -> 'b array -> 'a -> 'a

val iter2 : ('a -> 'b -> unit) -> 'a array -> 'b array -> unit

val map2 : ('a -> 'b -> 'c) -> 'a array -> 'b array -> 'c array

val for_all : ('a -> bool) -> 'a array -> bool

val exists : ('a -> bool) -> 'a array -> bool

val mem : 'a -> 'a array -> bool

val memq : 'a -> 'a array -> bool

val sort : ('a -> 'a -> int) -> 'a array -> unit

val stable_sort : ('a -> 'a -> int) -> 'a array -> unit

val fast_sort : ('a -> 'a -> int) -> 'a array -> unit

val to_seq : 'a array -> 'a Stdlib.Seq.t

val to_seqi : 'a array -> (int * 'a) Stdlib.Seq.t

val of_seq : 'a Stdlib.Seq.t -> 'a array

val unsafe_get : 'a array -> int -> 'a

val unsafe_set : 'a array -> int -> 'a -> unit

module Floatarray = Stdlib__array.Floatarray

val map_mut : ('a -> 'a) -> 'a array -> unit

Mutate the array by replacing each value x by f x

val of_list_mapi : (int -> 'a -> 'b) -> 'a list -> 'b array

of_list_mapi f l = of_list (List.mapi f i l) = mapi f (of_list l)

val of_list_map : ('a -> 'b) -> 'a list -> 'b array

of_list_map f l = of_list (List.map f l) = map f (of_list l)

Array scanning

val find_pair : ('a -> bool) -> 'a array -> int * 'a

Find the first value satisfying the predicate and return it with its index. Throw Not_found if no value satisfies the predicate

val find : ('a -> bool) -> 'a array -> 'a

Find the first value satisfying the predicate. Throw Not_found if no value satisfies the predicate

val find_index : ('a -> bool) -> 'a array -> int

Find the first index whose value satisfies the predicate. Throw Not_found if no value satisfies the predicate

val find_all_pairs : ('a -> bool) -> 'a t -> (int * 'a) list

Find all the values satisfying the predicate and return them with their index.

val find_all : ('a -> bool) -> 'a t -> 'a list

Find all the values satisfying the predicate

val find_all_indices : ('a -> bool) -> 'a t -> int list

Find all the indices whose value satisfies the predicate