Module Utils__BytesSeq

This module represent a byte sub view on a bytes object. Contrary to Bytes it is a non-owning immutable view. It do not prevent the original bytes from being modified, and the changes will be propagated in the view. It is additional sugar on top of Linksem's Byte_sequence_wrapper

About all the suffixed function:

All iteration function without suffix do the expected operation on char (as single bytes)

All iteration function with suffix nle do the expected operation on a sequence of integers of n bits as read in little endian.

All iteration function with suffix nbe do the expected operation on a sequence of integers of n bits as read in big endian.

All iteration function with suffix bs do the expected operation on a sequence of BytesSeq.t of specified length.

All iteration function with suffix bvle do the expected operation on a sequence of BitVec the specified size as read in little endian.

type t = Byte_sequence_wrapper.byte_sequence

Type inherited from linksem

val length : t -> int

Get the length of the byteseq in bytes

val equal : t -> t -> bool

Check if two byte sequence are equal byte for byte

val int_bytes : int

The size in bytes of an Ocaml int

Hexadecimal conversions

val to_hex : t -> string

Convert the byte sequence to an hexadecimal string

val to_hex_rev : t -> string

Convert the byte sequence to an reversed hexadecimal string. This will print it like a big-endian integer.

val of_hex : string -> t

Parse the string as hexadecimal like A4B767DF and create a bytes of this a binary data and then a bytesSeq view of it

Cutting the view

val sub : t -> int -> int -> t

sub bs start len Extract a sub range [start:start+len) of a byte sequence. This is O(1)

val front : int -> t -> t

front i bs Take the first i bytes of bs and discard the rest. Equivalent to sub bs 0 i

val back : int -> t -> t

back i bs Take the last i bytes of bs and discard the rest. Equivalent to sub bs i (length bs - i)

Interaction with bytes and raw string

val blit : t -> int -> bytes -> int -> int -> unit

blit src srcoff dst dstoff len copies len bytes from bytes sequence src, starting at index srcoff, to bytes dst, starting at index dstoff.

See Bytes.blit.

val of_bytes : bytes -> t

Create a view of the whole bytes

val of_string : string -> t

Create a view of the whole string as raw bytes

val to_string : t -> string

Create a copy of the view in a string

val bytes_sub : bytes -> int -> int -> t

Create a byte sequence view of a specified range of a bytes. See sub

Array conversions

val to_array : t -> char array

Convert to a char array

val of_array : char array -> t

Convert from a char array

Getters

val get : t -> int -> char

Get the bytes at the offset in the byte sequence

val unsafe_get : t -> int -> char

Unsafe version of get

val get16le : t -> int -> int

Get a 16 bit integer at the offset in the byte sequence as little endian

val get16be : t -> int -> int

Get a 16 bit integer at the offset in the byte sequence as big endian

val get32le : t -> int -> int32

Get a 32 bit integer at the offset in the byte sequence as little endian

val get32be : t -> int -> int32

Get a 32 bit integer at the offset in the byte sequence as big endian

val get64le : t -> int -> int64

Get a 64 bit integer at the offset in the byte sequence as little endian

val get64be : t -> int -> int64

Get a 64 bit integer at the offset in the byte sequence as big endian

val getintle : t -> int -> int

Get an Ocaml int at the offset in the byte sequence as little endian. The number of bytes read is 4 if Sys.int_size is 31 and 8 if Sys.int_size is 63

val getbs : len:int -> t -> int -> t

Get a byte sequence of length len at the offset in another byte sequence

val getbvle : size:int -> t -> int -> Utils.BitVec.t

Get a BitVec of size size at the offset in the byte sequence as little endian

val getintle_ze : t -> int -> int

Get an Ocaml int at the offset in the byte sequence as little endian. The number of bytes read is 4 if Sys.int_size is 31 and 8 if Sys.int_size is 63. If the read goes beyond the end of the sequence, instead of failing, zeros are read.

Iterators

Iterators over a byte sequence. If the length of the byte sequence is not a multiple of the step of the iteration then the trailing odd bytes are not iterated over.

val iter : (char -> unit) -> t -> unit

val iter16le : (int -> unit) -> t -> unit

val iter16be : (int -> unit) -> t -> unit

val iter32le : (int32 -> unit) -> t -> unit

val iter32be : (int32 -> unit) -> t -> unit

val iter64le : (int64 -> unit) -> t -> unit

val iter64be : (int64 -> unit) -> t -> unit

val iterbs : len:int -> (t -> unit) -> t -> unit

Iterate over the byte sequence by bytesequence of length len. If the total byte sequence is not of length a multiple of len then that iterated value will be shorter

val fold_left : ('a -> char -> 'a) -> 'a -> t -> 'a

val fold_left16le : ('a -> int -> 'a) -> 'a -> t -> 'a

val fold_left16be : ('a -> int -> 'a) -> 'a -> t -> 'a

val fold_left32le : ('a -> int32 -> 'a) -> 'a -> t -> 'a

val fold_left32be : ('a -> int32 -> 'a) -> 'a -> t -> 'a

val fold_left64le : ('a -> int64 -> 'a) -> 'a -> t -> 'a

val fold_left64be : ('a -> int64 -> 'a) -> 'a -> t -> 'a

val fold_leftbs : len:int -> ('a -> t -> 'a) -> 'a -> t -> 'a

List conversions

val to_list : t -> char list

val to_list16le : t -> int list

val to_list16be : t -> int list

val to_list32le : t -> int32 list

val to_list32be : t -> int32 list

val to_list64le : t -> int64 list

val to_list64be : t -> int64 list

val to_listbs : len:int -> t -> t list

Cut a byte sequence into a list of byte sequences of length len, (and a shorter last one if the total len is not a multiple of len)

Binary IO

val output : Stdlib.out_channel -> t -> unit

Output the raw data of the byte sequence on the output channel

val input : Stdlib.in_channel -> t

Output the raw date of the byte sequence of the input channel

Pretty Printing

val pp : t -> Utils.Pp.document

Pretty print a byte sequence as space separated bytes like ab cd ef.

Here "ab" is the byte number 0 and "ef" is the byte number 2.

val ppc : t -> Utils.Pp.document

Pretty print a byte sequence as an hexadecimal string like abcdef

Here "ab" is the byte number 0 and "ef" is the byte number 2.

This can also be seen as printing the bytesequence as a single integer encoded in big endian format.

val ppint : t -> Utils.Pp.document

Pretty print a byte sequence as an hexadecimal integer (in little endian). The byte order is reversed compared to ppc

For example the byte sequence ab cd ef will be printed as efcdab where "ab" is the byte number 0 and "ef" is the byte number 2.

val ppby : by:int -> t -> Utils.Pp.document

Pretty print a byte sequence by step of by bytes. Each block is pretty printed as an hex string like ppc and blocks are separated by spaces.

val ppbyint : by:int -> t -> Utils.Pp.document

Pretty print a byte sequence by step of by bytes. Each block is pretty printed as a reversed hex string i.e like an integer of length by. Thus each block will printed like with ppcint Blocks are separated by spaces.

For example to print a byte sequence as a space separated list of little-endian integers do:

ppbyint ~by:4 bs