Module Test.Cache

type key = Key.t

type value = Value.t

type epoch = UnitEpoch.t

type t = Make(Key)(Value)(UnitEpoch).t

The type that represent the cache in RAM.

val make : ?⁠fake:bool -> string -> epoch -> t

Build a new cache management object with a name and an epoch If fake is set, the cache will not touch the disk and behave as a plain Hashtbl.

val get_opt : t -> key -> value option

Get a value from the cache or None if no value is bound to the key

val get : t -> key -> value

Get a value from the cache or throws Not_found if no value is bound

val add : t -> key -> value -> unit

Add a new binding. Fails if a binding already exists

val remove : t -> key -> unit

Remove a binding (Also delete the representation on disk