Module Utils__Cache.Make

The Key must provide the Key interface, where the values must satisfy the Value interface.

The map is stored on the disk in a simple way. First the key is hashed and this hashed gives a filename with int_to_file. If there is no collision, then the value is stored in that file and the key is stored (optionally) in hash.key file (See Key for details).

If there is collision then the filename is a directory containing numbered files of all the key-value pairs. Again the value is in file named n when the key is (optionally) in file named n.key.

Parameters

Key : Key

Value : Value

Epoch : Epoch

Signature

type key = Key.t

The type of keys

type value = Value.t

The type of values

type epoch = Epoch.t

The type of the epoch

type t

The type that represent the cache in RAM.

val make : ?⁠fake:bool -> string -> epoch -> t

Build a new cache management object with a name and an epoch If fake is set, the cache will not touch the disk and behave as a plain Hashtbl.

val get_opt : t -> key -> value option

Get a value from the cache or None if no value is bound to the key

val get : t -> key -> value

Get a value from the cache or throws Not_found if no value is bound

val add : t -> key -> value -> unit

Add a new binding. Fails if a binding already exists

val remove : t -> key -> unit

Remove a binding (Also delete the representation on disk