Module Utils__Cmd.SocketServer

This module provide functionality for a socket server with which one can communicate on a sockets.

An example of use is in Isla.Server

type t

The abstract type of a socket server

val start : name:string -> (string -> cmd) -> t

Start the server with provided name and wait for it to connect to the socket. Then build the Server.t object. The function argument must take a socket name and give a valid command line to call the server process and make it connect to the socket.

val stop : t -> unit

Stop the server and cut the connection, wait for the subprocess to die and then delete the socket

May throw Crash on error.

val read_byte : t -> int

Read a single byte from the server

val read_string : t -> string

Read a string with the following format:

| header : 4 bytes | data : header bytes |

In other words, read a 4 bytes number, then read that number of bytes into a string

val write_string : t -> string -> unit

Write a string in the same binary format as read_string