Module Utils__Fun

More functional combinator. This module extends the base OCaml API of Fun.

include Stdlib.Fun

val id : 'a -> 'a

val const : 'a -> 'b -> 'a

val flip : ('a -> 'b -> 'c) -> 'b -> 'a -> 'c

val negate : ('a -> bool) -> 'a -> bool

val protect : finally:(unit -> unit) -> (unit -> 'a) -> 'a

exception Finally_raised of exn

val ($) : ('a -> 'b) -> 'a -> 'b

An other application operator. g @@ f @@ 4 replaces g (f 4) but h $ g 4 $ f 5 replaces h (g 4) (f 5)

@@ has higher precedence so h $ g @@ f 4 $ t @@ p 5 = h (g (f 4)) (t (p 5)) but in practice, I would advise not to mix them.

val (%>) : ('a -> 'b) -> ('b -> 'c) -> 'a -> 'c

Abstraction of piping above a parameter useful to refactor things like List.map f |> List.map g into List.map (f %> g).

The trivial definition is x |> (f %> g) = x |> f |> g

val tee : ('a -> unit) -> 'a -> 'a

When you want to run some imperative code on a value before continuing the pipeline

val curry : (('a * 'b) -> 'c) -> 'a -> 'b -> 'c

curry f a b = f (a, b)

val uncurry : ('a -> 'b -> 'c) -> ('a * 'b) -> 'c

uncurry f (a, b) = f a b

val ctrue : 'a -> bool

Shorthand for const true

val cfalse : 'a -> bool

Shorthand for const false