Module Utils__Pair

This module contain random utility functions dealing with pairs

val map : ('a -> 'c) -> ('b -> 'd) -> ('a * 'b) -> 'c * 'd

Map each function on one side of the pair

val iter : ('a -> unit) -> ('b -> unit) -> ('a * 'b) -> unit

Iter each function on one side of the pair

val swap : ('a * 'b) -> 'b * 'a

Swap the element of a pair

val compare : ?⁠fst:('a -> 'a -> int) -> ?⁠snd:('b -> 'b -> int) -> ('a * 'b) -> ('a * 'b) -> int

Compare a pair using provided comparison function. Both fst and snd default to the polymorphic compare

val equal : ?⁠fst:('a -> 'a -> bool) -> ?⁠snd:('b -> 'b -> bool) -> ('a * 'b) -> ('a * 'b) -> bool

Test pair equality using provided equality function. Both fst and snd default to the polymorphic equality

val make : 'a -> 'b -> 'a * 'b

Just build the pair, not useful on it's own but List.combine is just List.map2 Pair.make, and there are some other cases where it is handy for high-order programming

val split : 'a -> 'a * 'a

Just build a pair with twice the element. Useful in high order programming

val for_all : ('a -> bool) -> ('b -> bool) -> ('a * 'b) -> bool

Check that both individual predicates hold

val exists : ('a -> bool) -> ('b -> bool) -> ('a * 'b) -> bool

Check that at least one of the individual predicates holds