Module Utils__Protect

This module provide try-with-finally kind of exception handling.

It provides a new kind of protect function.

TODO: Think if this behavior is really useful compared to the standard protect

exception Protect_both of exn * exn

This exception it thrown when both function and protector have thrown

val protect : (unit -> 'a) -> (unit -> unit) -> 'a

protect f p runs f then p even if f throws.

If one of f or p throw, then that exception is transmitted as is If both throw, the pair of exceptions is encapsulated in Protect_both and thrown.

This behavior is slightly different from the standard protect.