Module type Utils__RngMap.LenObject

A module for type that have a concept of length in the rngMap context

type t

The type to be indexed by starting addresses, must have a length

val len : t -> int

The type of range end