Module Utils__Seq

This module is for extending the Seq module of the standard library

include Stdlib.Seq

type 'a t = unit -> 'a node

and 'a node = 'a Stdlib__seq.node =

| Nil

| Cons of 'a * 'a t

val empty : 'a t

val return : 'a -> 'a t

val map : ('a -> 'b) -> 'a t -> 'b t

val filter : ('a -> bool) -> 'a t -> 'a t

val filter_map : ('a -> 'b option) -> 'a t -> 'b t

val flat_map : ('a -> 'b t) -> 'a t -> 'b t

val fold_left : ('a -> 'b -> 'a) -> 'a -> 'b t -> 'a

val iter : ('a -> unit) -> 'a t -> unit

val add_count : ?⁠start:int -> 'a t -> (int * 'a) t

Add a counter starting at start (default 0) in front of each element of the sequence

val iota : ?⁠start:int -> int -> int t

Generate an integer sequence up to len. Optionally may start at start instead of 0

val iota_step_up : ?⁠start:int -> step:int -> endi:int -> int t

Generate an integer sequence up to endi by stepping step. Optionally may start at start instead of 0

val stop_at : ('a -> bool) -> 'a t -> 'a t

Make the sequence stop when the condition is met

val cons : 'a -> 'a t -> unit -> 'a node

Add a new element in front of the sequence. That element will appear first before the rest of the sequence.

Added to Stdlib in Ocaml 4.11

val find_map : ('a -> 'b option) -> 'a t -> 'b option

Applies the specified function to the elements of the sequence in order, and returns the first result of the form Some v, or None if no such result was returned.

See List.find_map