Module Z3

This module handles a Z3 server

For high level usage, call start or ensure_started then instantiate the Make functor and use operation in the section about Context less operation:

For a more medium level usage, you may want to manage your context manually before making requests. The best way of doing that is by using a ContextCounter.

Once you are in the correct context, you can use your instantiated version of Make to do operations such as: S.declare_var or S.send_assert to build your context, and then S.simplify, S.check of S.check_sat.

For low-level details (All function in the first two section of this module should probably be reserved to those who understand the implementation)

The module keeps Z3 as a child process and communicates through pipes using Utils.Cmd.IOServer.

start sends the introduction in intro.smt2 to Z3 so that it is available in any context. SMT answer are parsed from the pipe with Utils.Files.input_sexp. If the wrong number of answer is expected, the system will just deadlock.

Raw server management

val z3_trace : bool

A boolean enabling context tracing (For better error messages)

type context_elem = {

name : string;

mutable num : int;

}

An element of the context stack. It has a name and a declaration number.

The declaration number is incremented at each declaration in the context.

type context = context_elem list

The type of a SMT context stack

val start_context : context

The starting context

val context_elem_to_string : context_elem -> string

Give a string representation of a context_elem

type server = {

ioserver : Utils.Cmd.IOServer.t;

config : Config.File.Z3.t;

mutable context : context;

}

The type of a Z3 server

val server : server option Stdlib.ref

The global Z3 server

val get_server : unit -> server

Assume the server is started and returns it.

val raw_start : unit -> unit

Start Z3 without any checks and without sending intro or pushing

val raw_stop : unit -> unit

Stop Z3 without asking politely

val get_context_string : server -> string

Give a string representation of the current context for error reporting

val incr_context : server -> unit

Increment the declaration number of the top context_elem.

Request and answer management

This section handle raw and less raw direct interaction with the Z3 server.

The send_* function are for sending information to the server with a server handle. This avoided checking if the server is open at each declaration.

The read_* functions are the same the other way around.

The request and command are high level versions

val send_string : server -> string -> unit

Send a string to the server and increment the declaration number in the context

val send_smt : server -> ?ppv:('a -> Utils.Pp.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt -> unit

Send a smt statement to the server and increment the declaration number in the context

val read_string : server -> string

Read a string from the server (A Z3 answer is always a valid sexp)

val read_smt_ans : server -> Ast.rsmt_ans

Read a smt_ans from the server

val request : ?ppv:('a -> Utils.Pp.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt -> Ast.rsmt_ans

Make a request to the server and expect an answer. Will hang if there is no answer

val command : ?ppv:('a -> Utils.Pp.document) -> ('b, 'a, string, AstGen.Def.Size.t) AstGen.Ott.smt -> unit

Send a command or a declaration to the server

val expect_version : server -> Ast.rsmt_ans -> string

Expect a version answer and fails if it is not the case

val expect_exp : server -> Ast.rsmt_ans -> Ast.rexp

Expect an expression answer and fails if it is not the case

val get_version : unit -> string

Get the Z3 version string

val is_context_sat : server -> bool option

Check if the current context is sat

Full startup and shutdown

val stop : unit -> unit

Stop the Z3 server properly

val ensure_stopped : unit -> unit

Call stop if the server wasn't stopped

val start : unit -> unit

Start the Z3 Server and setup the intro

val ensure_started : unit -> unit

Call start if the server wasn't started

val ensure_started_get : unit -> server

Call start if the server wasn't started, then return the server

val reset : unit -> unit

Reset the Z3 server, forgetting everything. Useful for resetting in a test failure context, but probably shouldn't be used in normal operations

Context management

Z3 contexts are managed in a stack way.

There are implemented using (push) and (pop) on the Z3 side.

open_context can open a named context. However if it is expected to open a given context name multiple times, using ContextCounter is advised.

val open_context : server -> string -> unit

Open a new context with a name.

val close_context : server -> unit

Closes current context of the server

module ContextCounter (S : Utils.Logs.String) : sig ... end

Module for handling a context numbering scheme automatically.

High level interaction

This section provide functor that can be instantiated to get easy to use SMT functionality

module type Var = sig ... end

The functors in this section require a bit more support from variable than plain Exp.Var

module type S = sig ... end

module SimpContext : sig ... end

module CheckContext : sig ... end

module Make (Var : Var) : S with type var = Var.t

module Test : sig ... end