Module type Z3.S

type var

Variable declarations

The goal of those operation is to declare variables with their types to the SMT solver. The Htbl allow to declare every variable only once.

type exp = (var, Ast.no) Exp.Typed.t

The type of expression on which SMT operation are made

module Htbl : Stdlib.Hashtbl.S with type key = var

Hash tables over variables

val declare_var_always : server -> var -> unit

Declare the variable regardless of whether it has already been declared

val declare_var : server -> declared:unit Htbl.t -> var -> unit

Declare the variable if it's not in declared. In that case, it also add it to declared

val declare_vars : server -> declared:unit Htbl.t -> exp -> unit

Declare all not yet declared the variable in the expression, then add them to declared

Simplification

val simplify : server -> exp -> exp

Simplify the expression in the current context. All the variable must already have been declared.

val simplify_decl : server -> declared:unit Htbl.t -> exp -> exp

Literally just declare_vars then simplify

SMT checking

val send_assert : server -> exp -> unit

Assert that expression in the current context

val send_assert_decl : server -> declared:unit Htbl.t -> exp -> unit

Literally just declare_vars then send_assert

val check : server -> exp -> bool option

Check if this expression is always true in the current context. None is returned when the SMT solver didn't know

val check_sat : server -> exp -> bool option

Check if this expression is true on at least one assignment of the variables.ioserver None is returned when the SMT solver didn't know

val check_both : server -> exp -> bool option

Check if this expression is always true of always false. None is returned if neither can be proven.

This results in two calls to the SMT solver. one with check and one with check_sat

val simplify_subterms : server -> exp -> exp

val simplify_subterms_decl : server -> declared:unit Htbl.t -> exp -> exp

Context less operation

Those operations do not require a context to operate. They require not setup and tear-down. They are standalone and fully automated.

On the other hand doing multiple one of those in sequence if much less efficient than using properly the functions of previous section.

val simplify_full : exp -> exp

Do a standalone simplification in it's own context. Do not need anything to be already declared or any context to be opened.

val check_full : ?hyps:exp list -> exp -> bool option

Do a standalone check of whether the expression is implied by the list of hypothesis.

val check_sat_full : exp list -> bool option

Do a standalone check of whether the set of assertion is sat

val simplify_subterms_full : ?hyps:exp list -> exp -> exp