Module type Z3.Var

The functors in this section require a bit more support from variable than plain Exp.Var

include Exp.Var

type t

The type of variables

val equal : t -> t -> bool

Equality predicate that will be passed to expressions

val pp : t -> Utils.Pp.document

Pretty printer to be used, both for memory pretty printing and for sending memory to Z3

val ty : t -> Exp.ty

Get the type of the variable

val hash : t -> int

Hashing function to store variable in Hashtbls. Must be compatible with equal

val of_string : string -> t

Parser from a string. Must be the inverse of pp