Module Dw.Var

This module contain all the definition to handle local and global variables as defined in the DWARF information of the target file

type t = {

name : string;

param : bool;

ctype : Ctype.t;

locs : ((int * int) * Loc.t) list;

}

Type of a DWARF variable

type linksem_t = Dwarf.sdt_variable_or_formal_parameter

Type of a DWARF variable in linksem

val loc_merge : (('a * 'a) * Loc.t) list -> (('a * 'a) * Loc.t) list

Merge contiguous location lists

val clamp_z : Z.t -> int

Convert from Z.t to int, if there is an overflow, returns Int.max_int instead of throwing

val of_linksem : Elf.File.t -> Ctype.env -> linksem_t -> t

Create a DWARF variable from its linksem counterpart

val pp_raw : t -> Utils.Pp.document

Pretty print a variable