Module Exp

This module intends to provider a convenience expression module by lifting function like equality or pretty printing from variable to expressions.

It is restricted to typed expression as defined in Typed

For now it also do not support memory expression, but it will soon.

module ConcreteEval : sig ... end

This module provides a way of making concrete evaluation of an expression. The only required thing is a context.

module PpExp : sig ... end

This module provides a human readable pretty printing for Ast expressions

module Sums : sig ... end

This module provide sum manipulation functionality on top of typed expression Typed.t

module Typed : sig ... end

This module provide operation on typed expressions i.e expressions whose annotations are their SMT type (Ast.ty).

module Value : sig ... end

This module provide a type to represent concrete values.

type ty = Ast.no Ast.ty

The type of memory-less expressions

module type Var = sig ... end

This is the type signature for variable required by this module

module type S = sig ... end

The signature of the output module of Make

module Make : functor (Var : Var) -> S with type var = Var.t