Module type Exp.S

The signature of the output module of Make

type var

The type of variable provided in the functor

type t = (var, Ast.no) Typed.t

The type of expression on which this module works

val equal : t -> t -> bool

Test syntactic equality. a + b and b + a would test different under this predicate

val pp : t -> Utils.Pp.document

Pretty print the expression using PpExp

val pp_smt : t -> Utils.Pp.document

Pretty print the expression in SMTLIB language

val of_var : var -> t

Create an expression from a variable

val add_type : ('a, var, Ast.no, Ast.no) Ast.exp -> t

Convert a similar but untyped expression to an expression of type t