Module Isla

include Base

Aliases

include Isla_lang.AST

type enum = int * int

type lrng = Isla_lang__Isla_lang_ast.lrng =

| UnknownRng

| Generated of lrng

| Range of Stdlib.Lexing.position * Stdlib.Lexing.position

exception Parse_error_locn of lrng * string

val pp_lpos : Stdlib.Lexing.position -> PPrint.document

val pp_lrng : lrng -> PPrint.document

type bvmanyarith = Isla_lang__Isla_lang_ast.bvmanyarith =

| Bvand

| Bvor

| Bvxor

| Bvadd

| Bvmul

type bvcomp = Isla_lang__Isla_lang_ast.bvcomp =

| Bvult

| Bvslt

| Bvule

| Bvsle

| Bvuge

| Bvsge

| Bvugt

| Bvsgt

type bvarith = Isla_lang__Isla_lang_ast.bvarith =

| Bvnand

| Bvnor

| Bvxnor

| Bvsub

| Bvudiv

| Bvudivi

| Bvsdiv

| Bvsdivi

| Bvurem

| Bvsrem

| Bvsmod

| Bvshl

| Bvlshr

| Bvashr

type manyop = Isla_lang__Isla_lang_ast.manyop =

| And

| Or

| Bvmanyarith of bvmanyarith

| Concat

type unop = Isla_lang__Isla_lang_ast.unop =

| Not

| Bvnot

| Bvredand

| Bvredor

| Bvneg

| Extract of int * int

| ZeroExtend of int

| SignExtend of int

type binop = Isla_lang__Isla_lang_ast.binop =

| Eq

| Bvarith of bvarith

| Bvcomp of bvcomp

type accessor = Isla_lang__Isla_lang_ast.accessor =

| Field of string

type ty = Isla_lang__Isla_lang_ast.ty =

| Ty_Bool

| Ty_BitVec of int

| Ty_Enum of int

| Ty_Array of ty * ty

type 'a exp = 'a Isla_lang__Isla_lang_ast.exp =

| Var of int * 'a

| Bits of string * 'a

| Bool of bool * 'a

| Enum of enum * 'a

| Unop of unop * 'a exp * 'a

| Binop of binop * 'a exp * 'a exp * 'a

| Manyop of manyop * 'a exp list * 'a

| Ite of 'a exp * 'a exp * 'a exp * 'a

type valu = Isla_lang__Isla_lang_ast.valu =

| Val_Symbolic of int

| Val_Bool of bool

| Val_I of int * int

| Val_Bits of string

| Val_Enum of enum

| Val_String of string

| Val_Unit

| Val_NamedUnit of string

| Val_Vector of valu list

| Val_List of valu list

| Val_Struct of (string * valu) list

| Val_Poison

type accessor_list = Isla_lang__Isla_lang_ast.accessor_list =

| Nil

| Cons of accessor list

type 'a smt = 'a Isla_lang__Isla_lang_ast.smt =

| DeclareConst of int * ty

| DefineConst of int * 'a exp

| Assert of 'a exp

| DefineEnum of int

type valu_option = valu option

type valu_concrete = Isla_lang__Isla_lang_ast.valu_concrete =

| CVal_Bool of bool

| CVal_I of int * int

| CVal_Bits of string

| CVal_Enum of enum

| CVal_String of string

| CVal_Unit

| CVal_NamedUnit of string

| CVal_Vector of valu list

| CVal_List of valu list

| CVal_Struct of (string * valu) list

| CVal_Poison

type 'a event = 'a Isla_lang__Isla_lang_ast.event =

| Smt of 'a smt * 'a

| Branch of int * string * 'a

| ReadReg of string * accessor_list * valu * 'a

| WriteReg of string * accessor_list * valu * 'a

| ReadMem of valu * valu * valu * int * valu_option * 'a

| WriteMem of int * valu * valu * valu * int * valu_option * 'a

| BranchAddress of valu * 'a

| Barrier of valu * 'a

| CacheOp of valu * valu * 'a

| MarkReg of string * string * 'a

| Cycle of 'a

| Instr of valu * 'a

| Sleeping of int * 'a

| WakeRequest of 'a

| SleepRequest of 'a

type 'a trc = 'a Isla_lang__Isla_lang_ast.trc =

| Trace of 'a event list

type 'a exp_val = 'a Isla_lang__Isla_lang_ast.exp_val =

| EV_Bits of string * 'a

| EV_Bool of bool * 'a

| EV_Enum of enum * 'a

| EV_Unop of unop * 'a exp_val * 'a

| EV_Binop of binop * 'a exp_val * 'a exp_val * 'a

| EV_Manyop of manyop * 'a exp_val list * 'a

| EV_Ite of 'a exp_val * 'a exp_val * 'a exp_val * 'a

module Lexer = Isla_lang.Lexer

module Parser = Isla_lang.Parser

type loc = Stdlib.Lexing.position

type rtrc = lrng trc

The type of raw traces out of the parser

type revent = lrng event

The type of raw events out of the parser

type rsmt = lrng smt

The type of raw SMT declaration out of the parser

type rexp = lrng exp

The type of raw expressions out of the parser

IslaTrace parsing

exception ParseError of loc * string

Exception that represent an Isla parsing error

exception LexError of loc * string

Exception that represent an Isla lexing error

type lexer = Stdlib.Lexing.lexbuf -> Parser.token

type 'a parser = lexer -> Stdlib.Lexing.lexbuf -> 'a

val parse : 'a parser -> ?⁠filename:string -> Stdlib.Lexing.lexbuf -> 'a

Parse a single Isla instruction output from a Lexing.lexbuf

val parse_exp : ?⁠filename:string -> Stdlib.Lexing.lexbuf -> rexp

Parse a single Isla expression from a Lexing.lexbuf

val parse_exp_string : ?⁠filename:string -> string -> rexp

Parse a single Isla expression from a string

val parse_exp_channel : ?⁠filename:string -> Stdlib.in_channel -> rexp

Parse a single Isla expression from a channel

val parse_trc : ?⁠filename:string -> Stdlib.Lexing.lexbuf -> rtrc

Parse an Isla trace from a Lexing.lexbuf

val parse_trc_string : ?⁠filename:string -> string -> rtrc

Parse an Isla trace from a string

val parse_trc_channel : ?⁠filename:string -> Stdlib.in_channel -> rtrc

Parse an Isla trace from a channel

IslaTrace pretty printing

include Isla_lang.PP

val pp_raw_vvar : int -> PPrintEngine.document

val pp_raw_name : string -> PPrintEngine.document

val pp_raw_enum_ty : int -> PPrintEngine.document

val pp_raw_enum : Isla_lang__.Isla_lang_ast.enum -> PPrintEngine.document

val pp_raw_int : int -> PPrintEngine.document

val pp_raw_bvi : int -> PPrintEngine.document

val pp_raw_bv : string -> PPrintEngine.document

val pp_raw_str : string -> PPrintEngine.document

val pp_raw_ty : Isla_lang__.Isla_lang_ast.ty -> PPrintEngine.document

val pp_raw_bool : bool -> PPrintEngine.document

val pp_raw_unop : Isla_lang__.Isla_lang_ast.unop -> PPrintEngine.document

val pp_raw_bvarith : Isla_lang__.Isla_lang_ast.bvarith -> PPrintEngine.document

val pp_raw_bvcomp : Isla_lang__.Isla_lang_ast.bvcomp -> PPrintEngine.document

val pp_raw_binop : Isla_lang__.Isla_lang_ast.binop -> PPrintEngine.document

val pp_raw_bvmanyarith : Isla_lang__.Isla_lang_ast.bvmanyarith -> PPrintEngine.document

val pp_raw_manyop : Isla_lang__.Isla_lang_ast.manyop -> PPrintEngine.document

val pp_raw_exp : 'a Isla_lang__.Isla_lang_ast.exp -> PPrintEngine.document

val pp_raw_exp_val : 'a Isla_lang__.Isla_lang_ast.exp_val -> PPrintEngine.document

val pp_raw_smt : 'a Isla_lang__.Isla_lang_ast.smt -> PPrintEngine.document

val pp_raw_valu : Isla_lang__.Isla_lang_ast.valu -> PPrintEngine.document

val pp_raw_selem : (string * Isla_lang__.Isla_lang_ast.valu) -> PPrintEngine.document

val pp_raw_valu_concrete : Isla_lang__.Isla_lang_ast.valu_concrete -> PPrintEngine.document

val pp_raw_selem_concrete : (string * Isla_lang__.Isla_lang_ast.valu_concrete) -> PPrintEngine.document

val pp_raw_accessor : Isla_lang__.Isla_lang_ast.accessor -> PPrintEngine.document

val pp_raw_accessor_list : Isla_lang__.Isla_lang_ast.accessor_list -> PPrintEngine.document

val pp_raw_valu_option : Isla_lang__.Isla_lang_ast.valu_option -> PPrintEngine.document

val pp_raw_event : 'a Isla_lang__.Isla_lang_ast.event -> PPrintEngine.document

val pp_raw_trc : 'a Isla_lang__.Isla_lang_ast.trc -> PPrintEngine.document

val pp_vvar : int -> PPrintEngine.document

val pp_name : string -> PPrintEngine.document

val pp_enum_ty : int -> PPrintEngine.document

val pp_enum : Isla_lang__.Isla_lang_ast.enum -> PPrintEngine.document

val pp_int : int -> PPrintEngine.document

val pp_bvi : int -> PPrintEngine.document

val pp_bv : string -> PPrintEngine.document

val pp_str : string -> PPrintEngine.document

val pp_j : int -> string

val pp_ty : Isla_lang__.Isla_lang_ast.ty -> PPrintEngine.document

val pp_bool : bool -> PPrintEngine.document

val pp_unop : Isla_lang__.Isla_lang_ast.unop -> PPrintEngine.document

val pp_bvarith : Isla_lang__.Isla_lang_ast.bvarith -> PPrintEngine.document

val pp_bvcomp : Isla_lang__.Isla_lang_ast.bvcomp -> PPrintEngine.document

val pp_binop : Isla_lang__.Isla_lang_ast.binop -> PPrintEngine.document

val pp_bvmanyarith : Isla_lang__.Isla_lang_ast.bvmanyarith -> PPrintEngine.document

val pp_manyop : Isla_lang__.Isla_lang_ast.manyop -> PPrintEngine.document

val pp_exp : 'a Isla_lang__.Isla_lang_ast.exp -> PPrintEngine.document

val pp_exp_val : 'a Isla_lang__.Isla_lang_ast.exp_val -> PPrintEngine.document

val pp_smt : 'a Isla_lang__.Isla_lang_ast.smt -> PPrintEngine.document

val pp_valu : Isla_lang__.Isla_lang_ast.valu -> PPrintEngine.document

val pp_selem : (string * Isla_lang__.Isla_lang_ast.valu) -> PPrintEngine.document

val pp_valu_concrete : Isla_lang__.Isla_lang_ast.valu_concrete -> PPrintEngine.document

val pp_selem_concrete : (string * Isla_lang__.Isla_lang_ast.valu_concrete) -> PPrintEngine.document

val pp_accessor : Isla_lang__.Isla_lang_ast.accessor -> PPrintEngine.document

val pp_accessor_list : Isla_lang__.Isla_lang_ast.accessor_list -> PPrintEngine.document

val pp_valu_option : Isla_lang__.Isla_lang_ast.valu_option -> PPrintEngine.document

val pp_event : 'a Isla_lang__.Isla_lang_ast.event -> PPrintEngine.document

val pp_trc : 'a Isla_lang__.Isla_lang_ast.trc -> PPrintEngine.document

module Base : sig ... end

This module wraps all isla-lang functionality. No other module should directly touch the Isla_lang module.

module Cache : sig ... end

This module provide a caching system for isla trace on top of Server.

module Conv : sig ... end

module Manip : sig ... end

This module provide generic manipulation function of isla ast

module Preprocess : sig ... end

This module is about preprocessing isla traces. This includes:

module Run : sig ... end

This module provide facility to run Isla trace over states

module Server : sig ... end

This module is about launching isla as a background server and using it

module Test : sig ... end

This file about testing interaction with Isla for single instructions

module Type : sig ... end

This module is about type isla trace and register discovery.