Module Isla.Run

This module provide facility to run Isla trace over states

The main functions are trc for pure interface and trc_mut for imperative interface.

RunError will be thrown when something goes wrong.

It is for testing purpose only, otherwise use Trace.Run. Typing does not work, and some other expected behavior may not work either.

This module can be considered deprecated/legacy.

exception RunError of Isla__.Base.lrng * string

Exception that represent an Isla runtime error which should not happen

val run_error : Isla__.Base.lrng -> ('a, unit, string, 'b) Stdlib.format4 -> 'a

Throw a run error with the string part as formated by the format string

type value_context = State.exp Utils.HashVector.t

The contex of value that associate isla variable numbers to state expression

val get_var : Isla__.Base.lrng -> State.Exp.t Utils.HashVector.t -> int -> State.Exp.t

Get the expression associated to the free variable. Throw RunError, if the variable is no bound. The lrng is for error reporting.

val exp_conv_subst : value_context -> Base.rexp -> State.exp

Convert an Isla expression to an Ast by substituing all free variable with the bound expression in the value_context.

If a variable is not bound, throw RunError

val exp_of_valu : Isla__.Base.lrng -> State.Exp.t Utils.HashVector.t -> Isla__.Base.valu -> State.exp

Give the State.exp that represents the input valu.

A symbolic variable i is represented by the expression bound to it in the provided value_context.

Newly created expression are located with the provided lrng.

If the value is not convertible to a state expression, throw a RunError

val write_to_var : 'a -> 'b Utils.HashVector.t -> int -> 'b -> unit

This function write an expression to symbolic variable. The write is ignored if the variable was already set because isla guarantee that it would be the same value (Trusting Isla here)

val write_to_valu : 'a -> 'b Utils.HashVector.t -> Isla__.Base.valu -> 'b -> unit

This function write an expression to an valu.

If the valu is a variable, it is added to the context, otherwise nothing happens.

val event_mut : value_context -> State.t -> Base.revent -> unit

Run an event on State.t and a value_context by mutating both

val trc_mut : ?⁠vc:State.exp Utils.HashVector.t -> State.t -> Base.rtrc -> unit

This function run an isla trace on a state by mutation. If a vc is provided, then it is used and mutated according to the trace.

Any encountered branch are ignored and their assertion are added to the state

val trc : State.t -> Base.rtrc -> State.t

This function run an isla trace on a state and return the end state as a new state

It is just a wrapper of run_trc_mut that remove the imperative interface The new state is fresh and locked.