Module Run.Bb_lib

type trc = Trace.t

type state = State.t

type t = {

main : trc array;

}

Type of a basic block.

The main part is the traces of all the non-branching instruction

val from_binary : Utils.BytesSeq.t -> t

Take a binary block and call isla on all the instruction to get traces Also does the typing of traces for register discovery. TODO Support variable length instructions

val simplify_mut : t -> unit

Simplifies the traces in the basic block

val run_mut : ?⁠dwarf:Dw.t -> State.t -> t -> unit

Run a linear basic block on a state by mutation.

If dwarf is provided, the run is typed.

val run : ?⁠dwarf:Dw.t -> State.t -> t -> state

Run a linear basic block on a trace and return a new state

If dwarf is provided, the run is typed.

val pp : t -> PPrintEngine.document

Pretty print the basic block (The traces)