Module Base.Id

The type of a state ID. for now it's an integer, but it may change later In particular whether a state belong to O0 or O2 may be part of the id at some point.

type t

val to_string : t -> string

val of_string : string -> t

val equal : t -> t -> bool

val pp : t -> PPrintEngine.document