Module Tests.ExpGen

type ('v, 'm) exp = ('v, 'm) Exp.Typed.t

module Var : sig ... end

Test variables to instantiate variable dependent functors for testing.

module ExpT : sig ... end

module Z3 : sig ... end

module Gen : sig ... end

val shrink_propagate : ('v, 'm) Exp.Typed.t Tests.Common.Q.Shrink.t -> ('v, 'm) Exp.Typed.t -> (('m Ast.ty, 'v, Ast.no, 'm) Ast.exp -> unit) -> unit

Propagate an existing shrinker, by try it on all descendants of this. The The provided shrinker must preserve the type.

val from_gen : ?⁠shrink:ExpT.t Tests.Common.Q.Shrink.t -> (ExpT.var, Ast.no) exp Tests.Common.Q.Gen.t -> ExpT.t Tests.Common.Q.arbitrary

Generate an expression arbitrary from a generator (and optionally a shrinker)