Module Trace.Cache

This module provides a caching system for fully processed traces

The top level function to get traces from an opcode is get_traces. This is the function called by the Run.Runner.

module Opcode = Isla.Cache.Opcode

module Epoch = Isla.Cache.Epoch

module Traces : sig ... end

Store element of type Base.tlist on disk.

module TC : sig ... end

val cache : TC.t option Stdlib.ref

type config = Isla.Cache.config

val start : Isla.Cache.config -> unit

Start the caching system. Start Isla.Cache too

val stop : unit -> unit

Stop the caching system

val get_cache : unit -> TC.t

Get the cache and fails if the cache wasn't started

val get_traces : Utils.BytesSeq.t -> Base.t list

Get the traces of the opcode given. Use Isla.Server if the value is not in the cache

val get_instr : Utils.BytesSeq.t -> Instr.t

Get a full blown Instr from the opcode, going through the whole Isla pipeline if necessary.