Module Trace.Context

This module provide the type for a context to run a trace

Any information that should be required to run a trace but is not part of the state itself should be added here

type t = {

reg_writes : (State.Reg.t * State.tval) Utils.Vec.t;
Stores the delayed register writes

mem_reads : State.tval Utils.HashVector.t;
Stores the result of memory reads

state : State.t;

dwarf : Dw.t option;
Optionally DWARF information. If present, typing is enabled

}

The context to run a trace

val make_context : ?⁠dwarf:Dw.t -> State.t -> t

Build a context from a state

val expand_var : ctxt:t -> Base.Var.t -> Ast.no Ast.ty -> State.exp

Expand a Trace variable to a State expression, using the context

val typing_enabled : ctxt:t -> bool

Tell if typing should enabled with this context