Module Trace.Instr

This module provide the representation of an instruction. It only contain generic information about the opcode and not specific information about a place in the code.

module Reg = State.Reg

type trace_meta = {

trace : Base.t;

jump : bool;

footprint : Reg.t list;

}

A simple trace with it's metadata

type t = {

traces : trace_meta list;

length : int;
Bytes length

footprint : Reg.t list;

}

A full instruction representation

val footprint : t -> Reg.t list

Helper to access the footprint field

val trace_meta_of_trace : Base.t -> trace_meta

Compute the metadata of trace

val of_traces : Base.t list -> t

Generate an instruction data from a list of traces

val pp : t -> Utils.Pp.document

Pretty print the representation of an instruction