Module Utils.Cache

Utility

This section is not part of the external API, but creating an mli file here seemed needlessly annoying.

val base_dir : string

The name of the read-dwarf cache. .rdcache for now.

val find_dir : unit -> string

Find the current cache folder as described in the README.

When searching for a cache, this function will search if there already is a base_dir directory either in the current directory or one of its parent and use the closest one it find. If it finds none and need a cache, it will create a new directory named base_dir in the current directory. Therefore the returned directory always exists.

val removedir : string -> unit

Remove a directory and all it's content.

TODO: Make that Windows friendly

TODO move that in Files

val cleardir : string -> unit

Clear the content of directory

val int_to_file : int -> string

Convert an hash integer to a filename: 16 characters in hexadecimal

type file_type =

| NOPE

| FILE

| DIR of int

The type of an entry at a given hash

val file_type : string -> file_type

Give the file type of the given type. Do not count the number of entry in ther

val to_keyfile : string -> string

Transform a filename in the key filename

val movekey : string -> string -> unit

movekey old new moves a key correponding to the given filenames, if they exist

val removekey : string -> unit

Removes a key correponding to the given filename, if it exists

val clear_cache : string -> unit

Delete the supplied cache

val list_caches : unit -> string list

List all existing caches

val clear_all_caches : unit -> unit

Delete all the caches (and the base cache directory)

exception Exists

Thrown when adding a key that is already bound in cache

Input signatures

module type Key = sig ... end

The interface of keys.

module IntKey : Key with type Key.t = int

An implementation of Key on ints where hash is the identity and no file is written

module type Value = sig ... end

The interface of values: They just need to be able to be serialized to files. This is isn't plain Marshal because some values may want a human readable format.

module type Epoch = sig ... end

A cache can be indexed by an Epoch. When reading a cache with an incompatible epoch, then the cache is deleted on load. Use UnitEpoch to not have this functionality. Single caches currently do not support epochs.

module UnitEpoch : sig ... end

A dummy epochs implementation, if not epoch is needed

module IntEpoch : sig ... end

CacheMap

module type S = sig ... end

The output signature of Make

module Make : functor (Key : Key) -> functor (Value : Value) -> functor (Epoch : Epoch) -> S with type key = Key.t and type value = Value.t and type epoch = Epoch.t

The Key must provide the Key interface, where the values must satisfy the Value interface.

Single Cache

module type SingleS = sig ... end

The signature of the output of Single

module Single : functor (Value : Value) -> SingleS with type value = Value.t

The functor is to make a single cached value. This do not support epochs (Yet)

Command line operations

module Cmd : sig ... end

The cache command line. TODO documentation of the inside