Module Utils.HashVector

type 'a t

The type of the hash vector

val mem : 'a t -> int -> bool

Check if an element is set

val copy : 'a t -> 'a t

Copy the HashVector

exception Exists

val set : 'a t -> int -> 'a -> unit

Set a value. Create a new binding if necessary

val clear : 'a t -> int -> unit

Clear a binding, do nothing is the value is not bound.

val add : 'a t -> int -> 'a -> unit

Create a new binding. Throws Exists if a binding already exists

val get_opt : 'a t -> int -> 'a option

Return Some v if v is bound to the integer in the hash vector. Return None if nothing is bound

val get : 'a t -> int -> 'a

Return the value bound to the integer. Throw Not_found if the integer is not bound

val empty : unit -> 'a t

Create an empty hashVector.t

val bindings : 'a t -> (int * 'a) list

Returns a list of bindings in the hash vector

val of_seq : (int * 'a) Utils.Seq.t -> 'a t

Fill the hashVector from a sequence

val pp : ('a -> Utils.Pp.document) -> 'a t -> Utils.Pp.document

Pretty print a hashVector