Module Utils.List

include Stdlib.List

type 'a t = 'a list =

| ([])

| (::) of 'a * 'a list

val length : 'a list -> int

val compare_lengths : 'a list -> 'b list -> int

val compare_length_with : 'a list -> int -> int

val cons : 'a -> 'a list -> 'a list

val hd : 'a list -> 'a

val tl : 'a list -> 'a list

val nth : 'a list -> int -> 'a

val nth_opt : 'a list -> int -> 'a option

val rev : 'a list -> 'a list

val init : int -> (int -> 'a) -> 'a list

val append : 'a list -> 'a list -> 'a list

val rev_append : 'a list -> 'a list -> 'a list

val concat : 'a list list -> 'a list

val flatten : 'a list list -> 'a list

val iter : ('a -> unit) -> 'a list -> unit

val iteri : (int -> 'a -> unit) -> 'a list -> unit

val map : ('a -> 'b) -> 'a list -> 'b list

val mapi : (int -> 'a -> 'b) -> 'a list -> 'b list

val rev_map : ('a -> 'b) -> 'a list -> 'b list

val filter_map : ('a -> 'b option) -> 'a list -> 'b list

val fold_left : ('a -> 'b -> 'a) -> 'a -> 'b list -> 'a

val fold_right : ('a -> 'b -> 'b) -> 'a list -> 'b -> 'b

val iter2 : ('a -> 'b -> unit) -> 'a list -> 'b list -> unit

val map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val rev_map2 : ('a -> 'b -> 'c) -> 'a list -> 'b list -> 'c list

val fold_left2 : ('a -> 'b -> 'c -> 'a) -> 'a -> 'b list -> 'c list -> 'a

val fold_right2 : ('a -> 'b -> 'c -> 'c) -> 'a list -> 'b list -> 'c -> 'c

val for_all : ('a -> bool) -> 'a list -> bool

val exists : ('a -> bool) -> 'a list -> bool

val for_all2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val exists2 : ('a -> 'b -> bool) -> 'a list -> 'b list -> bool

val mem : 'a -> 'a list -> bool

val memq : 'a -> 'a list -> bool

val find : ('a -> bool) -> 'a list -> 'a

val find_opt : ('a -> bool) -> 'a list -> 'a option

val filter : ('a -> bool) -> 'a list -> 'a list

val find_all : ('a -> bool) -> 'a list -> 'a list

val partition : ('a -> bool) -> 'a list -> 'a list * 'a list

val assoc : 'a -> ('a * 'b) list -> 'b

val assoc_opt : 'a -> ('a * 'b) list -> 'b option

val assq : 'a -> ('a * 'b) list -> 'b

val assq_opt : 'a -> ('a * 'b) list -> 'b option

val mem_assoc : 'a -> ('a * 'b) list -> bool

val mem_assq : 'a -> ('a * 'b) list -> bool

val remove_assoc : 'a -> ('a * 'b) list -> ('a * 'b) list

val remove_assq : 'a -> ('a * 'b) list -> ('a * 'b) list

val split : ('a * 'b) list -> 'a list * 'b list

val combine : 'a list -> 'b list -> ('a * 'b) list

val sort : ('a -> 'a -> int) -> 'a list -> 'a list

val stable_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val fast_sort : ('a -> 'a -> int) -> 'a list -> 'a list

val sort_uniq : ('a -> 'a -> int) -> 'a list -> 'a list

val merge : ('a -> 'a -> int) -> 'a list -> 'a list -> 'a list

val to_seq : 'a list -> 'a Stdlib.Seq.t

val of_seq : 'a Stdlib.Seq.t -> 'a list

val repeat : int -> 'a -> 'a t

repeat n a return a list of n as.

val set_nth : 'a t -> int -> 'a -> 'a t

Set the nth value to the new value and return the modified list

val last : 'a t -> 'a

Give the last element of the list. Raise Invalid_argument if the list is empty.

val last_opt : 'a t -> 'a option

Give the last element of the list. Return None if the list is empty.

Iterators

val fold_left_same : ('a -> 'a -> 'a) -> 'a t -> 'a

Same as fold_left, but do not require a start element, instead the function start with the first element of the list: fold_left_same f [b1; ...; bn] = f (... (f (f b1 b2) ...) bn.

It will fail with Invalid_argument if the list is empty.

It can also be written as fold_left_same f l = fold_left f (hd l) (tl l)

val of_array_map : ('a -> 'b) -> 'a array -> 'b t

Map a function at the same time as we are creating a list from an array

Warning: The function is mapped from the right to the left (in case it has side-effects).

of_array_map f l = of_array (Array.map f l) = map f (of_array l)

val of_array_mapi : (int -> 'a -> 'b) -> 'a array -> 'c t

Same as of_array_map but with the index

val concat_map_rev : ('a -> 'b list) -> 'a t -> 'b t

Same as concat_map then rev

val concat_map : ('a -> 'b list) -> 'a t -> 'b t

Same as map then concat.

TODO: find a clean way of doing conditional compilation, otherwise this will shadow the official concat_map in 4.10

val find_map : ('a -> 'b option) -> 'a t -> 'b option

find_map f l applies f to the elements of l in order, and returns the first result of the form Some v, or None if f always return None

TODO: find a clean way of doing conditional compilation, otherwise this will shadow the official find_map in 4.10

val filter_opt : 'a option t -> 'a t

Takes a list of option and only keeps the Some. Equivalent to filter_map Fun.id

val partition_map : ('a -> 'b option) -> 'a list -> 'a t * 'b t

This function behaves as partition then a map. First we do a partition with the function assuming Some means true, then for all the extracted elements we map f on them and return the results.

Formally: partition_map f l = (filter (fun a -> f a = None) l, filter_map f l)

Removing elements

val remove : ('a -> bool) -> 'a t -> 'a list option

Remove the first element matching the predicate. Returns None if no element matches the predicate

val drop : int -> 'a t -> 'a t

Drop the specified number of item from the list. If n is greater than the size of the list, then return the empty list

val take_rev : int -> 'a t -> 'a t

Take the specified number of items from the list, but reverse If n is greater than the size of the list, then return the list reversed

Tail-recursive

val take : int -> 'a t -> 'a list

Take the specified number of items from the list. If n is greater than the size of the list, then return the list

l = take n l @ drop n l

val sub : pos:int -> len:int -> 'a t -> 'a list

sub l pos len return the sub-list of l starting at pos of length len

Sorted list manipulation

val merge_uniq : ('a -> 'a -> int) -> 'a t -> 'a t -> 'a t

Same as merge but duplicate elements are deleted. It is assumed that element are not duplicate in the argument (like if sorted with sort_uniq)

Sequences

val of_seq_rev : 'a Utils.Seq.t -> 'a t

Build a list from a sequence, but in reverse. Same as of_seq then rev

Equality and comparison

val equal : ('a -> 'b -> bool) -> 'a t -> 'b t -> bool

val mem : ('a -> 'b -> bool) -> 'a -> 'b t -> bool

val compare : ('a -> 'b -> int) -> 'a t -> 'b t -> int

If the list have the same length this a lexicographic compare. If one list is shorter, then the missing value a considered smaller than any actual values.

A mental model could to view list as infinite sequence of options that are None after the end of the list. Then it would be proper lexicographic ordering with Option.compare

List monad

val bind : 'a t -> ('a -> 'b list) -> 'b t

Monadic bind. It's just concat_map

val return : 'a -> 'a t

Monadic return

val short_combine : 'a t -> 'b t -> ('a * 'b) t

Same as combine but if a list is shorter then the elements of the longest list are discarded

val let+ : 'a list -> ('a -> 'b) -> 'b list

Applicative let binding. let+ x = xl in e = let* x = xl in return e

val and+ : 'a t -> 'b t -> ('a * 'b) t

Not strict applicative merge (short_combine). If both list have different length, the longer one is cropped

val let+! : 'a list -> ('a -> unit) -> unit

Iterative let binding (The expression in the in must be unit). This replace implicitly a unit member of the monad (that is assumed to be uninteresting) to a true unit. In other words, it's an iter: let+! x = l in e = List.iter (fun x -> e) l

val and+! : 'a list -> 'b list -> ('a * 'b) list

Strict applicative merge (combine). Will throw if lists have different length

val let* : 'a t -> ('a -> 'b list) -> 'b t

Monadic let binding

val prod : 'a t -> 'b list -> ('a * 'b) t

Do the Cartesian product of two lists

val and* : 'a t -> 'b list -> ('a * 'b) t

Monadic merge. let* x = xl and* y = yl in ... = let* x= xl in let* y = yl in ...