Module Utils.Logs

type 'a printf = ('a, Stdlib.out_channel, unit) Stdlib.format -> 'a

The type of normal logging function

type ('a, 'b) printf_fatal = ('a, Stdlib.out_channel, unit, 'b) Stdlib.format4 -> 'a

The type of a failing logging function, that will exit after printing it's message

type level =

| Base
The actual output. The only thing printed in quiet mode. Should only appear in CLI modules

| Err
An error message

| Warn
An warning

| Info
Details on what happens that should be understandable if the person do not know the corresponding module

| Debug
Everything happening in detail

The type of log level

val pp_level : level -> Utils.Pp.document

Pretty prints a level

val level_to_string : level -> string

Convert level to string

val level_to_header : level -> string

Convert level to string header like "[Error]". Base is the empty string

val set_default_level : level -> unit

Set a default level to all the modules. Erase local customized levels

val set_level : string -> level -> unit

Set level of a module by name

val level_conv : level Cmdliner.Arg.conv

Parser for log level on command line

val set_stdout_level : level -> unit

Set level below which the output goes to stdout

module type String = sig ... end

module Logger : functor (S : String) -> sig ... end

This declare a logger instance. The string parameter is the logger name. It should be the OCaml module name. The normal way of instantiating this module is:

val process_opts : bool -> bool list -> string list -> string list -> level -> unit

val term : unit Cmdliner.Term.t