Module RngMap.Make

How to make a RngMap from a LenObject

Parameters

Obj : LenObject

Signature

type obj = Obj.t

The type of the contained object

type obj_off = obj * int

The type of an object with an offset

type t

The type of the map from address ranges to obj

val empty : t

An empty RngMap

val is_in : objaddr:int -> obj -> int -> bool

Test if an address is inside the object at address objaddr

val at : t -> int -> obj

Get the object containing the address. Throw Not_found if no object contains the address

val at_opt : t -> int -> obj option

Get the object containing the address. None if no object contains the address

val at_off : t -> int -> obj_off

Get the object containing the address and the offset of the address inside the object

at_off map addr = (obj, off) 

means:

         |                      |           |         |
       map 0                 obj start    point    obj end
         |<--------------addr-------------->|
                                |<---off--->|
                                |<------len obj------>|

In other words, at_off allow a change of coordinate from the map frame to the object frame.

Throw Not_found if no object contains the address

val at_off_opt : t -> int -> obj_off option

Get the object containing the address and the offset of the address inside the object. See at_off for more explanation.

None if no object contains the address

val update : (obj -> obj) -> t -> int -> t

Update the binding containing the provided address. If no binding contained the address, this is a no-op

val map : (obj -> obj) -> t -> t

Map a function over all the objects

val mapi : (int -> obj -> obj) -> t -> t

Map a function over all the objects with their address

val iter : (obj -> unit) -> t -> unit

Iter a function over all the objects

val iteri : (int -> obj -> unit) -> t -> unit

Iter a function over all the objects with their address

val clear_at : t -> int -> t

Clear the object containing the address if any

val clear : t -> pos:int -> len:int -> t

Clear an area of the RngMap.

If an object is partially in the specified block. It will be removed entirely.

See clear_crop for a different behavior. See clear_bounds to allow some bounds to be infinity.

val clear_crop : t -> pos:int -> len:int -> crop:(pos:int -> len:int -> obj -> obj) -> t

Clear an area of the RngMap.

If a block is partially in the specified block, It will be cropped by using the provided crop function.

crop ~pos ~len obj is supposed to crop the object obj and keep only the segment [pos:pos +len) of it (in the object coordinate frame).

val clear_bounds : ?⁠start:int -> ?⁠endp:int -> t -> t

Same as clear but if a bound is missing, then we erase until infinity in that direction. The target interval is [start:endp).

In particular clear_bounds map = empty.

val add : t -> int -> obj -> t

Add an object at a specific address. The whole range of addresses covered by the object must be free

val addp : t -> obj_off -> t

Add an object at a specific address. The whole range of addresses covered by the object must be free

val bindings : t -> (int * obj) list

Give the list of bindings

val to_seq : ?⁠start:int -> ?⁠endp:int -> t -> (int * obj) Utils.Seq.t

Return a sequence of all the object overlapping the range [start:endp). The first and last element may not be entierly contained in the ranged. If any bound is unspecified, it goes to infinity in that direction.

In particular to_seq map will iterate the entiere RngMap