Module RngMap.PairLenObject

For types that do not have an inner length representation, we add it with a pair

Parameters

Obj : Object

Signature

type t = Obj.t * int

The type to be indexed by starting addresses, must have a length

val len : t -> int

The type of range end