Module Z3.ContextCounter

Module for handling a context numbering scheme automatically.

To use it do: module Whatever = ContextCounter(struct let str= "name here" end) Then you can use it with Whatever.openc () and Whatever.closec (). You can get the current instance number with Whatever.num() at any time.

Parameters

S : Utils.Logs.String

Signature

val counter : Utils.Counter.t

val openc : unit -> unit

Open a new context of that ContextCounter. Ensure the server is started

val num : unit -> int

Get the current context number

val closec : unit -> unit

Close a context opened with openc. Assert that the current context was indeed opened by this module